### 2020-12-18 #2919 @whedon
[PRE REVIEW]: Pychron: Automated Data Collection and Reduction for Noble Gas and Ar/Ar Geochronology
**Submitting author:** @jirhiker (<a href="http://orcid.org/0000-0003-4727-7472">Jake Ross</a>)
**Repository:** <a href="https://github.com/NMGRL/pychron" target ="_blank">https://github.com/NMGRL/pychron</a>
**Version:** v20.1
