### 2020-12-18 #2919 @kthyng
Hi @jirhiker and thanks for resubmitting. Did you include the JOSS paper in this repo? That is required for us to be able to review. Please add it in.
