### 2020-12-19 #2919 @danielskatz
👋 @jirhiker - your paper doesn't compile due to blank orcid lines - for authors without orcids, just remove the orcid line in the metadata.  Then use the command I did above to try to generate the PDF again
