### 2020-12-23 e23ad1c
fixed table config for analyses
