+++ pychron/spectrometer/pfeiffer/__init__.py
@@ -1,5 +1,5 @@
 # ===============================================================================
-# Copyright 2018 ross
+# Copyright 2020 ross
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,10 +13,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
+ERRORS = []
 
-from pychron.hardware.lakeshore.base_controller import BaseLakeShoreController
 
+class PfeifferMixin(object):
+    def handle_response(self, cmd, resp, *args, **kw):
+        if resp in ERRORS:
+            if hasattr(self, 'warning'):
+                self.warning('Command {}==>{},{}'.format(cmd, resp, ERRORS[resp]))
+            resp = None
+        return resp
 
-class Model335TemperatureController(BaseLakeShoreController):
-    pass
 # ============= EOF =============================================
