+++ pychron/lasers/laser_managers/ostech_laser_manager.py
@@ -1,5 +1,5 @@
 # ===============================================================================
-# Copyright 2018 ross
+# Copyright 2020 ross
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,10 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
+from pychron.lasers.laser_managers.base_lase_manager import BaseLaserManager
 
-from pychron.hardware.lakeshore.base_controller import BaseLakeShoreController
 
-
-class Model335TemperatureController(BaseLakeShoreController):
+class OsTechLaserManager(BaseLaserManager):
     pass
 # ============= EOF =============================================
