+++ pychron/hardware/__init__.py
@@ -63,6 +63,7 @@ HW_PACKAGE_MAP = {
     'NMGRLMagnetDumper': 'pychron.furnace.magnet_dumper',
     'LamontFurnaceControl': 'pychron.hardware.labjack.ldeo_furnace',
 
+    'Model325TemperatureController': 'pychron.hardware.lakeshore.model325',
     'Model335TemperatureController': 'pychron.hardware.lakeshore.model335',
     'Model336TemperatureController': 'pychron.hardware.lakeshore.model336',
     'MKSSRG': 'pychron.hardware.gauges.mks.srg',

[…]
