+++ pychron/lasers/tasks/laser_task.py
@@ -23,7 +23,7 @@ from pychron.envisage.view_util import open_view
 from pychron.lasers.pattern.pattern_maker_view import PatternMakerView
 # ============= standard library imports ========================
 # ============= local library imports  ==========================
-from pychron.pychron_constants import FUSIONS_CO2, FUSIONS_DIODE
+from pychron.pychron_constants import FUSIONS_CO2, FUSIONS_DIODE, OSTECH_DIODE
 
 
 class BaseLaserTask(BaseHardwareTask):
@@ -104,6 +104,40 @@ class FusionsTask(BaseLaserTask):
             #         if self.manager.use_video:
             #             v = self.manager.degasser_factory()
             #             self.window.application.open_view(v)
+
+
+class OsTechDiodeTask(BaseLaserTask):
+    id = 'pychron.ostech.diode'
+    name = OSTECH_DIODE
+
+    def create_central_pane(self):
+        if self.manager.mode == 'client':
+            from pychron.lasers.tasks.panes.ostech import OsTechDiodeClientPane
+            return OsTechDiodeClientPane(model=self.manager)
+        else:
+            from pychron.lasers.tasks.panes.ostech import OsTechDiodePane
+            return OsTechDiodePane(model=self.manager)
+
+    def create_dock_panes(self):
+        if self.manager.mode == 'client':
+            return []
+        else:
+            from pychron.lasers.tasks.panes.ostech import OsTechDiodeStagePane
+            # from pychron.lasers.tasks.panes.diode import FusionsDiodeControlPane
+            # from pychron.lasers.tasks.panes.diode import FusionsDiodeSupplementalPane
+            # from pychron.lasers.tasks.laser_panes import PulsePane
+            # from pychron.lasers.tasks.laser_panes import OpticsPane
+            # from pychron.lasers.tasks.laser_panes import AuxilaryGraphPane
+
+            return [
+                OsTechDiodeStagePane(model=self.manager),
+                # FusionsDiodeControlPane(model=self.manager),
+                # FusionsDiodeSupplementalPane(model=self.manager),
+
+                # PulsePane(model=self.manager),
+                # OpticsPane(model=self.manager),
+                # AuxilaryGraphPane(model=self.manager)
+            ]
 
 
 class AblationCO2Task(BaseLaserTask):
