+++ pychron/pychron_constants.py
@@ -186,6 +186,7 @@ NULL_EXTRACT_DEVICES = [EXTRACT_DEVICE, LINE_STR, NO_EXTRACT_DEVICE, None, '']
 CRYO = 'Cryo'
 FUSIONS_UV = 'Fusions UV'
 FUSIONS_DIODE = 'Fusions Diode'
+OSTECH_DIODE = 'OsTech Diode'
 FUSIONS_CO2 = 'Fusions CO2'
 FUSIONS = [FUSIONS_CO2, FUSIONS_DIODE, FUSIONS_UV]
 

[…]
