### 2021-01-05 3c8ecb4
fixed stage_manager default
