### 2021-01-05 98307db
update stage pane
