+++ pychron/lasers/tasks/laser_panes.py
@@ -189,17 +189,28 @@ class StageControlPane(TraitsDockPane):
         return tabs
 
     def traits_view(self):
-        pgrp = HGroup(UItem('stage_manager.calibrated_position_entry',
-                            tooltip='Enter a position e.g 1 for a hole, '
-                                    'or 3,4 for X,Y'),
-                      icon_button_editor('stage_manager.autocenter_button', 'find',
-                                         tooltip='Do an autocenter at the current location',
-                                         enabled_when='stage_manager.autocenter_manager.use_autocenter'),
-                      icon_button_editor('stage_manager.manual_override_position_button', 'edit-move',
-                                         tooltip='Manual define the X,Y coordinates for current position',
-                                         enabled_when='stage_manager.calibrated_position_entry'),
-                      label='Calibrated Position',
-                      show_border=True)
+        if self.model.stage_manager.__class__.__name__ == 'VideoStageManager':
+            pgrp = HGroup(UItem('stage_manager.calibrated_position_entry',
+                                tooltip='Enter a position e.g 1 for a hole, '
+                                        'or 3,4 for X,Y'),
+                          icon_button_editor('stage_manager.autocenter_button', 'find',
+                                             tooltip='Do an autocenter at the current location',
+                                             enabled_when='stage_manager.autocenter_manager.use_autocenter'),
+                          icon_button_editor('stage_manager.manual_override_position_button', 'edit-move',
+                                             tooltip='Manual define the X,Y coordinates for current position',
+                                             enabled_when='stage_manager.calibrated_position_entry'),
+                          label='Calibrated Position',
+                          show_border=True)
+        else:
+            pgrp = HGroup(UItem('stage_manager.calibrated_position_entry',
+                                tooltip='Enter a position e.g 1 for a hole, '
+                                        'or 3,4 for X,Y'),
+                          icon_button_editor('stage_manager.manual_override_position_button', 'edit-move',
+                                             tooltip='Manual define the X,Y coordinates for current position',
+                                             enabled_when='stage_manager.calibrated_position_entry'),
+                          label='Calibrated Position',
+                          show_border=True)
+
         hgrp = HGroup(UItem('stage_manager.stop_button'),
                       UItem('stage_manager.home'),
                       UItem('stage_manager.home_option',
