### 2021-01-05 99ca311
updated cosmo correction with correct math from cameron
