+++ pychron/hardware/newport/newport_motion_controller.py
@@ -147,44 +147,6 @@ class NewportMotionController(MotionController):
         if 'x' in self.axes:
             return self.axes['x'].id == 2
 
-    def get_current_xy(self):
-        # x, y = None, None
-        # if self.mode == 'grouped':
-        #     f = self.ask('{}HP'.format(self.groupobj.id), verbose=True)
-        #     # cmd = '{}TP?;{}TP?'.format(self.axes['x'].id, self.axes['y'].id)
-        #     # f = self.ask(cmd, verbose=True)
-        #     # args = f.split(',')[:2]
-        #     try:
-        #         f = f.strip()
-        #         args = f.split('\n')
-        #         x, y = map(float, map(str.strip, args))
-        #
-        #         ax = self.axes['x']
-        #         x = self._sign_correct(x, 'x', ratio=False) / ax.drive_ratio
-        #
-        #         ax = self.axes['y']
-        #         y = self._sign_correct(y, 'y', ratio=False) / ax.drive_ratio
-        #     except BaseException, e:
-        #         # import traceback
-        #         # traceback.print_exc()
-        #         self.warning('get_current_xy failed. {}'.format(e))
-        #         x = self.get_current_position('x')
-        #         y = self.get_current_position('y')
-        #
-        # else:
-        x = self.get_current_position('x')
-        y = self.get_current_position('y')
-        if x is None or y is None:
-            return
-
-        return x, y
-
-        # cmd = '{}TP?;{}TP?'.format(self.axes['x'].id, self.axes['y'].id)
-        # # cmd = self._build_query('TP')
-        # f = self.ask(cmd, verbose=True)
-
-        # args = f.split(',')[:2]
-
     def get_current_position(self, aid):
         if isinstance(aid, str):
             if aid in self.axes:
