### 2021-01-08 056a0e0
updated ngx for combo atona/multiplier measurements
