+++ pychron/hardware/zaber/zaber_motion_controller.py
@@ -13,46 +13,112 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from traits.api import Str, Int
+from traits.api import Str, CInt, Enum
+
+from zaber_motion import Library
+Library.enable_device_db_store()
+
 from pychron.hardware.motion_controller import MotionController
-from zaber_motion import Connection, Units
+from pychron.hardware.zaber.axis import ZaberAxis
+from zaber_motion import Units
+
+class AsciiWrapper:
+    def __init__(self, devices):
+        self._device=devices[0]
+
+    def get_axis(self, a):
+        return self._device.get_axis(a)
+
+    def is_busy(self):
+        return self._device.all_axes.is_busy()
+
+
+class BinaryWrapper:
+    def __init__(self, devices):
+        self._devices = devices
+
+    def get_axis(self, a):
+        return self._devices[a]
+
+    def is_busy(self):
+        return any((d.is_busy() for d in self._devices))
 
 
 class ZaberMotionController(MotionController):
     port = Str
-    device_id = Int
+    device_id = CInt
+    baudrate = CInt
+    protocol = Enum('ASCII', 'Binary')
     _device = None
-    _units = Units.LENGTH_MILLIMETERS
+    _units = Units.LENGTH_MILLIMETRES
+    _wrapper = None
+    move_enabled = False
 
     def load(self, *args, **kw):
         config = self.get_configuration()
         if config:
             section = 'Communications'
-            self.set_attribute(config, section, 'port')
-            self.set_attribute(config, section, 'device_id', optional=True, default=0)
+            self.set_attribute(config, 'port', section, 'port')
+            self.set_attribute(config, 'device_id', section, 'device_id', optional=True, default=0)
+            self.set_attribute(config, 'baudrate', section, 'baudrate', optional=True, default=9600)
+            self.set_attribute(config, 'protocol', section, 'protocol', optional=True, default='ASCII')
+
+            self.axes_factory(config)
+
+            return True
+
+    def initialize(self, *args, **kw):
+        return True
 
     def open(self, *args, **kw):
-        connection = Connection.open_serial_port(self.port)
+        self.debug('scanning port {}'.format(self.port))
+        if self.protocol=='ASCII':
+            from zaber_motion.ascii import Connection
+        else:
+            from zaber_motion.binary import Connection
+
+        connection = Connection.open_serial_port(self.port, self.baudrate)
         device_list = connection.detect_devices()
-        self._device = device_list[self.device_id]
+        self.debug(device_list)
+        if self.protocol == 'ASCII':
+            klass = AsciiWrapper
+        else:
+            klass = BinaryWrapper
+
+        self._wrapper = klass(device_list)
+        return True
+
+    def _get_simulation(self):
+        return False
 
     def linear_move(self, x, y, block=True, *args, **kw):
         xaxis = self.single_axis_move('x', x, block=False)
         yaxis = self.single_axis_move('y', y, block=False)
 
         if block:
-            xaxis.wait_until_idle()
-            yaxis.wait_until_idle()
+            if self.protocol == 'ASCII':
+                xaxis.wait_until_idle()
+                yaxis.wait_until_idle()
+            else:
+                while 1:
+                    if not xaxis.is_busy() and not yaxis.is_busy():
+                        return
 
     def single_axis_move(self, key, value, block=True, *args, **kw):
         axis = self._get_device_axis(key)
-        axis.move_absolute(value, self._units, wait_until_idle=block)
+
+        if self.move_enabled:
+            if self.protocol == 'ASCII':
+                axis.move_absolute(value, self._units, wait_until_idle=block)
+            else:
[…]
