+++ pychron/pychron_constants.py
@@ -188,7 +188,10 @@ FUSIONS_UV = 'Fusions UV'
 FUSIONS_DIODE = 'Fusions Diode'
 OSTECH_DIODE = 'OsTech Diode'
 FUSIONS_CO2 = 'Fusions CO2'
+CHROMIUM_CO2 = 'Chromium CO2'
+ABLATION_CO2 = 'Ablation CO2'
 FUSIONS = [FUSIONS_CO2, FUSIONS_DIODE, FUSIONS_UV]
+LASER_PLUGINS = [a.replace(' ', '') for a in (FUSIONS_CO2, FUSIONS_DIODE, CHROMIUM_CO2, ABLATION_CO2, OSTECH_DIODE)]
 
 NO_BLANK_CORRECT = (BLANK, DETECTOR_IC, BACKGROUND)
 
