### 2021-01-13 ea1fa65
added labjack based gauge controller
