+++ pychron/hardware/adc/adc_device.py
@@ -30,18 +30,22 @@ class PolynomialMapperMixin(HasTraits):
     def load_mapping(self, config):
         conv = 'Conversion'
         if config.has_section(conv):
-            pmapper = PolynomialMapper()
-            coeffs = self.config_get(config, conv, 'coefficients')
-            pmapper.parse_coefficient_string(coeffs)
-            pmapper.output_low = self.config_get(config, conv, 'output_low', cast='float')
-            pmapper.output_high = self.config_get(config, conv, 'output_high', cast='float')
-
+            pmapper = self.factory(config, conv)
             self.poly_mapper = pmapper
             self.set_attribute(config, 'mapped_name', conv, 'name')
 
             if self.mapped_name:
                 u = self.config_get(config, conv, 'units', default='')
                 self.graph_ytitle = '{} ({})'.format(self.mapped_name.capitalize(), u)
+
+    def factory(self, config, section):
+        pmapper = PolynomialMapper()
+        coeffs = self.config_get(config, section, 'coefficients')
+        pmapper.parse_coefficient_string(coeffs)
+        pmapper.output_low = self.config_get(config, section, 'output_low', cast='float')
+        pmapper.output_high = self.config_get(config, section, 'output_high', cast='float')
+
+        return pmapper
 
 
 class ADCDevice(AbstractDevice, PolynomialMapperMixin):
