+++ pychron/hardware/__init__.py
@@ -71,7 +71,9 @@ HW_PACKAGE_MAP = {
     'GenericDevice': 'pychron.hardware.generic_device',
 
     'T4Actuator': 'pychron.hardware.actuators.t4_actuator',
-    'U3Actuator': 'pychron.hardware.actuators.u3_actuator'
+    'U3Actuator': 'pychron.hardware.actuators.u3_actuator',
+    'ProXRActuator': 'pychron.hardware.actuators.proxr_actuator',
+    'U3GaugeController': 'pychron.hardware.labjack.u3_gauge_controller'
 
 }
 
