+++ pychron/hardware/actuators/u3_actuator.py
@@ -20,8 +20,7 @@ from pychron.hardware.labjack.base_u3_lv import BaseU3LV
 
 class U3Actuator(BaseU3LV, GPActuator):
     def _actuate(self, obj, action):
-        self.set_channel_state(get_switch_address(obj), action.lower() == 'open')
-        return True
+        return BaseU3LV.set_channel_state(self, get_switch_address(obj), action.lower() == 'open')
 
     def get_channel_state(self, obj, **kw):
         return BaseU3LV.get_channel_state(self, get_switch_address(obj), **kw)
