+++ pychron/hardware/labjack/base_u3_lv.py
@@ -22,13 +22,16 @@ try:
 except ImportError:
     NullHandleException = BaseException
 
-
+import re
 # ============= local library imports  ==========================
+
+DIORE = re.compile(r'^[FCE]IO\d+$')
 
 
 class BaseU3LV:
     _device = None
     _dio_mapping = None
+    _dio_channels = None
 
     def load(self, *args, **kw):
         self._dio_mapping = {}
@@ -61,14 +64,25 @@ class BaseU3LV:
             for option in config.options(section):
                 u3channel = config.get(section, option)
                 mapping[option] = getattr(u3, u3channel)
-        self._dio_mapping = mapping
+            self._dio_mapping = mapping
+
+        elif config.has_section('DIOConfig'):
+            channellist = config.get('DIOConfig', 'channellist')
+            self._dio_channels = [getattr(u3, c.strip()) for c in channellist.split(',')]
 
         return True
 
     def initialize(self, *args, **kw):
+        chs = None
+        if self._dio_mapping:
+            chs = [v for v in self._dio_mapping.values()]
+        elif self._dio_channels:
+            chs = self._dio_channels
 
-        chs = [v for v in self._dio_mapping.values() if v not in (u3.CIO0, u3.CIO1, u3.CIO2, u3.CIO3)]
-        self._device.configDigital(*chs)
+        if chs:
+            chs = [v for v in chs if v not in (u3.CIO0, u3.CIO1, u3.CIO2, u3.CIO3)]
+            print('configuring {}'.format(chs))
+            self._device.configDigital(*chs)
 
         return True
 
@@ -80,13 +94,18 @@ class BaseU3LV:
         @return:
         """
         pin = self._get_pin(ch)
+        print('set channel state {} {} state={}'.format(ch, pin, state))
         if pin is not None:
             self._device.setDOState(pin, int(not state))
+            return True
 
     def get_channel_state(self, ch):
         pin = self._get_pin(ch)
+        print('get chanel state {} {}'.format(ch, pin))
         if pin is not None:
-            return self._device.getDIOState(pin)
+            r = self._device.getDIOState(pin)
+            print('got channel state {} {}'.format(r, not r))
+            return not r
 
     # def read_dac_channel(self, ch):
     #     v = self._device.getFIOState(ch)
@@ -111,10 +130,14 @@ class BaseU3LV:
 
     # private
     def _get_pin(self, ch):
-        try:
-            return self._dio_mapping[str(ch)]
-        except KeyError:
-            self.warning('Invalid channel {}'.format(ch))
-            self.warning('DIOMapping {}'.format(self._dio_mapping))
+        ch = str(ch)
+        if DIORE.match(ch):
+            return getattr(u3, ch)
+        else:
+            try:
+                return self._dio_mapping[ch]
+            except KeyError:
+                self.warning('Invalid channel {}'.format(ch))
+                self.warning('DIOMapping {}'.format(self._dio_mapping))
 
 # ============= EOF =============================================
