+++ pychron/dvc/__init__.py
@@ -163,7 +163,6 @@ def _analysis_path(runid, repository, modifier=None, extension='.json', mode='r'
         if not os.path.isdir(root):
             os.mkdir(root)
 
-    print(runid)
     if UUID_RE.match(runid):
         sublen = 2, 5
     elif WISCAR_ID_RE.match(runid):
