### 2021-01-21 21a9bd3
implemented rescale for spectrum
