### 2021-01-22 dc67172
fixed persistence spec settings added force_sublen to force new analyses to use sublen=2
