+++ pychron/pychron_constants.py
@@ -254,8 +254,9 @@ QTEGRA_INTEGRATION_TIMES = [0.065536, 0.131072, 0.262144, 0.524288,
 
 QTEGRA_DEFAULT_INTEGRATION_TIME = 1.048576
 ISOTOPX_INTEGRATION_TIMES = [1.0, 2.0, 3.0, 4.0, 5.0, 10.0, 20.0, 100.0]
+QUADERA_INTEGRATION_TIMES = [1.0,]
 ISOTOPX_DEFAULT_INTEGRATION_TIME = 1
-
+QUADERA_DEFAULT_INTEGRATION_TIME = 1
 DEFAULT_INTEGRATION_TIME = 1
 
 K_DECAY_CONSTANTS = {'Min et al., 2000': (5.80e-11, 0.099e-10, 4.883e-10, 0.014e-10),
