+++ pychron/hardware/fiber_light.py
@@ -48,9 +48,9 @@ class FiberLight(AbstractDevice):
 
         self._cdevice = None
         if klass is not None:
-            if klass=='ArduinoFiberLightModule':
+            if klass == 'ArduinoFiberLightModule':
                 package = 'pychron.hardware.arduino.arduino_fiber_light_module'
-            elif klass=='U3FiberLight':
+            elif klass == 'U3FiberLight':
                 package = 'pychron.hardware.labjack.u3_fiber_light'
             factory = self.get_factory(package, klass)
             self._cdevice = factory(name=klass,
@@ -66,7 +66,7 @@ class FiberLight(AbstractDevice):
 
     def _write_intensity(self, v):
         if self._cdevice:
-            self._cdevice.set_intensity(v/100)
+            self._cdevice.set_intensity(v / 100)
 
     def read_state(self):
         if self._cdevice is not None:

[…]
