### 2021-01-26 7f4bf18
save standard ratios when doing ic factor fit
