### 2021-01-26 0aff10a
fixed dump icfactor
