+++ pychron/hardware/zaber/zaber_motion_controller.py
@@ -14,56 +14,55 @@
 # limitations under the License.
 # ===============================================================================
 from traits.api import Str, CInt, Enum
+from zaber.serial import BinarySerial, BinaryDevice
+# from zaber_motion import Library
 
-from zaber_motion import Library
-
-Library.enable_device_db_store()
+# Library.enable_device_db_store()
 
 from pychron.hardware.motion_controller import MotionController
 from pychron.hardware.zaber.axis import ZaberAxis
-from zaber_motion import Units
 
 
-class AsciiWrapper:
-    def __init__(self, devices):
-        self._device = devices[0]
+# from zaber_motion import Units
+#
+#
+# class AsciiWrapper:
+#     def __init__(self, devices):
+#         self._device = devices[0]
+#
+#     def get_axis(self, a):
+#         return self._device.get_axis(a)
+#
+#     def is_busy(self):
+#         return self._device.all_axes.is_busy()
+#
+#
+# class BinaryWrapper:
+#     def __init__(self, devices):
+#         self._devices = devices
+#
+#     def get_axis(self, a):
+#         return self._devices[a]
+#
+#     def is_busy(self):
+#         return any((d.is_busy() for d in self._devices))
 
-    def get_axis(self, a):
-        return self._device.get_axis(a)
-
-    def is_busy(self):
-        return self._device.all_axes.is_busy()
-
-
-class BinaryWrapper:
-    def __init__(self, devices):
-        self._devices = devices
-
-    def get_axis(self, a):
-        return self._devices[a]
-
-    def is_busy(self):
-        return any((d.is_busy() for d in self._devices))
-
-
-class ZaberMotionController(MotionController):
+class LegacyBinaryZaberMotionController(MotionController):
     port = Str
     device_id = CInt
     baudrate = CInt
-    protocol = Enum('ASCII', 'Binary')
     _device = None
-    _units = Units.LENGTH_MILLIMETRES
-    _wrapper = None
-    move_enabled = False
+    # _units = Units.LENGTH_MILLIMETRES
+    # _wrapper = None
 
     def load(self, *args, **kw):
+
         config = self.get_configuration()
         if config:
             section = 'Communications'
             self.set_attribute(config, 'port', section, 'port')
             self.set_attribute(config, 'device_id', section, 'device_id', optional=True, default=0)
             self.set_attribute(config, 'baudrate', section, 'baudrate', optional=True, default=9600)
-            self.set_attribute(config, 'protocol', section, 'protocol', optional=True, default='ASCII')
 
             self.axes_factory(config)
 
@@ -74,20 +73,11 @@ class ZaberMotionController(MotionController):
 
     def open(self, *args, **kw):
         self.debug('scanning port {}'.format(self.port))
-        if self.protocol == 'ASCII':
-            from zaber_motion.ascii import Connection
-        else:
-            from zaber_motion.binary import Connection
+        bs = BinarySerial(self.port)
 
-        connection = Connection.open_serial_port(self.port, self.baudrate)
-        device_list = connection.detect_devices()
-        self.debug(device_list)
-        if self.protocol == 'ASCII':
-            klass = AsciiWrapper
-        else:
-            klass = BinaryWrapper
+        for a in self.axes.values():
+            a.device = BinaryDevice(bs, a.device_id)
 
-        self._wrapper = klass(device_list)
         return True
 
     def _get_simulation(self):
@@ -96,54 +86,144 @@ class ZaberMotionController(MotionController):
     def linear_move(self, x, y, block=True, *args, **kw):
         xaxis = self.single_axis_move('x', x, block=False)
         yaxis = self.single_axis_move('y', y, block=False)
+        return
 
-        if block:
-            if self.protocol == 'ASCII':
-                xaxis.wait_until_idle()
[…]
