### 2021-01-27 #2919 @pvermees
I am confused. Have I been formally invited to review this paper, or have I simply been added to the GitHub page as an observer? I must confess that I am not very familiar with JOSS. Is it common for this journal to publish 2-page papers? Does such a short paper even require peer review? From [https://github.com/openjournals/joss](https://github.com/openjournals/joss) it appears that the review process is a formality. Would it suffice if I confirm that Pychron is a top notch piece of software? @jirhiker certainly deserves to get a DOI for his work!
