### 2021-01-27 #2919 @danielskatz
👋 @pvermees - @jirhiker suggested you as a potential reviewer (which he should have done without the `@` so that you wouldn't have been notified). The next step is for the editor (@richardjgowers in this case) to invite you (and/or others).

Re: JOSS review criteria, see https://joss.readthedocs.io/en/latest/review_criteria.html.  In brief, the review is of the software as much as or more than the short paper.  The process is most definitely not 'a formality', except in the case of software that has previously been carefully reviewed by rOpenSci.
