### 2021-01-29 b13155d
removed github authentication for windows
