### 2021-01-29 dc82507
removed kiva from pychron_constants.py
