### 2021-02-02 7280cec
fixed conditional checking for blanks
