+++ pychron/git_archive/repo_manager.py
@@ -817,8 +817,7 @@ class GitRepoManager(Loggable):
                             pass
 
                         if self.confirmation_dialog('There appears to be a problem with {}.'
-                                                    '\n\nWould you like to accept the master copy'.format(self.name)):
-
+                                                                            '\n\nWould you like to accept the master copy'.format(self.name)):
                             try:
                                 repo.git.pull('-X', 'theirs', '--commit', '--no-edit')
                                 return True
@@ -827,7 +826,7 @@ class GitRepoManager(Loggable):
                                 if clean:
                                     if self.confirmation_dialog('''You have untracked files that could be an issue. 
 {}
- 
+
 You like to delete them and try again?'''.format(clean)):
                                         try:
                                             repo.git.clean('-fd')
@@ -843,9 +842,45 @@ You like to delete them and try again?'''.format(clean)):
                                 else:
                                     self.warning_dialog('Failed pulling changes for {}'.format(self.name))
                                 return
+
                         else:
-                            return
-                self._resolve_conflicts(branch, remote, accept_our, accept_their, quiet)
+                            self._resolve_conflicts(branch, remote, accept_our, accept_their, quiet)
+
+                        # try:
+                        #     repo.git.merge('--abort')
+                        # except GitCommandError:
+                        #     pass
+#
+#                         if self.confirmation_dialog('There appears to be a problem with {}.'
+#                                                     '\n\nWould you like to accept the master copy'.format(self.name)):
+#
+#                             try:
+#                                 repo.git.pull('-X', 'theirs', '--commit', '--no-edit')
+#                                 return True
+#                             except GitCommandError:
+#                                 clean = repo.git.clean('-n')
+#                                 if clean:
+#                                     if self.confirmation_dialog('''You have untracked files that could be an issue.
+# {}
+#
+# You like to delete them and try again?'''.format(clean)):
+#                                         try:
+#                                             repo.git.clean('-fd')
+#                                         except GitCommandError:
+#                                             self.warning_dialog('Failed to clean repository')
+#                                             return
+#
+#                                         try:
+#                                             repo.git.pull('-X', 'theirs', '--commit', '--no-edit')
+#                                             return True
+#                                         except GitCommandError:
+#                                             self.warning_dialog('Failed pulling changes for {}'.format(self.name))
+#                                 else:
+#                                     self.warning_dialog('Failed pulling changes for {}'.format(self.name))
+#                                 return
+#                         else:
+#                             return
+#                 self._resolve_conflicts(branch, remote, accept_our, accept_their, quiet)
             else:
                 self.debug('merging {} commits'.format(behind))
                 self._git_command(lambda g: g.merge('FETCH_HEAD'), 'GitRepoManager.smart_pull/!ahead')
@@ -1017,7 +1052,7 @@ You like to delete them and try again?'''.format(clean)):
     # private
     def _resolve_conflicts(self, branch, remote, accept_our, accept_their, quiet):
         conflict_paths = self._get_conflict_paths()
-        self.debug('conflict_paths: {}'.format(conflict_paths))
+        self.debug('resolve conflict_paths: {}'.format(conflict_paths))
         if conflict_paths:
             mm = MergeModel(conflict_paths,
                             branch=branch,
@@ -1029,7 +1064,7 @@ You like to delete them and try again?'''.format(clean)):
                 mm.accept_their()
             else:
                 mv = MergeView(model=mm)
-                mv.edit_traits()
+                mv.edit_traits(kind='livemodal')
         else:
             if not quiet:
                 self.information_dialog('There were no conflicts identified')
[…]
