### 2021-02-03 f702b82
added repeat run to queue mod conditional
