+++ pychron/experiment/conditional/conditional.py
@@ -430,7 +430,9 @@ class ActionConditional(AutomatedRunConditional):
             action()
 
 
-MODIFICATION_ACTIONS = ('Skip Next Run', 'Skip N Runs', 'Skip Aliquot', 'Skip to Last in Aliquot', 'Set Extract')
+MODIFICATION_ACTIONS = ('Skip Next Run', 'Skip N Runs', 'Skip Aliquot',
+                        'Skip to Last in Aliquot', 'Set Extract',
+                        'Repeat Run')
 
 
 class ExtractionStr(BaseStr):
@@ -471,10 +473,11 @@ class QueueModificationConditional(AutomatedRunConditional):
     action = Enum(MODIFICATION_ACTIONS)
     extraction_str = ExtractionStr
 
-    def do_modifications(self, queue, current_run):
+    def do_modifications(self, current_run, executor, queue):
         runs = queue.cleaned_automated_runs
         func = getattr(self, self.action.lower().replace(' ', '_'))
-        func(runs, current_run)
+        if func(queue, runs, current_run):
+            executor.queue_modified = True
         queue.refresh_table_needed = True
 
     def _from_dict_hook(self, cd):
@@ -482,7 +485,12 @@ class QueueModificationConditional(AutomatedRunConditional):
             if tag in cd:
                 setattr(self, tag, cd[tag])
 
-    def _skip_n_runs(self, runs, current_run, n=None):
+    def _repeat_run(self, queue, runs, current_run):
+        spec = current_run.spec.tocopy()
+        queue.automated_runs.insert(spec, 0)
+        return True
+
+    def _skip_n_runs(self, queue, runs, current_run, n=None):
         if n is None:
             n = self.nskip
 
@@ -490,10 +498,10 @@ class QueueModificationConditional(AutomatedRunConditional):
             r = runs[i]
             r.skip = True
 
-    def _skip_next_run(self, runs, current_run):
+    def _skip_next_run(self, queue, runs, current_run):
         self._skip_n_runs(runs, current_run, 1)
 
-    def _skip_aliquot(self, runs, current_run):
+    def _skip_aliquot(self, queue, runs, current_run):
 
         identifier = current_run.spec.identifier
         aliquot = current_run.spec.aliquot
@@ -504,7 +512,7 @@ class QueueModificationConditional(AutomatedRunConditional):
             if r.identifier == identifier and r.aliquot == aliquot:
                 r.skip = True
 
-    def _skip_to_last_in_aliquot(self, runs, current_run):
+    def _skip_to_last_in_aliquot(self, queue, runs, current_run):
         identifier = current_run.spec.identifier
         for i, r in enumerate(runs):
             if r.is_special():
@@ -520,7 +528,7 @@ class QueueModificationConditional(AutomatedRunConditional):
             except IndexError:
                 pass
 
-    def _set_extract(self, runs, current_run):
+    def _set_extract(self, queue, runs, current_run):
 
         es = self.extraction_str
 
