### 2021-02-04 c84a5d5
added auto linear vs parabolic
