+++ pychron/pychron_constants.py
@@ -97,9 +97,10 @@ INTEGRATED = 'Total Integrated'
 DEFAULT_INTEGRATED = '{} else Valid Integrated'.format(PLATEAU)
 VALID_INTEGRATED = 'Valid Integrated'
 PLATEAU_INTEGRATED = '{} Integrated'.format(PLATEAU)
-
+AUTO_LINEAR_PARABOLIC = 'Auto Linear/Parabolic'
+AUTO_N = 'Auto N'
 FIT_TYPES = ['Linear', 'Parabolic', 'Cubic',
-             'Average', 'Exponential', WEIGHTED_MEAN, 'Custom']
+             'Average', 'Exponential', WEIGHTED_MEAN, 'Custom', AUTO_LINEAR_PARABOLIC]
 
 FIT_ERROR_TYPES = [SD, SEM, MSEM, 'CI', 'MonteCarlo']
 SERIES_FIT_TYPES = [NULL_STR] + FIT_TYPES
