+++ pychron/pipeline/results/iso_evo.py
@@ -18,8 +18,8 @@ from traits.api import Str, Float
 from pychron.core.helpers.formatting import floatfmt
 from pychron.pipeline.results.base import BaseResult
 
-GOODNESS_TAGS = ('int_err', 'slope', 'outlier', 'curvature', 'rsquared', 'signal_to_blank')
-GOODNESS_NAMES = ('Intercept Error', 'Slope', 'Outliers', 'Curvature', 'RSquared', 'Blank/Signal %')
+GOODNESS_TAGS = ('int_err', 'slope', 'outlier', 'curvature', 'rsquared', 'signal_to_blank', 'signal_to_baseline')
+GOODNESS_NAMES = ('Intercept Error', 'Slope', 'Outliers', 'Curvature', 'RSquared', 'Blank/Signal %', 'Baseline/Signal %')
 INVERTED_GOODNESS = ('rsquared',)
 
 
@@ -36,6 +36,7 @@ class IsoEvoResult(BaseResult):
     curvature_goodness = None
     rsquared_goodness = None
     signal_to_blank_goodness = None
+    signal_to_baseline_goodness = None
 
     int_err_threshold = None
     slope_threshold = None
@@ -44,6 +45,7 @@ class IsoEvoResult(BaseResult):
     rsquared_threshold = None
     signal_to_blank_threshold = None
     signal_to_baseline_threshold = None
+    signal_to_baseline_percent_threshold = None
 
     int_err = None
     slope = None
