### 2021-02-08 d50b439
added ability to checkout branches you didn't create locally
