+++ pychron/hardware/watlow/base_ezzone.py
@@ -31,6 +31,42 @@ from pychron.hardware.watlow import sensor_map, tc_map, itc_map, isensor_map, he
 from six.moves import zip
 
 
+class Protocol:
+    def get_register(self, key):
+        return self.mapping[key]
+
+    def read(self, *args, **kw):
+        pass
+
+    def write(self, *args, **kw):
+        pass
+
+
+class StandardProtocol(Protocol):
+    mapping = {'baud_rate': [17002, 17002],
+               'csp': 7001,
+               'osp': 7002,
+               'cm': 8001,
+               'hag': 8003,
+               'cpb': 8012,
+               'ti': 8006,
+               'td': 8007,
+               'db': 8008}
+
+
+class ModbusProtocol(Protocol):
+    mapping = {'baud_rate': [2484, 2504],
+               'csp': 2160,
+               'osp': 2162,
+               'cm': 1880,
+               'hag': 1884,
+               'hpb': 1890,
+               'cpb': 1892,
+               'ti': 1894,
+               'td': 1896,
+               'db': 1898}
+
+
 class BaseWatlowEZZone(HasTraits):
     """
         WatlowEZZone represents a WatlowEZZone PM PID controller.
@@ -165,6 +201,8 @@ class BaseWatlowEZZone(HasTraits):
     min_output_scale = Float
     max_output_scale = Float
 
+    use_modbus = Bool(True)
+
     # def _get_use_calibrated_temperature(self):
     #    return self._use_calibrated_temperature and self.calibration is not None
 
@@ -217,6 +255,11 @@ class BaseWatlowEZZone(HasTraits):
         return te
 
     def initialize(self, *args, **kw):
+        if self.use_modbus:
+            self.protocol = ModbusProtocol()
+        else:
+            self.protocol = StandardProtocol()
+
         if self.read_baudrate(port=1):
             # set open loop and closed loop to zero
             self.disable()
@@ -270,9 +313,18 @@ class BaseWatlowEZZone(HasTraits):
             self._process_memory_len += 2
 
     def report_pid(self):
-        pid_attrs = ['_Ph_', '_Pc_', '_I_', '_D_']
+
         self.info('read pid parameters')
-        pid_vals = self.read(1890, nregisters=8, nbytes=21, response_type='float')
+        if self.use_modbus:
+            pid_attrs = ['_Ph_', '_Pc_', '_I_', '_D_']
+            pid_vals = self.read(1890, nregisters=8, nbytes=21, response_type='float')
+        else:
+            ph = self.read_heat_proportional_band()
+            pc = self.read_cool_proportional_band()
+            i = self.read_time_integral()
+            d = self.read_time_derivative()
+            pid_vals = (ph, pc, i, d)
+
         if pid_vals:
             self.info('======================== PID =====================')
             for pa, pv in zip(pid_attrs, pid_vals):
@@ -288,6 +340,7 @@ class BaseWatlowEZZone(HasTraits):
             hdb = self.read_heat_dead_band()
             self.info('heat dead band: {}'.format(hdb))
             self.info('==================================================')
+
         return pid_vals
 
     def initialization_hook(self):
@@ -425,6 +478,9 @@ class BaseWatlowEZZone(HasTraits):
     def load_additional_args(self, config):
         """
         """
+
+        self.set_attribute(config, 'use_modbus', 'Communications', 'use_modbus', cast='boolean', default=True)
+
         self.set_attribute(config, 'use_pid_bin', 'Output', 'use_pid_bin', cast='boolean', default=False)
         self.set_attribute(config, 'min_output_scale', 'Output', 'scale_low', cast='float')
[…]
