+++ pychron/hardware/watlow/base_ezzone.py
@@ -48,10 +48,23 @@ class StandardProtocol(Protocol):
                'osp': 7002,
                'cm': 8001,
                'hag': 8003,
+               'hpb': 8009,
                'cpb': 8012,
                'ti': 8006,
                'td': 8007,
-               'db': 8008}
+               'db': 8008,
+               'sen1': 4005,
+               'sen2': 4005,  # sen1 not defined in docs
+               'tc1': 4006,
+               'tc2': 4006,
+               'hhy': 8010,
+               'atsp': 8025,
+               'tagr': 8024,
+               'ttun': 8022,
+               'tbnd': 8034,
+               'tgn': 8035,
+               'fn': 18002
+               }
 
 
 class ModbusProtocol(Protocol):
@@ -64,7 +77,19 @@ class ModbusProtocol(Protocol):
                'cpb': 1892,
                'ti': 1894,
                'td': 1896,
-               'db': 1898}
+               'db': 1898,
+               'sen1': 368,
+               'sen2': 448,
+               'tc1': 370,
+               'tc2': 450,
+               'hhy': 1900,
+               'atsp': 1918,
+               'tagr': 1916,
+               'ttun': 8022,
+               'tbnd': 1912,
+               'tgn': 1914,
+               'fn': 722
+               }
 
 
 class BaseWatlowEZZone(HasTraits):
@@ -315,8 +340,9 @@ class BaseWatlowEZZone(HasTraits):
     def report_pid(self):
 
         self.info('read pid parameters')
+        pid_attrs = ['_Ph_', '_Pc_', '_I_', '_D_']
+
         if self.use_modbus:
-            pid_attrs = ['_Ph_', '_Pc_', '_I_', '_D_']
             pid_vals = self.read(1890, nregisters=8, nbytes=21, response_type='float')
         else:
             ph = self.read_heat_proportional_band()
@@ -489,13 +515,7 @@ class BaseWatlowEZZone(HasTraits):
         self.set_attribute(config, 'setpointmax', 'Setpoint', 'max', cast='float')
 
         self.set_attribute(config, 'memory_blocks_enabled', 'MemoryBlock', 'enabled', cast='boolean')
-        if self.memory_blocks_enabled:
-            self.set_attribute(config, 'program_memory_blocks', 'MemoryBlock', 'program', cast='boolean')
-            #            if self.program_memory_blocks:
-        #                self.memory_blocks=[]
-        #                for option in config.options('MemoryBlock'):
-        #                    if option.startswith('block'):
-        #                        self.memory_blocks.append(config.get('MemoryBlock',option))
+        self.set_attribute(config, 'program_memory_blocks', 'MemoryBlock', 'program', cast='boolean')
 
         coeffs = self.config_get(config, 'Calibration', 'coefficients')
         if coeffs:
@@ -760,7 +780,7 @@ class BaseWatlowEZZone(HasTraits):
             A positive value keeps heating and cooling outputs from fighting each other.
         """
         self.info('setting dead_band = {:0.3f}'.format(v))
-        register =self.protocol.get_register('db')
+        register = self.protocol.get_register('db')
         self.write(register, v, nregisters=2, **kw)
 
     # ===============================================================================
@@ -792,7 +812,7 @@ class BaseWatlowEZZone(HasTraits):
 
     def set_analog_input_sensor_type(self, input_id, value, **kw):
         self.info('set input sensor type {}'.format(value))
-        register = 368 if input_id == 1 else 448
+        register = self.protocol.get_register('sen{}'.format(input_id))
         v = value if isinstance(value, int) else isensor_map[value]
         self.write(register, v, **kw)
         if v == 95:
@@ -827,7 +847,8 @@ class BaseWatlowEZZone(HasTraits):
[…]
