+++ pychron/pipeline/engine.py
@@ -55,7 +55,7 @@ from pychron.pipeline.pipeline_defaults import ISOEVO, BLANKS, ICFACTOR, IDEO, S
     MASSSPEC_REDUCED, DEFINE_EQUILIBRATION, CA_CORRECTION_FACTORS, K_CORRECTION_FACTORS, \
     FLUX_VISUALIZATION, CSV_RAW_DATA_EXPORT, COMPOSITE, SIMPLE_ANALYSIS_TABLE, MASS_SPEC_FLUX, PYSCRIPT, RATIO_SERIES, \
     CSV_SPEC, ARAR_IDEO, ARAR_SPEC, ARAR_INVERSE_ISOCHRON, ARAR_SIMPLE_ANALYSIS_TABLE, RUNID_EDIT, CORRELATION_IDEO, \
-    COSMOGENIC
+    COSMOGENIC, FLUX_EXPORT
 from pychron.pipeline.plot.editors.figure_editor import FigureEditor
 from pychron.pipeline.plot.editors.ideogram_editor import IdeogramEditor
 from pychron.pipeline.plot.editors.spectrum_editor import SpectrumEditor
@@ -895,7 +895,8 @@ class PipelineEngine(Loggable):
                    ('History', (('Ideogram', HISTORY_IDEOGRAM),
                                 ('Spectrum', HISTORY_SPECTRUM))),
                    ('Share', (('CSV Analyses Export', CSV_ANALYSES_EXPORT),
-                              ('CSV Raw Data Export', CSV_RAW_DATA_EXPORT))),
+                              ('CSV Raw Data Export', CSV_RAW_DATA_EXPORT),
+                              ('CSV Flux Export', FLUX_EXPORT))),
                    ('Transfer', (('Mass Spec Reduced', MASSSPEC_REDUCED),
                                  ('Mass Spec Flux', MASS_SPEC_FLUX))),
                    ('Scripting', (('PyScript', PYSCRIPT),))]
