+++ pychron/pipeline/nodes/__init__.py
@@ -45,10 +45,10 @@ from pychron.pipeline.nodes.grouping import GroupingNode, GraphGroupingNode, Sub
 from pychron.pipeline.nodes.ia import SetInterpretedAgeNode
 from pychron.pipeline.nodes.mass_spec_reduced import MassSpecReducedNode, MassSpecFluxNode
 from pychron.pipeline.nodes.ml import MLDataNode, MLRegressionNode
-from pychron.pipeline.nodes.persist import CosmogenicCorrectionPersistNode
 from pychron.pipeline.nodes.persist import DVCPersistNode, PDFFigureNode, \
     BlanksPersistNode, IsotopeEvolutionPersistNode, ICFactorPersistNode, FluxPersistNode, \
-    XLSXAnalysisTablePersistNode, InterpretedAgePersistNode, DefineEquilibrationPersistNode
+    XLSXAnalysisTablePersistNode, InterpretedAgePersistNode, DefineEquilibrationPersistNode, \
+    FluxMonitorMeansPersistNode, CosmogenicCorrectionPersistNode
 from pychron.pipeline.nodes.push import PushNode
 from pychron.pipeline.nodes.report import ReportNode
 from pychron.pipeline.nodes.review import ReviewNode
