### 2021-02-15 d306333
fixed show isotope regression popup from recall
