### 2021-02-15 b7b6875
added pyrometer mixin
