+++ pychron/lasers/laser_managers/fusions_laser_manager.py
@@ -15,34 +15,17 @@
 # ===============================================================================
 
 # =============enthought library imports=======================
-from __future__ import absolute_import
-from __future__ import print_function
 from apptools.preferences.preference_binding import bind_preference
-from traits.api import DelegatesTo, Property, Instance, Str, List, Dict, \
-    on_trait_change, Event, Bool, Any, Button
-from traitsui.api import VGroup, Item, HGroup, spring
+from traits.api import DelegatesTo, Instance, Str, List, Dict, \
+    on_trait_change, Event, Bool, Any
+from traitsui.api import HGroup, spring, UReadonly, Readonly
 
-from .laser_manager import LaserManager
 from pychron.core.helpers.strtools import to_bool
-from pychron.core.ui.thread import Thread
+from pychron.core.pychron_traits import BorderVGroup
 from pychron.hardware.fiber_light import FiberLight
 from pychron.hardware.fusions.fusions_logic_board import FusionsLogicBoard
-from pychron.mv.degas.degasser import Degasser
+from pychron.lasers.laser_managers.laser_manager import LaserManager
 from pychron.response_recorder import ResponseRecorder
-
-
-# from pychron.core.ui.gui import invoke_in_main_thread
-# from pychron.lasers.laser_managers.degas_manager import DegasManager
-
-
-# class OpticsView(Viewable):
-# #    title = 'Optics'
-#    laser_controller = Instance(FusionsLogicBoard)
-#    def traits_view(self):
-#        item = Item('laser_controller', show_label=False,
-#                    editor=InstanceEditor(view='control_view'),
-#                    style='custom')
-#        return self.view_factory(item, id='optics_view')
 
 
 class FusionsLaserManager(LaserManager):
@@ -52,32 +35,6 @@ class FusionsLaserManager(LaserManager):
     laser_controller = Instance(FusionsLogicBoard)
     fiber_light = Instance(FiberLight)
     response_recorder = Instance(ResponseRecorder)
-
-    #    optics_view = Instance(OpticsView)
-
-    #    beam = DelegatesTo('laser_controller')
-    #    beammin = DelegatesTo('laser_controller')
-    #    beammax = DelegatesTo('laser_controller')
-    #    update_beam = DelegatesTo('laser_controller')
-    #    beam_enabled = DelegatesTo('laser_controller')
-    # #    beam_enabled = Bool(True)
-    #
-    #    zoom = DelegatesTo('laser_controller')
-    #    zoommin = DelegatesTo('laser_controller')
-    #    zoommax = DelegatesTo('laser_controller')
-    #    update_zoom = DelegatesTo('laser_controller')
-    #    execute_button = DelegatesTo('laser_script_executor')
-    #    execute_label = DelegatesTo('laser_script_executor')
-
-    pointer = Event
-    pointer_state = Bool(False)
-    pointer_label = Property(depends_on='pointer_state')
-
-    step_heat_manager = None
-
-    lens_configuration = Str('gaussian')
-    lens_configuration_dict = Dict
-    lens_configuration_names = List
 
     power_timer = None
     brightness_timer = None
@@ -97,8 +54,6 @@ class FusionsLaserManager(LaserManager):
     _data_manager_lock = None
 
     _current_rid = None
-
-    #    brightness_meter = Instance(BrightnessPIDManager)
 
     chiller = Any
 
@@ -134,17 +89,6 @@ class FusionsLaserManager(LaserManager):
     def open_motor_configure(self):
         self.laser_controller.open_motor_configure()
 
-    #     def _record_fired(self):
-    #         if self._recording_power_state:
-    #             save = self.db_save_dialog()
-    #             self.stop_power_recording(delay=0, save=save)
-    #         else:
[…]
