### 2021-02-17 96d1227
added include all positions to csv flux export
