### 2021-02-18 cdd67e1
fixed weighted mean MSEM error calculation. Was doing SEM *sqrt(MSWD) instead of SE*sqrt(MSWD). SE is the Taylor error aka sum of weights, SEM is STD/sqrt(n)
