+++ pychron/pipeline/editors/flux_results_editor.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ===============================================================================
 from operator import attrgetter
-
+import csv
 # ============= enthought library imports =======================
 from numpy import average, array, diff, arctan, Inf
 from scipy.stats import mode
@@ -24,10 +24,12 @@ from traitsui.extras.checkbox_column import CheckboxColumn
 from traitsui.table_column import ObjectColumn
 from uncertainties import nominal_value, std_dev
 
+from pychron.core.helpers.filetools import add_extension
 from pychron.core.helpers.formatting import calc_percent_error, floatfmt
 from pychron.core.helpers.iterfuncs import groupby_key
 from pychron.core.stats import validate_mswd
 from pychron.envisage.icon_button_editor import icon_button_editor
+from pychron.paths import paths
 from pychron.pipeline.editors.flux_visualization_editor import BaseFluxVisualizationEditor
 from pychron.pipeline.plot.plotter.arar_figure import SelectionFigure
 from pychron.processing.flux import mean_j
@@ -92,7 +94,10 @@ MONITOR_COLUMNS = [
            label='%',
            format_func=ff)]
 
-MONITOR_EDITOR = TableEditor(columns=MONITOR_COLUMNS, sortable=False, reorderable=False)
+MONITOR_EDITOR = TableEditor(columns=MONITOR_COLUMNS,
+                             selected='selected_monitors',
+                             selection_mode='rows',
+                             sortable=False, reorderable=False)
 
 
 class FluxPosition(HasTraits):
@@ -167,10 +172,11 @@ class FluxPosition(HasTraits):
 
 
 class FluxResultsEditor(BaseFluxVisualizationEditor, SelectionFigure):
-    save_all_button = Event
-    save_unknowns_button = Event
+    selected_monitors = List
+    toggle_use_button = Button('Toggle Use')
+    toggle_save_button = Button('Toggle Save')
+    save_monitor_flux_csv_button = Button('Monitor Fluxes CSV')
     recalculate_button = Button('Calculate')
-
     suppress_metadata_change = Bool(False)
     _analyses = List
 
@@ -183,9 +189,6 @@ class FluxResultsEditor(BaseFluxVisualizationEditor, SelectionFigure):
         # if self.geometry:
         #     self.set_positions(analyses)
         #     self.predict_values()
-
-    def _recalculate_button_fired(self):
-        self.predict_values()
 
     def set_positions(self, monitors, unk=None):
         self.debug('setting positions mons={}, unks={}'.format(len(monitors), len(unk) if unk else 0))
@@ -208,7 +211,7 @@ class FluxResultsEditor(BaseFluxVisualizationEditor, SelectionFigure):
             vs = abs(diff(vs))
             vs = vs[vs.astype(bool)].mean()
         else:
-            vs = [p[1]/p[0] if p[0] else Inf for p in geom]
+            vs = [p[1] / p[0] if p[0] else Inf for p in geom]
             vs = arctan(vs)
 
             vs = abs(diff(vs))
@@ -290,6 +293,62 @@ class FluxResultsEditor(BaseFluxVisualizationEditor, SelectionFigure):
 
         self.predict_values(refresh=True)
 
+    def _save_monitor_flux_csv(self, p=None):
+        if p is None:
+            p = self.save_file_dialog(default_filename='{}{}_MonitorFluxes.csv'.format(self.irradiation, self.level),
+                                      default_directory=paths.data_dir)
+
+        if p:
+            p = add_extension(p, '.csv')
+            attrs = ['hole_id',
+                     'identifier',
+                     'sample',
+                     'n',
+                     'saved_j',
+                     'saved_jerr',
+                     'percent_saved_error',
+                     'mean_j',
[…]
