+++ pychron/pipeline/pipeline_defaults.py
@@ -325,8 +325,8 @@ FLUX = """
 required:
 nodes:
   - klass: FindFluxMonitorsNode
-#    irradiation: NM-300
-#    level: E
+    irradiation: NM-312
+    level: P
   - klass: FluxMonitorsNode
   - klass: FitFluxNode
   - klass: ReviewNode
