+++ pychron/core/regression/flux_regressor.py
@@ -17,8 +17,9 @@
 # ============= standard library imports ========================
 from operator import itemgetter
 
-from numpy import asarray, column_stack, ones_like, array, average, ravel
+from numpy import asarray, column_stack, ones_like, array, average, ravel, zeros_like
 # ============= local library imports  ==========================
+from scipy.interpolate import Rbf, bisplrep, bisplev, griddata
 from statsmodels.regression.linear_model import WLS, OLS
 # ============= enthought library imports =======================
 from traits.api import Bool, Int
@@ -26,6 +27,7 @@ from traits.api import Bool, Int
 from pychron.core.geometry.geometry import calc_distances
 from pychron.core.regression.base_regressor import BaseRegressor
 from pychron.core.regression.ols_regressor import MultipleLinearRegressor
+from pychron.core.stats.idw import Invdisttree
 
 
 class SpecialFluxRegressor(BaseRegressor):
@@ -45,6 +47,65 @@ class SpecialFluxRegressor(BaseRegressor):
 
     def _calculate_coefficient_errors(self):
         return ''
+
+
+class InterpolationRegressor(SpecialFluxRegressor):
+    def predict(self, pts):
+        return zeros_like(pts)
+
+    def predict_error(self, pts, **kw):
+        return zeros_like(pts)
+
+
+class BSplineRegressor(InterpolationRegressor):
+    def calculate(self):
+        x, y = self.clean_xs.T
+
+        z = self.clean_ys
+        # self.rbf = Rbf(x, y, z)
+
+        self._tck = bisplrep(x, y, z, kx=4, ky=4)
+
+    def predict_grid(self, x, y):
+        znew = bisplev(x, y, self._tck)
+        return znew
+
+
+class RBFRegressor(InterpolationRegressor):
+    rbf_kind = 'multiquadric'
+
+    def calculate(self):
+        x, y = self.clean_xs.T
+
+        z = self.clean_ys
+        self.rbf = Rbf(x, y, z, function=self.rbf_kind)
+
+    def predict_grid(self, x, y):
+        return self.rbf(x, y)
+
+
+class GridDataRegressor(InterpolationRegressor):
+    method = 'cubic'
+
+    def calculate(self):
+        pass
+
+    def predict_grid(self, x, y):
+        return griddata(self.clean_xs, self.clean_ys, (x, y), method=self.method).T
+
+
+class IDWRegressor(InterpolationRegressor):
+    def calculate(self):
+        leafsize = 10
+        known = self.clean_xs
+        z = self.clean_ys
+        self._invdisttree = Invdisttree(known, z, leafsize=leafsize, stat=1)
+
+    def predict(self, pts):
+        nnear = 8  # 8 2d, 11 3d => 5 % chance one-sided -- Wendel, mathoverflow.com
+        eps = .1  # approximate nearest, dist <= (1 + eps) * true nearest
+        p = 2  # weights ~ 1 / distance**p
+        return self._invdisttree(pts, nnear=nnear, eps=eps, p=p)
 
 
 class NearestNeighborFluxRegressor(SpecialFluxRegressor):
