### 2021-02-21 #2919 @pvermees
Yes I can try to find some time. Note that I am not a practitioner of
40Ar/39Ar geochronology so I won't be able to install the software on a
mass spectrometer. So all that I will be able to test is the data
processing side of Pychron.

I assume that I will receive detailed instructions on how to conduct the
review?

P

On Sun, 21 Feb 2021 at 09:23, Richard Gowers <notifications@github.com>
wrote:

> @pvermees <https://github.com/pvermees> would you have time to do the
> review? Thanks
>
