### 2021-02-22 0d9c106
flux updates. addressed issue #1942
