+++ pychron/lasers/tasks/panes/ostech.py
@@ -51,9 +51,9 @@ class OsTechDiodeSupplementalPane(SupplementalPane):
         v = View(VGroup(UCustom('temperature_controller',
                         editor=InstanceEditor(view='control_view')),
                    label='Watlow'),
-            # VGroup(UItem('pyrometer', style='custom'),
-            #        label='Pyrometer',
-            #        ),
+             VGroup(UItem('pyrometer', style='custom'),
+                    label='Pyrometer',
+                    ),
             # VGroup(Item('control_module_manager', show_label=False, style='custom',
             #             ),
             #        #                      show_border = True,
