+++ pychron/hardware/watlow/base_ezzone.py
@@ -41,6 +41,18 @@ class Protocol:
     def write(self, *args, **kw):
         pass
 
+    def fromtemp(self, t):
+        return t
+
+    def totemp(self, t):
+        return t
+
+    def ctof(self, c):
+        return c*9/5.+32
+
+    def ftoc(self, f):
+        return (f-32)*5/9.
+
 
 class StandardProtocol(Protocol):
     mapping = {'baud_rate': [17002, 17002],
@@ -63,9 +75,30 @@ class StandardProtocol(Protocol):
                'ttun': 8022,
                'tbnd': 8034,
                'tgn': 8035,
-               'fn': 18002
+               'aut': 8026,
+               'fn': 18002,
+               'shi': 18010,
+               'slo': 18009,
+               'ishi': 4016,
+               'islo': 4015,
+               'ain1': 4001,
+               'ain2': 4001,
+               'hpr': 8011,
+               'pva1': 8031,
+               'pva2': 8031,
+               'ica': 4012,
+               'ier': 4002
                }
 
+    def fromtemp(self, t):
+        """
+        standard protocol temp also in f
+
+        """
+        return self.ftoc(t)
+
+    def totemp(self, t):
+        return self.ctof(t)
 
 class ModbusProtocol(Protocol):
     mapping = {'baud_rate': [2484, 2504],
@@ -85,10 +118,22 @@ class ModbusProtocol(Protocol):
                'hhy': 1900,
                'atsp': 1918,
                'tagr': 1916,
-               'ttun': 8022,
+               'ttun': 1910,
                'tbnd': 1912,
                'tgn': 1914,
-               'fn': 722
+               'aut': 1920,
+               'fn': 722,
+               'shi': 738,
+               'slo': 736,
+               'ishi': 390,
+               'islo': 388,
+               'ain1': 360,
+               'ain2': 440,
+               'hpr': 1904,
+               'pva1': 402,
+               'pva2': 402,
+               'ica': 382,
+               'ier': 362
                }
 
 
@@ -382,18 +427,24 @@ class BaseWatlowEZZone(HasTraits):
 
         self.report_pid()
 
+        e = self.read_analog_input_error()
+        c = self.read_analog_input_calibration()
+        self.debug('analog input error={}'.format(e))
+        self.debug('analog input cal={}'.format(c))
+
         self.info('read input/output scaling')
         if not self.simulation:
-            try:
-                osl, osh = self.read(736, nregisters=4, nbytes=13)
-                isl, ish = self.read(388, nregisters=4, nbytes=13)
+            if isinstance(self.protocol, ModbusProtocol):
+                try:
+                    osl, osh = self.read(736, nregisters=4, nbytes=13)
+                    isl, ish = self.read(388, nregisters=4, nbytes=13)
 
-                self._output_scale_low = osl
-                self._output_scale_high = osh
-                self._input_scale_low = isl
-                self._input_scale_high = ish
-            except TypeError:
-                pass
+                    self._output_scale_low = osl
+                    self._output_scale_high = osh
+                    self._input_scale_low = isl
+                    self._input_scale_high = ish
+                except TypeError:
+                    pass
 
         attrs = [
             ('read_autotune_setpoint', '_autotune_setpoint'),
@@ -401,14 +452,15 @@ class BaseWatlowEZZone(HasTraits):
             ('read_tru_tune_enabled', '_enable_tru_tune'),
             ('read_tru_tune_band', '_tru_tune_band'),
             ('read_tru_tune_gain', '_tru_tune_gain'),
-
-            #               ('read_output_scale_low', '_output_scale_low'),
-            #               ('read_output_scale_high', '_output_scale_high'),
[…]
