+++ app_utils/install.py
@@ -175,28 +175,50 @@ def ask_config():
     for k, v in config.items():
         print('{:<20s}: {}'.format(k, v))
 
-    config['pip_requirements'] = 'uncertainties peakutils qimage2ndarray'
-    config['pip_git_requirements'] = ['git+https://github.com/enthought/chaco.git#egg=chaco',
-                                      'git+https://github.com/enthought/enable.git#egg=enable']
-
-    creq = 'pip qt numpy statsmodels scikit-learn PyYAML yaml traitsui envisage sqlalchemy ' \
-           'Reportlab lxml xlrd xlwt xlsxwriter requests keyring pillow gitpython cython pytables ' \
-            'pyproj pymysql certifi jinja2 swig {}'.format(config['qt_bindings'])
-
-    if IS_MAC:
-        creq = '{} python.app'.format(creq)
-
-    if config['install_gis_plugin']:
-        creq = '{} '.format('qgis')
-
-
-    config['conda_requirements'] = creq
+    build_requirements(config)
 
     print()
     print()
     if yes('Continue? [y]/n >> '):
         return config
 
+
+def build_requirements(cfg):
+    pip_reqs = ['uncertainties',
+                'peakutils',
+                'qimage2ndarray']
+    pip_git_reqs = ['git+https://github.com/enthought/chaco.git#egg=chaco',
+                    'git+https://github.com/enthought/enable.git#egg=enable']
+
+    conda_reqs = ['numpy', 'statsmodels', 'scikit-learn', 'PyYAML', 'yaml', 'traits=5', 'traitsui=6', 'pyface=6',
+                  'envisage', 'sqlalchemy', 'Reportlab', 'lxml', 'xlrd', 'xlwt', 'xlsxwriter', 'requests', 'keyring',
+                  'pillow', 'gitpython', 'cython', 'pytables', 'pyproj', 'pymysql', 'certifi', 'jinja2', 'swig=3',
+                  cfg['qt_bindings']]
+
+    if IS_MAC:
+        conda_reqs.append('python.app')
+
+    if cfg['install_gis_plugin']:
+        conda_reqs.append('qgis')
+
+    cfg['pip_requirements'] = pip_reqs
+    cfg['pip_git_requirements'] = pip_git_reqs
+    cfg['conda_requirements'] = conda_reqs
+
+# config['pip_requirements'] = 'uncertainties peakutils qimage2ndarray'
+# config['pip_git_requirements'] =
+#
+# creq = 'pip qt numpy statsmodels scikit-learn PyYAML yaml traits=5 traitsui=6 pyface=6 envisage sqlalchemy ' \
+#        'Reportlab lxml xlrd xlwt xlsxwriter requests keyring pillow gitpython cython pytables ' \
+#         'pyproj pymysql certifi jinja2 swig=3 {}'.format(config['qt_bindings'])
+#
+# if IS_MAC:
+#     creq = '{} python.app'.format(creq)
+#
+# if config['install_gis_plugin']:
+#     creq = '{} '.format('qgis')
+
+# config['conda_requirements'] = creq
 
 def yes(msg):
     return input(msg) in ('', 'y', 'yes', 'Yes', 'YES')
@@ -244,14 +266,14 @@ def install_conda(cfg):
 
     # install deps
     subprocess.call(['conda', 'install', '--yes',
-                     '--name', env_name] + cfg['conda_requirements'].split(' '))
+                     '--name', env_name] + cfg['conda_requirements'])
 
     if IS_MAC:
-        subprocess.call(['conda', 'activate', cfg['conda_env_name']])
[…]

[…]
