+++ pychron/pyscripts/extraction_line_pyscript.py
@@ -200,14 +200,15 @@ class ExtractionPyScript(ValvePyScript):
 
     @verbose_skip
     @command_register
-    def set_cryo(self, value, block=False):
-        result = self._manager_action(('set_cryo', (value,), {'block': block}), protocol=EL_PROTOCOL)
+    def set_cryo(self, value, block=False, delay=1):
+        result = self._manager_action(('set_cryo', (value,), {'block': block, 'delay': delay}), protocol=EL_PROTOCOL)
+
         self.debug('set cyro result={}'.format(result))
         return result
 
     @verbose_skip
     @command_register
-    def get_cryo_temp(self, value):
+    def get_cryo_temp(self, value=1):
         result = self._manager_action(('get_cryo_temp', (value,), {}), protocol=EL_PROTOCOL)
         return result
 
@@ -356,7 +357,7 @@ class ExtractionPyScript(ValvePyScript):
             func = self._make_waitfor_func(*func_or_tuple, func_kw=func_kw)
         else:
             func = func_or_tuple
-            args = inspect.getargspec(func).args
+            args = inspect.getfullargspec(func).args
             if len(args) == 1:
                 include_time = True
             elif len(args) == 2:
