+++ pychron/dvc/dvc.py
@@ -290,10 +290,11 @@ class DVC(Loggable):
                 mrepo.open_repo(root)
             else:
                 url = self.make_url(self.meta_repo_name)
-                if url:
-                    self.debug('cloning meta repo url={}'.format(url))
-                    path = os.path.join(paths.dvc_dir, name)
-                    self.meta_repo.clone(url, path)
+                if url is not None:
+                    if url:
+                        self.debug('cloning meta repo url={}'.format(url))
+                        path = os.path.join(paths.dvc_dir, name)
+                        self.meta_repo.clone(url, path)
                 else:
                     self.debug('no url returned for MetaData repository. You need to clone your MetaData repository '
                                'manually')
