### 2021-03-12 c0f25cc
fixed pause from a extraction script
