+++ pychron/pipeline/nodes/__init__.py
@@ -50,12 +50,12 @@ from pychron.pipeline.nodes.persist import DVCPersistNode, PDFFigureNode, \
     XLSXAnalysisTablePersistNode, InterpretedAgePersistNode, DefineEquilibrationPersistNode, \
     FluxMonitorMeansPersistNode, CosmogenicCorrectionPersistNode
 from pychron.pipeline.nodes.push import PushNode
+from pychron.pipeline.nodes.recent_runs import RecentRunsNode
 from pychron.pipeline.nodes.report import ReportNode
 from pychron.pipeline.nodes.review import ReviewNode
 from pychron.pipeline.nodes.runid_edit import RunIDEditNode
 from pychron.pipeline.nodes.scripting import PyScriptNode
 from pychron.pipeline.nodes.table import InterpretedAgeTableNode, GroupAnalysisTableNode, SubGroupAnalysisTableNode
-
 
 class NodeFactory:
     def __init__(self, name, factory):

[…]
