### 2021-03-16 26cbb89
addressed issue #1946.  updates for enable 5
