+++ pychron/hardware/core/communicators/modbustcp_communicator.py
@@ -1,0 +1,29 @@
+# ===============================================================================
+# Copyright 2021 ross
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ===============================================================================
+from pymodbus.client.sync import ModbusTcpClient
+
+from pychron.hardware.core.communicators.communicator import Communicator
+
+
+class ModbustcpCommunicator(Communicator):
+    def open(self, *args, **kw):
+        a = self.address
+        self.handle = ModbusTcpClient(a)
+        return True
+
+    def initialize(self, *args, **kw):
+        return self.handle.connect()
+# ============= EOF =============================================
