+++ pychron/hardware/actuators/t4_actuator.py
@@ -19,8 +19,9 @@ from pychron.hardware.labjack.base_t4 import BaseT4
 
 class T4Actuator(BaseT4, GPActuator):
     def _actuate(self, obj, action):
-        self.set_channel_state(obj.address, action.lower()=='open')
+        self.set_channel_state(obj.address, action.lower() == 'open')
         return True
+
     def get_channel_state(self, obj, **kw):
         if isinstance(obj, str):
             addr = obj
@@ -28,6 +29,6 @@ class T4Actuator(BaseT4, GPActuator):
             addr = obj.state_address
             if not addr:
                 addr = obj.address
-                
+
         return BaseT4.get_channel_state(self, addr, **kw)
 # ============= EOF =============================================
