+++ pychron/pipeline/results/iso_evo.py
@@ -13,17 +13,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from traits.api import Str, Float
+from traits.api import Str, Float, Property, cached_property
 
 from pychron.core.helpers.formatting import floatfmt
 from pychron.pipeline.results.base import BaseResult
 
-GOODNESS_TAGS = ('int_err', 'slope', 'outlier', 'curvature', 'rsquared', 'signal_to_blank', 'signal_to_baseline')
-GOODNESS_NAMES = ('Intercept Error', 'Slope', 'Outliers', 'Curvature', 'RSquared', 'Blank/Signal %', 'Baseline/Signal %')
+GOODNESS_TAGS = ('int_err', 'slope', 'outlier', 'curvature', 'rsquared',
+                 'signal_to_blank', 'signal_to_baseline',
+                 'smart_filter')
+GOODNESS_NAMES = ('Intercept Error', 'Slope', 'Outliers', 'Curvature', 'RSquared', 'Blank/Signal %',
+                  'Baseline/Signal %', 'Smart Filter')
 INVERTED_GOODNESS = ('rsquared',)
+
+ISO_EVO_RESULT_ARGS = ('intercept_value', 'intercept_error', 'slope', 'outlier', 'curvature', 'rsquared')
 
 
 class IsoEvoResult(BaseResult):
+    isotope_obj = None
     n = Str
     fit = Str
     intercept_value = Float
@@ -37,6 +43,7 @@ class IsoEvoResult(BaseResult):
     rsquared_goodness = None
     signal_to_blank_goodness = None
     signal_to_baseline_goodness = None
+    smart_filter_goodness = None
 
     int_err_threshold = None
     slope_threshold = None
@@ -46,6 +53,7 @@ class IsoEvoResult(BaseResult):
     signal_to_blank_threshold = None
     signal_to_baseline_threshold = None
     signal_to_baseline_percent_threshold = None
+    smart_filter_threshold = None
 
     int_err = None
     slope = None
@@ -54,6 +62,15 @@ class IsoEvoResult(BaseResult):
     rsquared = None
     signal_to_blank = None
     signal_to_baseline = None
+    smart_filter = None
+
+    hover_text = Property
+
+    @cached_property
+    def _get_hover_text(self):
+        a = 'RunID: {}'.format(self.record_id)
+        lines = [a]
+        return lines
 
     @property
     def goodness(self):
