+++ pychron/user/tasks/plugin.py
@@ -15,7 +15,6 @@
 # ===============================================================================
 
 # ============= enthought library imports =======================
-from __future__ import absolute_import
 from envisage.ui.tasks.task_factory import TaskFactory
 
 # ============= standard library imports ========================
