+++ pychron/pipeline/engine.py
@@ -956,7 +956,8 @@ class PipelineEngine(Loggable):
 
         def func(editor):
             vs = self.selected.unknowns
-            editor.set_items(vs)
+            if editor:
+                editor.set_items(vs)
             self.state.unknowns = vs
             for node in self.pipeline.nodes:
                 if isinstance(node, UnknownNode) and node is not self.selected:
@@ -1163,9 +1164,10 @@ class PipelineEngine(Loggable):
                 setattr(self, lc, 0)
                 setattr(self, lr, n)
 
-            if getattr(self, lc) >= getattr(self, lr) or n == 1:
-                setattr(self, lc, 0)
-                func(editor)
+        if getattr(self, lc) >= getattr(self, lr) or n == 1:
+            setattr(self, lc, 0)
+            func(editor)
+            if editor:
                 editor.refresh_needed = True
 
     def _dclicked_changed(self, new):
