+++ pychron/envisage/browser/sample_view.py
@@ -77,11 +77,11 @@ irrad_grp = BorderVGroup(HGroup(UItem('irradiation_enabled',
                          label='Irradiations')
 
 project_grp = BorderVGroup(UItem('projects',
-                                 height=100,
                                  editor=FilterTabularEditor(editable=False,
                                                             enabled_cb='project_enabled',
                                                             use_fuzzy=True,
                                                             column_index=-1,
+                                                            stretch_last_section=False,
                                                             refresh='refresh_needed',
                                                             selected='selected_projects',
                                                             adapter=ProjectAdapter(),
@@ -111,9 +111,10 @@ simple_mass_spectrometer_grp = BorderHGroup(UItem('mass_spectrometers_enabled',
                                             label='Mass Spectrometer')
 
 pi_grp = Group(UItem('principal_investigators',
-                     height=100,
+                     height=-75,
                      editor=FilterTabularEditor(editable=False,
                                                 use_fuzzy=True,
+                                                stretch_last_section=False,
                                                 enabled_cb='principal_investigator_enabled',
                                                 refresh='refresh_needed',
                                                 selected='selected_principal_investigators',
