+++ pychron/pipeline/nodes/fit.py
@@ -331,12 +331,13 @@ class FitIsotopeEvolutionNode(FitNode):
                 if goodness_threshold:
                     int_err_goodness = bool(pe < goodness_threshold)
 
-                bs = iso.baseline.error
-                signal_to_baseline = abs(bs / i * 100)
                 signal_to_baseline_threshold = f.signal_to_baseline_goodness
                 signal_to_baseline_percent_threshold = f.signal_to_baseline_percent_goodness
                 signal_to_baseline_goodness = None
+                signal_to_baseline = 0
                 if hasattr(iso, 'baseline'):
+                    bs = iso.baseline.error
+                    signal_to_baseline = abs(bs / i * 100)
                     if signal_to_baseline_threshold and signal_to_baseline_percent_threshold:
                         if signal_to_baseline > signal_to_baseline_threshold:
                             signal_to_baseline_goodness = bool(pe < signal_to_baseline_percent_threshold)
