### 2021-03-26 a1206bb
addressed issue #1954 added print component
