+++ pychron/hardware/actuators/__init__.py
@@ -90,4 +90,4 @@ PACKAGES = dict(AgilentGPActuator='{}.agilent.agilent_gp_actuator'.format(base),
                 DummyGPActuator='{}.dummy_gp_actuator'.format(abase),
                 RPiGPIO='{}.rpi_gpio'.format(base),
                 T4Actuator='{}.t4_actuator'.format(abase),
-                PLC2000Actuator='{}.plc2000_gp_actuator'.format(abase))
+                PLC2000GPActuator='{}.plc2000_gp_actuator'.format(abase))
