+++ pychron/experiment/experiment_executor.py
@@ -171,6 +171,7 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
 
     ratio_change_detection_enabled = Bool(False)
     execute_open_queues = Bool(True)
+    use_preceding_blank = Bool(True)
 
     # dvc
     use_dvc_persistence = Bool(False)
@@ -190,9 +191,8 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
 
     _prev_blanks = Dict
     _prev_baselines = Dict
-    _prev_blank_runid = String
     _err_message = String
-    _prev_blank_id = Long
+
     _ratios = Dict
     _failure_counts = Dict
     _excluded_uuids = Dict
@@ -253,6 +253,7 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
                  'experiment_type',
                  'laboratory',
                  'ratio_change_detection_enabled',
+                 'use_preceding_blank',
                  'execute_open_queues')
         self._preference_binder(prefid, attrs)
 
@@ -750,18 +751,29 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
         self.debug('{} finished'.format(run.runid))
         if self.is_alive():
             self.debug('spec analysis type {}'.format(spec.analysis_type))
-            if spec.analysis_type.startswith('blank'):
-                pb = run.get_baseline_corrected_signals()
-                if pb is not None:
-                    self._prev_blank_runid = run.spec.runid
-                    # self._prev_blank_id = run.spec.analysis_dbid
-                    self._prev_blanks = pb
-                    self.debug('previous blanks ={}'.format(pb))
+            self._set_prev(run)
 
         run.teardown()
 
         self.measuring_run = None
         self.debug('join run finished')
+
+    def _set_prev(self, run):
+        if hasattr(run, 'get_baseline_corrected_signal_dict'):
+            d = (run.record_id, run.get_baseline_corrected_signal_dict())
+            self._prev_blanks[run.analysis_type] = d
+            self._prev_blanks['blank'] = d
+            self._prev_baselines = run.get_baseline_dict()
+        else:
+            if run.analysis_type.startswith('blank'):
+                pb = run.get_baseline_corrected_signals()
+                if pb is not None:
+                    d = (run.spec.runid, pb)
+                    self._prev_blanks[run.spec.analysis_type] = d
+                    self._prev_blanks['blank'] = d
+                    self.debug('previous blanks ={}'.format(d))
+
+            self._prev_baselines = run.get_baselines()
 
     def _do_run(self, run):
         self._set_thread_name(run.runid)
@@ -888,10 +900,12 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
         t.join()
 
         self.debug('{} finished'.format(run.runid))
-        if run.analysis_type.startswith('blank'):
-            pb = run.get_baseline_corrected_signals()
-            if pb is not None:
-                self._prev_blanks = pb
+        self._set_prev(run)
+        # if run.analysis_type.startswith('blank'):
+        #     pb = run.get_baseline_corrected_signals()
+        #     if pb is not None:
+        #         self._prev_blanks[run.spec.analysis_type] = (run.spec.runid, pb)
+        #         self._prev_blanks['blank'] = (run.spec.runid, pb)
 
         do_after(1000, run.teardown)
 
@@ -1178,7 +1192,12 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
                   'use_db_persistence', 'use_dvc_persistence', 'use_xls_persistence'):
             setattr(arun, k, getattr(self, k))
 
[…]
