+++ pychron/experiment/automated_run/persistence.py
@@ -247,6 +247,7 @@ class AutomatedRunPersister(BasePersister):
             set_preference(preferences, self, attr, 'pychron.experiment.{}'.format(attr), cast)
 
         set_preference(preferences, self, 'use_massspec_database', 'pychron.massspec.database.enabled', to_bool)
+
     # ===============================================================================
     # data writing
     # ===============================================================================
@@ -456,100 +457,99 @@ class AutomatedRunPersister(BasePersister):
         #     self.datahub.update_spec(run_spec)
 
         cp = self._current_data_frame
-
         ln = run_spec.labnumber
-        aliquot = run_spec.aliquot
+        # aliquot = run_spec.aliquot
 
         if save_local:
             # save to local sqlite database for backup and reference
             self._local_db_save()
 
-        # save to a database
-        db = self.datahub.get_db('isotopedb')
-        if not db or not db.connected:
-            self.warning('No database instance. Not saving post measurement to isotopedb database')
-        else:
-            with db.session_ctx() as sess:
-                pt = time.time()
-
-                lab = db.get_labnumber(ln)
-
-                endtime = get_datetime().time()
-                self.info('analysis finished at {}'.format(endtime))
-
-                un = run_spec.username
-                dbuser = db.get_user(un)
-                if dbuser is None:
-                    self.debug('user= {} does not existing. adding to database now'.format(un))
-                    dbuser = db.add_user(un)
-
-                self.debug('adding analysis identifier={}, aliquot={}, '
-                           'step={}, increment={}'.format(ln, aliquot,
-                                                          run_spec.step,
-                                                          run_spec.increment))
-                a = db.add_analysis(lab,
-                                    user=dbuser,
-                                    uuid=run_spec.uuid,
-                                    endtime=endtime,
-                                    aliquot=aliquot,
-                                    step=run_spec.step,
-                                    increment=run_spec.increment,
-                                    comment=run_spec.comment,
-                                    whiff_result=self.per_spec.whiff_result)
-                sess.flush()
-                run_spec.analysis_dbid = a.id
-                run_spec.analysis_timestamp = a.analysis_timestamp
-
-                experiment = db.get_experiment(self.dbexperiment_identifier, key='id')
-                if experiment is not None:
-                    # added analysis to experiment
-                    a.experiment_id = experiment.id
-                else:
-                    self.warning('no experiment found for {}'.format(self.dbexperiment_identifier))
-
-                # save measurement
-                meas = self._save_measurement(db, a)
-                # save extraction
-                ext = self._db_extraction_id
-                if ext is not None:
-                    dbext = db.get_extraction(ext, key='id')
-                    a.extraction_id = dbext.id
-                    # save sensitivity info to extraction
-                    self._save_sensitivity(dbext, meas)
-
-                else:
-                    self.debug('no extraction to associate with this run')
-
-                self._save_spectrometer_info(db, meas)
-
-                # add selected history
-                db.add_selected_histories(a)
-                # self._save_isotope_info(a, ss)
-                self._save_isotope_data(db, a)
-
-                # save ic factor
-                self._save_detector_intercalibration(db, a)
-
-                # save blanks
-                self._save_blank_info(db, a)
-
-                # save peak center
-                self._save_peak_center(db, a, cp)
-
-                # save monitor
[…]
