+++ pychron/regex.py
@@ -42,7 +42,7 @@ MDD_PARAM_REGEX = re.compile(r'E= {3}(?P<E>\d+.\d+) {7}\+\- {4}(?P<Eerr>\d+.\d+)
                              r'{11}Ordinate= {3}(?P<O>\d+.\d+) {7}\+\- {2}(?P<Oerr>\d+.\d+)')
 
 
-STEPREGEX = re.compile(r'^[A-Z]+$')
+STEPREGEX = re.compile(r'^[a-zA-Z]+$|^[0-9]+$')
 
 DETREGEX = re.compile(r'^(H1|H2|L1|L2|CDD|AX)(\(CDD\))?$')
 
