+++ pychron/image/basler_pylon_camera.py
@@ -13,6 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
+from pypylon.genicam import GenericException
+
 from pychron.core.yaml import yload
 
 try:
@@ -88,11 +90,14 @@ class BaslerPylonCamera(Loggable):
     def read(self):
         ok, img = False, None
         if self._cam and not self._setting_config:
-            res = self._cam.RetrieveResult(0, pylon.TimeoutHandling_Return)
-            if res.IsValid():
-                ok = True
-                img = res.Array
-            res.Release()
+            try:
+                res = self._cam.RetrieveResult(0, pylon.TimeoutHandling_Return)
+                if res.IsValid():
+                    ok = True
+                    img = res.Array
+                res.Release()
+            except GenericException:
+                pass
 
         return ok, img
 
