+++ pychron/tx/factories.py
@@ -52,6 +52,11 @@ class FusionsDiodeFactory(LaserFactory):
 class FusionsUVFactory(LaserFactory):
     _name = 'FusionsUV'
 
+
+class OsTechDiodeFactory(LaserFactory):
+    _name = 'OsTechDiode'
+
+
 from pychron.paths import paths
 path = os.path.join(paths.log_dir, 'pps.log.json')
 
