### 2021-04-06 6015603
updated el canvas loaders
