+++ pychron/canvas/canvas2D/scene/extraction_line_scene.py
@@ -18,67 +18,18 @@
 # ============= standard library imports ========================
 import os
 
-from numpy.core.numeric import Inf
 from traits.api import Dict
 
 # ============= local library imports  ==========================
-from pychron.canvas.canvas2D.scene.canvas_parser import get_volume, CanvasParser
-from pychron.canvas.canvas2D.scene.primitives.connections import Tee, Fork, Elbow, Connection
+from pychron.canvas.canvas2D.scene.base_scene_loader import floatify, colorify
 from pychron.canvas.canvas2D.scene.primitives.lasers import Laser, CircleLaser
-from pychron.canvas.canvas2D.scene.primitives.primitives import Label, BorderLine, Line, Image, ValueLabel
+from pychron.canvas.canvas2D.scene.primitives.primitives import Label, BorderLine, Image, ValueLabel
 from pychron.canvas.canvas2D.scene.primitives.pumps import Turbo, IonPump
-from pychron.canvas.canvas2D.scene.primitives.rounded import RoundedRectangle, CircleStage, Spectrometer, Getter
-from pychron.canvas.canvas2D.scene.primitives.valves import RoughValve, Valve, Switch, ManualSwitch
+from pychron.canvas.canvas2D.scene.primitives.rounded import CircleStage, Spectrometer, Getter
+from pychron.canvas.canvas2D.scene.primitives.valves import RoughValve, Valve
 from pychron.canvas.canvas2D.scene.scene import Scene
-from pychron.core.helpers.strtools import to_bool
-from pychron.core.yaml import yload
-from pychron.extraction_line.switch_parser import SwitchParser
-from pychron.paths import paths
-
-KLASS_MAP = {'turbo': Turbo,
-             'laser': Laser,
-             'ionpump': IonPump,
-             'getter': Getter,
-             'spectrometer': Spectrometer,
-             'circle_stage': CircleStage,
-             'circle_laser': CircleLaser}
-
-RECT_TAGS = ('stage', 'laser', 'spectrometer',
-             'turbo', 'getter', 'tank',
-             'ionpump', 'gauge', 'rectangle',
-             'circle_stage', 'circle_laser')
-
-SWITCH_TAGS = ('switch', 'valve', 'rough_valve', 'manual_valve')
-
-
-def get_offset(elem, default=None):
-    offset = elem.get('offset')
-    if default is None:
-        default = 0, 0
-
-    txt = ''
-    if offset:
-        txt = offset
-        x, y = floatify(offset)
-    else:
-        x, y = default
-    return x, y, txt
-
-
-def floatify(a, delim=','):
-    if not isinstance(a, str):
-        a = a.text.strip()
-
-    return [float(i) for i in a.split(delim)]
-
-
-def colorify(a):
-    if not isinstance(a, str):
-        a = a.text.strip()
-
-    if a.startswith('0x'):
-        a = int(a, 16)
-    return a
+from pychron.canvas.canvas2D.scene.xml_scene_loader import XMLLoader
+from pychron.canvas.canvas2D.scene.yaml_scene_loader import YAMLLoader
 
 
 class ExtractionLineScene(Scene):
@@ -89,7 +40,7 @@ class ExtractionLineScene(Scene):
         self.overlays = []
         self.reset_layers()
 
-        origin, color_dict, valve_dimension = self._load_config(configpath, canvas)
+        origin, color_dict, valve_dimension, images = self._load_config(configpath, canvas)
         if pathname.endswith('.yaml') or pathname.endswith('.yml'):
             klass = YAMLLoader
         else:
@@ -99,15 +50,15 @@ class ExtractionLineScene(Scene):
 
         loader.load_switchables(self, valvepath)
         loader.load_rects(self)
-        #
-        # loader.load_pipettes()
-        # loader.load_markup()
[…]
