### 2021-04-07 65fefac
fixed set_laser_output
