+++ pychron/processing/analyses/view/regression_view.py
@@ -15,19 +15,46 @@
 # ===============================================================================
 from chaco.plot_containers import HPlotContainer
 from enable.component_editor import ComponentEditor
-from traits.api import HasTraits, Instance
+from traits.api import HasTraits, Instance, Any
 from traitsui.api import View, UItem
 
+from pychron.core.helpers.formatting import format_percent_error, errorfmt
 from pychron.graph.stacked_graph import StackedGraph
 from pychron.graph.stacked_regression_graph import StackedRegressionGraph
+from pychron.graph.tools.regression_inspector import RegressionInspectorTool
+from pychron.pychron_constants import PLUSMINUS
+
+
+class AnalysisRegressionInspectorTool(RegressionInspectorTool):
+    analysis = Any
+
+    def assemble_lines(self):
+        lines = super(AnalysisRegressionInspectorTool, self).assemble_lines()
+        an = self.analysis
+        a = an.age
+
+        ef = errorfmt(a, an.age_err)
+        ef_wo_j = errorfmt(a, an.age_err_wo_j)
+        lines.insert(0, 'Age={:0.4f} {}{} w/o_J={}'.format(a, PLUSMINUS, ef, ef_wo_j))
+        return lines
+
+
+class AnalysisRegressionGraph(StackedRegressionGraph):
+    analysis = Any
+
+    def regression_inspector_factory(self, line):
+        tool = AnalysisRegressionInspectorTool(component=line, analysis=self.analysis)
+        return tool
 
 
 class RegressionView(HasTraits):
     name = 'Regressions'
     container = Instance(HPlotContainer)
+    analysis = Any
 
     def initialize(self, an):
         an.load_raw_data()
+        self.analysis = an
         self.setup_graph(an)
 
     def setup_graph(self, an):
@@ -36,8 +63,8 @@ class RegressionView(HasTraits):
 
         container_dict = {'spacing': 5, 'stack_order': 'top_to_bottom'}
         sg = StackedGraph(container_dict=container_dict)
-        bg = StackedRegressionGraph(container_dict=container_dict)
-        ig = StackedRegressionGraph(container_dict=container_dict)
+        bg = AnalysisRegressionGraph(container_dict=container_dict, analysis=an)
+        ig = AnalysisRegressionGraph(container_dict=container_dict, analysis=an)
 
         isos = an.sorted_values(reverse=False)
 
@@ -71,9 +98,9 @@ class RegressionView(HasTraits):
             ig.set_x_limits(min_=0, max_=max(iso.offset_xs) * 1.05, plotid=i)
 
         ig.refresh()
+        ig.on_trait_change(self.handle_regression, 'regression_results')
 
         for i, baseline in enumerate(baselines):
-
             p = bg.new_plot(ytitle=baseline.detector, xtitle='Time (s)', title='Baseline')
             bg.add_axis_tool(p, p.x_axis)
             bg.add_axis_tool(p, p.y_axis)
@@ -92,6 +119,18 @@ class RegressionView(HasTraits):
         container.add(bg.plotcontainer)
 
         self.container = container
+
+    def handle_regression(self, new):
+        if new:
+            for plot, regressor in new:
+                for k, iso in self.analysis.isotopes.items():
+                    yt = plot.y_axis.title
+                    if k == yt or '{}({})'.format(iso.name, iso.detector) == yt:
+                        iso.set_fit(regressor.get_fit_dict())
+                        break
+
+            self.analysis.calculate_age(force=True)
+            self.analysis.analysis_view.refresh()
 
     def traits_view(self):
         v = View(UItem('container', style='custom', editor=ComponentEditor()),
