### 2021-04-10 65e5a93
addressed issue #1971  use IntSparseTicks for analysis_number
