### 2021-04-11 c7d099a
added mean label location to ideogram option
