+++ pychron/options/views/ideogram_views.py
@@ -198,8 +198,9 @@ class IdeogramSubOptions(SubOptions):
 
 class IdeogramAppearance(AppearanceSubOptions):
     def traits_view(self):
-        mi = BorderHGroup(UItem('mean_indicator_fontname'),
-                          UItem('mean_indicator_fontsize'),
+        mi = BorderVGroup(HGroup(UItem('mean_indicator_fontname'),
+                          UItem('mean_indicator_fontsize')),
+                          Item('display_mean_location'),
                           label='Mean Indicator')
 
         ee = BorderHGroup(UItem('error_info_fontname'),
