+++ pychron/hardware/isotopx_spectrometer_controller.py
@@ -20,6 +20,8 @@ from traits.api import Str, HasTraits
 from apptools.preferences.preference_binding import bind_preference
 # ============= standard library imports ========================
 from threading import RLock, Lock
+from queue import Queue
+import time
 # ============= local library imports  ==========================
 
 from pychron.hardware.core.core_device import CoreDevice
@@ -29,15 +31,36 @@ class NGXController(CoreDevice):
     username = Str('')
     password = Str('')
     lock = None
-
+    canceled = False
+    
+    def ask(self, cmd, *args, **kw):
+        resp = super(NGXController, self).ask(cmd, *args, **kw)
+        if any((cmd.startswith(t) for t in ('GetValveStatus', 'OpenValve', 'CloseValve'))):
+            if resp and resp.strip() not in ('E00', 'OPEN', 'CLOSED'):
+                #self.event_buf.push(resp)
+                self.debug('retrying')
+                time.sleep(0.5)
+                return self.ask(cmd, *args, **kw)
+        return resp
+    
+    #def read(self, *args, **kw):
+    #    if self.event_buffer.empty():
+    #        resp = super(NGXController, self).read(*args, **kw)
+    #    else:
+    #        resp = self.event_buffer.get()
+    #    return resp
+    
     def set(self, *args, **kw):
         return HasTraits.set(self, *args, **kw)
 
+            
     def initialize(self, *args, **kw):
         ret = super(NGXController, self).initialize(*args, **kw)
-
+        
         # trying a new locking mechanism see ngx.trigger for more details
         self.lock = Lock()
+     #   self.event_buffer = Queue()
+        
         if ret:
             resp = self.read()
 
@@ -47,7 +70,7 @@ class NGXController(CoreDevice):
             if resp:
                 self.info('NGX-{}'.format(resp))
                 self.ask('Login {},{}'.format(self.username, self.password))
-
+            
             return True
 
 
