+++ pychron/spectrometer/jobs/peak_center.py
@@ -171,7 +171,7 @@ class BasePeakCenter(HasTraits):
                            max_=max([start, end]))
 
         def get_reference_intensity():
-            keys, signals, t = spec.get_intensities(trigger=True)
+            keys, signals, t, inc = spec.get_intensities(trigger=True)
             idx = keys.index(self.reference_detector.name)
             return signals[idx]
 
