### 2021-04-13 0eac1b5
added fixed size figures. added ability to import options files
