+++ pychron/options/layout.py
@@ -73,6 +73,7 @@ class FigureLayout(HasTraits):
     rows = Int(1)
     columns = Int(2)
     fixed = Enum('column', 'row', 'filled_grid')
+    fixed_width = Int(0)
 
     row_enabled = Property(depends_on='fixed')
     column_enabled = Property(depends_on='fixed')
@@ -123,12 +124,14 @@ class FigureLayout(HasTraits):
         self.items.append(LayoutItem(kind=kind))
 
     def traits_view(self):
-        rc_grp = VGroup(HGroup(Item('rows',
+        rc_grp = VGroup(Item('fixed_width'),
+                        HGroup(Item('rows',
                                     enabled_when='row_enabled'),
                                Item('columns',
                                     enabled_when='column_enabled'
                                     ),
-                               Item('fixed')),
+                               Item('fixed'),
+                               enabled_when='not fixed_width'),
                         label='Layout', show_border=True)
         v = View(rc_grp, resizable=True)
         return v
