### 2021-04-14 9956802
added integrated_intensity to get_intensity/intensities
