+++ pychron/pipeline/plot/plotter/ideogram.py
@@ -669,6 +669,9 @@ class Ideogram(BaseArArFigure):
                                  error=we,
                                  nsigma=opt.nsigma,
                                  color=plotkw['color'],
+                                 group_marker=plotkw.get('marker', 'circle'),
+                                 group_marker_size=plotkw.get('marker_size', 1),
+                                 display_group_marker=opt.display_group_marker,
                                  group_id=gid,
                                  location=opt.display_mean_location,
                                  visible=opt.display_mean_indicator,
