+++ app_utils/install.py
@@ -57,7 +57,7 @@ INITIALIZATION_TXT = '''<root>
             <plugin enabled="false">Query</plugin>
             <plugin enabled="false">GitLab</plugin>
         </general>
-        <hardware>
+        <hardware>{}
         </hardware>
         <social>
         </social>
@@ -69,6 +69,13 @@ INITIALIZATION_TXT = '''<root>
     </plugins>
 </root>
 '''
+
+HARDWARE_PLUGIN_TXT = """
+            <plugin enabled="true">ExtractionLine
+                <manager enabled="true">switch_manager
+                <device enabled="true">switch_controller</device>
+                </manager>
+            </plugin>"""
 
 DEFAULTS_TXT = '''#script defaults file
 #defines default pyscripts for the analysis types
@@ -305,128 +312,46 @@ CANVAS_XML = """<?xml version="1.0" ?>
 CANVAS_TXT = """connection: []
 getter: []
 hconnection:
-- end:
-    name: '2'
-    offset: ''
-  name: '1_2'
-  start:
-    name: '1'
-    offset: ''
+  - end:
+      name: 'Spec'
+      offset: ''
+    start:
+      name: 'B'
+      offset: ''
 ionpump:
-- border_width: 5
-  color: 50,150,100,255
-  dimension: 6.0, 5.0
-  display_name: IonPump
-  fill: true
-  name: IonPump
-  translation: -30.0,20.0
+  - name: IonPump
+    color: 234,165,57
+    dimension: 8,3
+    translation: -24.0,-8
 laser: []
 manualswitch: []
 spectrometer:
-- border_width: 5
-  color: 250,120,150,255
-  dimension: 10.0, 10.0
-  display_name: QMS
-  fill: true
-  name: QMS
-  translation: -28.0,-15.0
+  - name: Spec
+    border_width: 5
+    color: 250,120,150,255
+    dimension: 10.0, 10.0
+    translation: -25.0,0.0
 stage: []
 switch: []
 turbo:
-- border_width: 5
-  color: 250,250,0,255
+- color: 250,250,0,255
   dimension: 4.0, 4.0
-  display_name: Turbo
-  fill: true
-  name: Turbo
   translation: -11.0,-28.0
+  name: Turbo
 valve:
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '0'
-  translation: -27.0,14.5
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '1'
-  translation: -22.0,0.0
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '2'
-  translation: -14.0,0.0
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '3'
-  translation: -10.0,-15.0
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '4'
-  translation: -6.0,5.0
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '5'
-  translation: -6.0,-5.0
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
-  display_name: null
-  fill: true
-  name: '6'
-  translation: 2.0,11.5
-- border_width: 3
-  color: 255,0,0,255
-  dimension: 3.0, 3.0
[…]
