### 2021-05-05 49a222c
1) GPIBInstrument from pyvisa has no ask (should be query) in GPIB Communicator. 2) GPIB Communicator does not accept ‘verbose’ keyword, so removed from base_controller. 3) Calling the update method with getattr in scanable_device returns the method, so () is needed to actually call the function. I’m not sure if this is a python 3 change.
