### 2021-05-05 6009d6c
GPIB Communicator will now accept ‘verbose’ keyword from base_controller
