### 2021-05-06 c131eed
fixed kca aux plot.
