+++ pychron/hardware/gauges/plc2000/plc2000_gauge_controller.py
@@ -33,11 +33,13 @@ class PLC2000GaugeController(BaseGaugeController, CoreDevice, ModbusMixin):
 
         if isinstance(name, str):
             gauge = self.get_gauge(name)
-            register = gauge.channel
+            channel = gauge.channel
         else:
-            register = name.channel
+            channel = name.channel
 
         if name is not None:
+            # register = channel-1
+            register = int(channel) - 1
             try:
                 pressure = self._read_float(register)
             except BaseException as e:
