+++ pychron/hardware/core/modbus.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ===============================================================================
 from pymodbus.constants import Endian
-from pymodbus.payload import BinaryPayloadDecoder
+from pymodbus.payload import BinaryPayloadDecoder, BinaryPayloadBuilder
 
 
 class ModbusMixin:
@@ -23,8 +23,34 @@ class ModbusMixin:
     """
     def _read_float(self, register, *args, **kw):
         result = self._read_holding_registers(int(register), 2, *args, **kw)
+        return self._decode_float(result)
+
+    def _read_input_float(self, register, *args, **kw):
+        result = self._read_input_registers(int(register), 2, *args, **kw)
+        return self._decode_float(result)
+
+    def _decode_float(self, result):
         decoder = BinaryPayloadDecoder.fromRegisters(result.registers, Endian.Big, wordorder=Endian.Little)
         return decoder.decode_32bit_float()
+
+    def _get_payload(self, value, is_float=True):
+        builder = BinaryPayloadBuilder(byteorder=Endian.Big, wordorder=Endian.Little)
+        if is_float:
+            builder.add_32bit_float(value)
+        else:
+            builder.add_32bit_int(int(value))
+        payload = builder.build()
+        return payload
+
+    def _write_int(self, register, value, *args, **kw):
+        payload = self._get_payload(value, is_float=False)
+        self.debug('writing int register={} payload={} value={}'.format(register, payload, value))
+        self._write_registers(register, payload, skip_encode=True)
+
+    def _write_float(self, register, value, *args, **kw):
+        payload = self._get_payload(value)
+        self.debug('writing float register={} payload={} value={}'.format(register, payload, value))
+        self._write_registers(register, payload, skip_encode=True)
 
     def _func(self, funcname, *args, **kw):
         if self.communicator:
