+++ pychron/hardware/axis.py
@@ -70,6 +70,9 @@ class Axis(ConfigLoadable):
     def ask(self, cmd):
         return self.parent.ask(cmd)
 
+    @property
+    def calibrated_velocity(self):
+        return min(self.max_velocity, max(self.min_velocity, self.velocity))
     # private
     def _get_velocity(self):
         return self._velocity
