+++ pychron/hardware/core/motion/motion_profiler.py
@@ -42,7 +42,7 @@ class MotionProfiler(ConfigLoadable):
     max_acceleration = Float(50)
 
     def load(self, p):
-        attrs = ['max_velocity', 'max_transit_time',
+        attrs = ['max_velocity', 'min_velocity', 'max_transit_time',
                  'min_acceleration_time', 'velocity_tol',
                  'acceleration_tol', 'deceleration_tol',
                  'max_acceleration']
@@ -94,6 +94,7 @@ class MotionProfiler(ConfigLoadable):
 
         vel = (displacement * 2 * acc / 3.) ** 0.5
         vel = min(self.max_velocity, vel)
+        vel = max(self.min_velocity, vel)
 
         if cnt > 200 or \
                         acc >= self.max_acceleration or \
