+++ pychron/hardware/core/base_core_device.py
@@ -261,9 +261,10 @@ class BaseCoreDevice(HasCommunicator, ConsumerMixin):
         for i in range(ntries + 1):
             resp = self._parse_response(self.ask(cmd, verbose=verbose))
             if verbose:
-                m = 'repeat command {} response = {} len={} '.format(i + 1,
-                                                                     resp,
-                                                                     len(str(resp)) if resp is not None else None)
+                resp = resp or ''
+                resp = resp.strip()
+                n = len(str(resp))
+                m = 'repeat command {} response = {} len={} '.format(i + 1, resp, n)
                 self.debug(m)
 
             if break_val and resp == break_val:
