### 2021-05-10 d3dbba8
fixed n/total_n. improved tick_both_sides
