### 2021-05-11 23b9de4
Added Lakeshore 330 (single output, no output number accepted in commands). Also needed to add remove_eol function to gpib_communicator to mimic serial behavior.
