### 2021-05-11 069d642
XMLLoader needs load_widgets for ExtractionLineScene until XML support is dropped
