### 2021-05-12 1d41756
added approximate_polygon_center3
