+++ pychron/hardware/fusions/fusions_logic_board.py
@@ -197,8 +197,7 @@ class FusionsLogicBoard(CoreDevice):
         """
         interlocks = self.check_interlocks()
         if not interlocks:
-
-            resp = self.repeat_command('ENBL 1', check_val='OK')
+            resp = self.repeat_command('ENBL 1', check_val='OK', ntries=200, delay=3)
             if resp == 'OK' or self.simulation:
                 return True
 
