+++ pychron/core/geometry/geometry.py
@@ -18,7 +18,7 @@
 # ============= standard library imports ========================
 import math
 
-from numpy import array, vstack, mean, average, hstack, zeros, gradient
+from numpy import array, vstack, mean, average, hstack, zeros, gradient, asarray, ones_like, column_stack
 # ============= local library imports  ==========================
 from numpy.linalg import norm
 
@@ -251,6 +251,109 @@ def approximate_polygon_center(pts, r, weight=True):
         mcy = average(cys, weights=ws)
 
     return mcx, mcy
+
+
+TRANSPOSE = array([[0, -1], [1, 0]])
+
+
+def get_intersections(p0, r0, p1, r1):
+    """
+    https://stackoverflow.com/questions/55816902/finding-the-intersection-of-two-circles
+
+
+    :return:
+    """
+
+    x0, y0 = p0
+    x1, y1 = p1
+
+    # circle 1: (x0, y0), radius r0
+    # circle 2: (x1, y1), radius r1
+    dx = x1 - x0
+    dy = y1 - y0
+    d = (dx**2+dy**2)**0.5
+    # d = math.sqrt((x1 - x0) ** 2 + (y1 - y0) ** 2)
+
+    # non intersecting
+    if d > r0 + r1:
+        return None
+    # One circle within other
+    if d < abs(r0 - r1):
+        return None
+    # coincident circles
+    if d == 0 and r0 == r1:
+        return None
+    else:
+        a = (r0 ** 2 - r1 ** 2 + d ** 2) / (2 * d)
+        h = (r0**2-a**2)**0.5
+        # h = math.sqrt(r0 ** 2 - a ** 2)
+        # x2 = x0 + a * (x1 - x0) / d
+        # y2 = y0 + a * (y1 - y0) / d
+        # x3 = x2 + h * (y1 - y0) / d
+        # y3 = y2 - h * (x1 - x0) / d
+        #
+        # x4 = x2 - h * (y1 - y0) / d
+        # y4 = y2 + h * (x1 - x0) / d
+
+        dxd = dx/d
+        dyd = dy/d
+
+        x2 = x0 + a * dxd
+        y2 = y0 + a * dyd
+        x3 = x2 + h * dyd
+        y3 = y2 - h * dxd
+
+        x4 = x2 - h * dyd
+        y4 = y2 + h * dxd
+
+        return x3, y3, x4, y4
+
+
+def approximate_polygon_center3(pts, r, width, height, weight=True):
+    """
+    get 2 points
+    if d btw points is > 2*r skip
+    find the 2 other points that make the 2 isosceles triangles with side r. This is also
+    the intersection of 2 circles centers at the 2 points with radius r
+    take weighted mean. weight by distance to average
+
+    :param pts:
+    :param r:
+    :param weight:
+    :return:
+    """
+    n = len(pts)
+    xs, ys = [], []
+    for k in (3, 4, 5):
+        m = n / k
+        for i, p_i in enumerate(pts):
+            j = (i + m) % n
+            p_j = pts[j]
+            npts = get_intersections(p_i, r, p_j, r)
+            if npts:
+                x3, y3, x4, y4 = npts
+                if 0 < x3 < width:
+                    xs.append(x3)
+                if 0 < x4 < width:
+                    xs.append(x4)
+                if 0 < y3 < height:
+                    ys.append(y3)
+                if 0 < y4 < height:
+                    ys.append(y4)
+
+    xs = array(xs)
+    ys = array(ys)
+    if weight:
+        cxs = ones_like(xs)*xs.mean()
+        cys = ones_like(xs)*ys.mean()
[…]
