+++ pychron/hardware/core/base_core_device.py
@@ -250,7 +250,8 @@ class BaseCoreDevice(HasCommunicator, ConsumerMixin):
 
     def repeat_command(self, cmd, ntries=2, check_val=None, check_type=None,
                        break_val=None,
-                       verbose=True, **kw):
+                       verbose=True,
+                       delay=None, **kw):
 
         if isinstance(cmd, tuple):
             cmd = self._build_command(*cmd)
@@ -277,6 +278,8 @@ class BaseCoreDevice(HasCommunicator, ConsumerMixin):
                 if resp == check_val:
                     break
                 else:
+                    if delay:
+                        time.sleep(delay)
                     continue
 
             if check_type is not None:
