+++ pychron/core/geometry/geometry.py
@@ -18,7 +18,7 @@
 # ============= standard library imports ========================
 import math
 
-from numpy import array, vstack, mean, average, hstack, zeros, gradient, asarray, ones_like, column_stack
+from numpy import array, vstack, mean, average, hstack, zeros, gradient, asarray, ones_like, column_stack, ones
 # ============= local library imports  ==========================
 from numpy.linalg import norm
 
@@ -177,8 +177,9 @@ def find_arc_center(p1, p2, r):
     """
 
     from scipy.optimize import fsolve
-    cx, cy = fsolve(arc_cost_func, [0, 0], args=(p1, p2, r))
-    return cx, cy
+    x, info, status, message = fsolve(arc_cost_func, [0, 0], args=(p1, p2, r), full_output=True)
+    if status == 1:
+        return x
 
 
 def approximage_polygon_center2(pts, r, weight=True):
@@ -223,34 +224,44 @@ def approximate_polygon_center(pts, r, weight=True):
     """
 
     n = len(pts)
-    cxs = []
-    cys = []
+    # cxs = []
+    # cys = []
 
     pts = array(pts)
     pts = vstack((pts, pts))
-    for i in range(2 * n - 5):
-        p1 = pts[i]
-        p2 = pts[i + 4]
-        cx, cy = find_arc_center(p1, p2, r)
-        cxs.append(cx)
-        cys.append(cy)
+    # for i in range(2 * n - 5):
+    #     p1 = pts[i]
+    #     p2 = pts[i + 4]
+    m = int(n / 3)
+    # for i in range(n):
+    #     p1 = pts[i]
+    #     p2 = pts[(i + m) % n]
+    #
+    #     cx, cy = find_arc_center(p1, p2, r)
+    #     cxs.append(cx)
+    #     cys.append(cy)
 
-    mcx = mean(cxs)
-    mcy = mean(cys)
+    cs = [find_arc_center(pts[i], pts[(i + m) % n], r) for i in range(n)]
+    cs = [c for c in cs if c is not None]
+    if cs:
+        cxs, cys = zip(*cs)
 
-    if weight:
-        # pts = sort_clockwise(pts, pts)
-        #        cenx, ceny = calculate_centroid(array(pts))
-        cxs = array(cxs)
-        cys = array(cys)
-        # weight each arc center by the inverse distance to the centroid
-        #        ws = ((cxs - cenx) ** 2 + (cys - ceny) ** 2) ** -0.5
-        # weight each arc center by the inverse distance to the mean
-        ws = ((cxs - mcx) ** 2 + (cys - mcy) ** 2) ** -2
-        mcx = average(cxs, weights=ws)
-        mcy = average(cys, weights=ws)
+        mcx = mean(cxs)
+        mcy = mean(cys)
 
-    return mcx, mcy
+        if weight:
+            # pts = sort_clockwise(pts, pts)
+            #        cenx, ceny = calculate_centroid(array(pts))
+            cxs = array(cxs)
+            cys = array(cys)
+            # weight each arc center by the inverse distance to the centroid
+            #        ws = ((cxs - cenx) ** 2 + (cys - ceny) ** 2) ** -0.5
+            # weight each arc center by the inverse distance to the mean
+            ws = ((cxs - mcx) ** 2 + (cys - mcy) ** 2) ** -2
+            mcx = average(cxs, weights=ws)
+            mcy = average(cys, weights=ws)
+
+        return mcx, mcy
 
 
 TRANSPOSE = array([[0, -1], [1, 0]])
[…]
