+++ pychron/mv/locator.py
@@ -23,7 +23,7 @@ from traits.api import Float
 from numpy import array, histogram, argmax, zeros, asarray, ones_like, \
     nonzero, max, arange, argsort, invert, median, mean, zeros_like
 from operator import attrgetter
-from skimage.morphology import watershed
+# from skimage.morphology import watershed
 from skimage.draw import polygon, circle, circle_perimeter, circle_perimeter_aa
 from scipy import ndimage
 from skimage.exposure import rescale_intensity
@@ -181,71 +181,149 @@ class Locator(Loggable):
         if inverted:
             src = invert(src)
 
-        blocksize_step = search.get('blocksize_step', 5)
-        seg = RegionSegmenter(use_adaptive_threshold=search.get('use_adaptive_threshold', False),
-                              blocksize=search.get('blocksize', 20),
-                              use_watershed=True)
+        # blocksize_step = search.get('blocksize_step', 5)
+        # seg = RegionSegmenter(use_adaptive_threshold=search.get('use_adaptive_threshold', False),
+        #                       blocksize=search.get('blocksize', 20),
+        #                       use_watershed=True)
+        seg = RegionSegmenter()
         fa = self._get_filter_target_area(shape, dim)
-        phigh, plow = None, None
+        # phigh, plow = None, None
 
-        for sdir in (1, -1):
-            for low, high in self._generate_steps(src, search, sdir)():
-                st = time.time()
-                if not self.alive:
-                    self.debug('canceled')
-                    return
+        low, high, prev_m = None, None, None
+        for i in range(10):
+            try:
+                low, high = self._get_bounds(src, search, low, high, prev_m, i)
+            except ValueError:
+                break
 
-                # self.debug('bandwidth low={}, high={}, {}'.format(low, high, sdir))
+            st = time.time()
+            if not self.alive:
+                self.debug('canceled')
+                return
 
-                if inverted:
-                    low = 255 - low
-                    high = 255 - high
+            if inverted:
+                low = 255 - low
+                high = 255 - high
 
-                seg.threshold_low = low
-                seg.threshold_high = high
+            self.debug('bandwidth low={}, high={}'.format(low, high))
+            seg.threshold_low = low
+            seg.threshold_high = high
 
-                if seg.threshold_low == plow and seg.threshold_high == phigh:
-                    return
+            # if seg.threshold_low == plow and seg.threshold_high == phigh:
+            #     return
 
-                plow = seg.threshold_low
-                phigh = seg.threshold_high
+            # plow = seg.threshold_low
+            # phigh = seg.threshold_high
 
-                nsrc = seg.segment(src)
-                seg.blocksize += blocksize_step
+            nsrc = seg.segment(src)
+            # seg.blocksize += blocksize_step
 
-                m = nsrc.mean()
-                if (m > 190 and sdir == -1) or (m < 5 and sdir == 1):
-                    self.debug('image mean out of bounds 5,190 m={}'.format(m))
-                    continue
+            prev_m = nsrc.mean()
+            # if (m > 190 and sdir == -1) or (m < 5 and sdir == 1):
+            if prev_m > 190 or prev_m < 5:
+                self.debug('image mean out of bounds 5,190 m={}'.format(prev_m))
+                continue
 
-                nf = colorspace(nsrc)
+            nf = colorspace(nsrc)
 
-                # draw contours
-                targets = self._find_polygon_targets(nsrc, frame=nf)
-                if set_image and image is not None:
-                    image.set_frame(nf)
+            # draw contours
+            targets = self._find_polygon_targets(nsrc, frame=nf)
+            if set_image and image is not None:
+                image.set_frame(nf)
 
-                if targets:
+            if targets:
+                # filter targets
+                if filter_targets:
[…]
