+++ pychron/mv/locator.py
@@ -14,18 +14,17 @@
 # ===============================================================================
 
 # ============= enthought library imports =======================
-import time
-from collections import deque
-
 from traits.api import Float
 # ============= standard library imports ========================
 
-from numpy import array, histogram, argmax, zeros, asarray, ones_like, \
-    nonzero, max, arange, argsort, invert, median, mean, zeros_like
+import time
 from operator import attrgetter
-# from skimage.morphology import watershed
-from skimage.draw import polygon, circle, circle_perimeter, circle_perimeter_aa
+from collections import deque
+from numpy import array, histogram, argmax, zeros, asarray, ones_like, \
+    nonzero, max, arange, argsort, invert, median, mean, zeros_like, ones, count_nonzero, unique
 from scipy import ndimage
+from skimage.morphology import watershed
+from skimage.draw import polygon, circle, circle_perimeter, circle_perimeter_aa
 from skimage.exposure import rescale_intensity
 from skimage.filters import gaussian
 from skimage import feature
@@ -55,6 +54,10 @@ def draw_circle_perimeter(frame, center_x, center_y, radius, color):
     cy, cx = circle_perimeter(int(center_y), int(center_x), int(radius))
     cy, cx = _coords_inside_image(cy, cx, frame.shape)
     frame[cy, cx] = color
+
+
+def _binary_percent(src):
+    return count_nonzero(src) / src.size
 
 
 class Locator(Loggable):
@@ -190,7 +193,7 @@ class Locator(Loggable):
         # phigh, plow = None, None
 
         low, high, prev_m = None, None, None
-        for i in range(10):
+        for i in range(100):
             try:
                 low, high = self._get_bounds(src, search, low, high, prev_m, i)
             except ValueError:
@@ -217,12 +220,18 @@ class Locator(Loggable):
 
             nsrc = seg.segment(src)
             # seg.blocksize += blocksize_step
-
-            prev_m = nsrc.mean()
+            # per = count_nonzero(nsrc) / nsrc.size
+            # per = 1
+            per = _binary_percent(nsrc)
+            # print(count_nonzero(nsrc), nsrc.shape, nsrc.size, per)
+            if per > 0.65:
+                self.debug('image too bright binary percent {}'.format(per))
+                break
+            # prev_m = nsrc.mean()
             # if (m > 190 and sdir == -1) or (m < 5 and sdir == 1):
-            if prev_m > 190 or prev_m < 5:
-                self.debug('image mean out of bounds 5,190 m={}'.format(prev_m))
-                continue
+            # if prev_m > 190 or prev_m < 5:
+            #     self.debug('image mean out of bounds 5,190 m={}'.format(prev_m))
+            #     continue
 
             nf = colorspace(nsrc)
 
@@ -312,89 +321,93 @@ class Locator(Loggable):
             return self._cached_threshold.pop()
 
         if low is None:
-            band = 128
+            band = 32
             low = src.mean() - band / 2
             high = low + band
 
         else:
-            step = 5
-            if m < 5:  # too much of the scene is dark need to lower the high threshold
-                high -= step
-            elif m>190: # too much of the scene is light need to
-                pass
+            step = 1
+            low += step
+            # high -= step
+            # if m < 5:  # too much of the scene is dark need to lower the high threshold
+            #     high -= step
+            # elif m>190: # too much of the scene is light need to
+            #     pass
+            # else:
 
-        return low, high
+        return low, low
+        # return low, high
 
-    def _generate_steps(self, src, search, shift_dir):
-        self.debug('generate steps search={}'.format(search))
-        self.debug('cached_threshold={}, {}'.format(self._cached_threshold, id(self)))
-        if search.get('use_adaptive_threshold'):
[…]
