### 2021-05-14 d283493
clean up locator
