+++ pychron/mv/locator.py
@@ -162,10 +162,7 @@ class Locator(Loggable):
         """
             use a segmentor to segment the image
         """
-        self.debug('use thrshold caching={}'.format(use_threshold_caching))
-
-        if search is None:
-            search = {}
+        self.debug('use threshold caching={}'.format(use_threshold_caching))
 
         if preprocess:
             if not isinstance(preprocess, dict):
@@ -184,230 +181,65 @@ class Locator(Loggable):
         if inverted:
             src = invert(src)
 
-        # blocksize_step = search.get('blocksize_step', 5)
-        # seg = RegionSegmenter(use_adaptive_threshold=search.get('use_adaptive_threshold', False),
-        #                       blocksize=search.get('blocksize', 20),
-        #                       use_watershed=True)
         seg = RegionSegmenter()
         fa = self._get_filter_target_area(shape, dim)
-        # phigh, plow = None, None
 
-        low, high, prev_m = None, None, None
-        for i in range(100):
-            try:
-                low, high = self._get_bounds(src, search, low, high, prev_m, i)
-            except ValueError:
-                break
+        stargets = []
+        sm = src.mean()
+        for step in (1, -1):
+            if step == 1:
+                threshold = sm - 64
+            else:
+                threshold = sm + 64
 
-            st = time.time()
-            if not self.alive:
-                self.debug('canceled')
-                return
+            for i in range(255):
+                st = time.time()
+                if not self.alive:
+                    self.debug('canceled')
+                    return
 
-            if inverted:
-                low = 255 - low
-                high = 255 - high
+                if self._cached_threshold:
+                    threshold = self._cached_threshold.pop()
+                else:
+                    threshold += step
 
-            self.debug('bandwidth low={}, high={}'.format(low, high))
-            seg.threshold_low = low
-            seg.threshold_high = high
+                if inverted:
+                    # low = 255 - low
+                    # high = 255 - high
+                    threshold = 255 - threshold
 
-            # if seg.threshold_low == plow and seg.threshold_high == phigh:
-            #     return
+                self.debug('threshold={}'.format(threshold))
+                nsrc = seg.segment(src, threshold)
+                per = _binary_percent(nsrc)
+                if per > 0.65:
+                    self.debug('image too bright binary percent {}'.format(per))
+                    break
 
-            # plow = seg.threshold_low
-            # phigh = seg.threshold_high
+                nf = colorspace(nsrc)
+                targets = self._find_polygon_targets(nsrc, frame=nf)
+                if set_image and image is not None:
+                    image.set_frame(nf)
 
-            nsrc = seg.segment(src)
-            # seg.blocksize += blocksize_step
-            # per = count_nonzero(nsrc) / nsrc.size
-            # per = 1
-            per = _binary_percent(nsrc)
-            # print(count_nonzero(nsrc), nsrc.shape, nsrc.size, per)
-            if per > 0.65:
-                self.debug('image too bright binary percent {}'.format(per))
-                break
-            # prev_m = nsrc.mean()
-            # if (m > 190 and sdir == -1) or (m < 5 and sdir == 1):
-            # if prev_m > 190 or prev_m < 5:
-            #     self.debug('image mean out of bounds 5,190 m={}'.format(prev_m))
-            #     continue
+                if targets:
+                    # filter targets
+                    if filter_targets:
+                        targets = self._filter_targets(image, frame, dim, targets, fa)
+                    elif convexity_filter:
+                        targets = [t for t in targets if t.perimeter_convexity > convexity_filter]
 
-            nf = colorspace(nsrc)
+                et = time.time() - st
+                self.debug('targets={} et={}'.format(len(targets), et))
+                if targets:
+                    # remember this thresholding for the next time
[…]
