+++ pychron/mv/autocenter_manager.py
@@ -165,9 +165,12 @@ class AutoCenterManager(MachineVisionManager):
 class CO2AutocenterManager(AutoCenterManager):
     # private
     def _get_locator(self, *args, **kw):
-        from pychron.mv.co2_locator import CO2Locator
-        return CO2Locator(pxpermm=self.pxpermm, pixel_depth=self.video.pixel_depth)
-
+        if self.locator:
+            loc = self.locator
+        else:
+            from pychron.mv.co2_locator import CO2Locator
+            loc = CO2Locator(pxpermm=self.pxpermm, pixel_depth=self.video.pixel_depth)
+        return loc
 
 class DiodeAutocenterManager(AutoCenterManager):
     # private
