+++ pychron/experiment/experiment_executor.py
@@ -759,14 +759,14 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
         self.debug('join run finished')
 
     def _set_prev(self, run):
-        at = run.spec.analysis_type
         if hasattr(run, 'get_baseline_corrected_signal_dict'):
             d = (run.record_id, run.get_baseline_corrected_signal_dict())
             if self.use_preceding_blank:
-                self._prev_blanks[at] = d
+                self._prev_blanks[run.analysis_type] = d
                 self._prev_blanks['blank'] = d
             self._prev_baselines = run.get_baseline_dict()
         else:
+            at = run.spec.analysis_type
             if self.use_preceding_blank and at.startswith('blank'):
                 pb = run.get_baseline_corrected_signals()
                 if pb is not None:
@@ -1197,7 +1197,7 @@ class ExperimentExecutor(Consoleable, PreferenceMixin):
         try:
             pb = self._prev_blanks[spec.analysis_type]
         except KeyError:
-            pb = self._prev_blanks.get('blank', ({}, ''))
+            pb = self._prev_blanks.get('blank', ('', {}))
 
         arun.previous_blanks = pb
         arun.previous_baselines = self._prev_baselines
