### 2021-05-26 abc4084
added new flux visualization
