+++ pychron/pipeline/editors/flux_visualization_editor.py
@@ -14,9 +14,10 @@
 # limitations under the License.
 # ===============================================================================
 from operator import itemgetter
+from pprint import pprint
 
-from numpy import linspace, meshgrid, arctan2, sin, cos, vstack, array, zeros, diff, argwhere
-from traits.api import Instance, Int, Str, Float, Property, List, on_trait_change
+from numpy import linspace, meshgrid, arctan2, sin, cos, vstack, array, zeros, diff, argwhere, isnan, argmin, argmax
+from traits.api import Instance, Int, Str, Float, Property, List, on_trait_change, Button
 from traitsui.api import View, UItem, VGroup, HGroup, TableEditor, Tabbed, HSplit
 from traitsui.table_column import ObjectColumn
 from uncertainties import nominal_value, std_dev
@@ -39,10 +40,13 @@ from pychron.graph.regression_graph import RegressionGraph
 from pychron.graph.stacked_graph import StackedGraph
 from pychron.graph.tools.data_tool import DataTool, DataToolOverlay
 from pychron.options.layout import FigureLayout
+from pychron.paths import paths
 from pychron.pipeline.editors.irradiation_tray_overlay import IrradiationTrayOverlay
 from pychron.pychron_constants import LEAST_SQUARES_1D, MATCHING, BRACKETING, WEIGHTED_MEAN, BOWL, PLANE, \
     WEIGHTED_MEAN_1D, MSEM, NN, format_mswd, BSPLINE, RBF, GRIDDATA, IDW
 
+HEADER_KEYS = ['hole_id', 'identifier', 'x', 'y', 'z', 'saved_j', 'saved_jerr',
+               'mean_j', 'mean_jerr', 'mean_j_mswd', 'min_j', 'max_j', 'variation_percent_j']
 
 def make_grid(r, n):
     xi = linspace(-r, r, n)
@@ -281,9 +285,14 @@ class BaseFluxVisualizationEditor(BaseTraitsEditor):
         reg.calculate()
         return reg
 
-    def _model_flux(self, reg, r):
-        n = reg.n * 10
+    def _model_flux(self, reg, r, total=True, n=None, origin=None):
+        if n is None:
+            n = reg.n * 10
+
         gx, gy = make_grid(r, n)
+        if origin:
+            gx += origin[0]
+            gy += origin[1]
 
         nz = zeros((n, n))
         ne = zeros((n, n))
@@ -299,9 +308,9 @@ class BaseFluxVisualizationEditor(BaseTraitsEditor):
                 nominals = reg.predict(pts)
                 nz[i] = nominals
 
-        print(nz.shape)
-        self.max_j = nz.max()
-        self.min_j = nz.min()
+        if total:
+            self.max_j = nz.max()
+            self.min_j = nz.min()
 
         return gx, gy, nz, ne
 
@@ -326,26 +335,27 @@ class BaseFluxVisualizationEditor(BaseTraitsEditor):
                                      geometry=self.geometry,
                                      show_labels=self.show_labels)
         self.irradiation_tray_overlay = ito
-        center_plot.overlays.append(ito)
+        center_plot.underlays.append(ito)
 
         gx, gy, m, me = self._model_flux(reg, r)
         # self._visualization_update(gx, gy, m, me, reg.xs, reg.ys)
 
-        r = r * 1.1
-        s, p = cg.new_series(z=m,
-                             xbounds=(-r, r),
-                             ybounds=(-r, r),
-                             levels=self.plotter_options.levels,
-                             cmap=self.plotter_options.color_map_name,
-                             colorbar=True,
-                             style='contour')
+        b = r
+        center, p = cg.new_series(z=m,
+                                  xbounds=(-b, b),
+                                  ybounds=(-b, b),
[…]
