### 2021-06-11 5576a3c
added save_all_runs preference
