+++ pychron/experiment/tasks/experiment_task.py
@@ -41,7 +41,8 @@ from pychron.globals import globalv
 from pychron.lasers.laser_managers.ilaser_manager import ILaserManager
 from pychron.paths import paths
 from pychron.pipeline.plot.editors.figure_editor import FigureEditor
-from pychron.pychron_constants import SPECTROMETER_PROTOCOL, DVC_PROTOCOL, COCKTAIL, AIR, BLANK, FUSIONS_UV
+from pychron.pychron_constants import SPECTROMETER_PROTOCOL, DVC_PROTOCOL, COCKTAIL, AIR, BLANK, FUSIONS_UV, STATES, \
+    INVALID, EXTRACTION, MEASUREMENT, CANCELED, TRUNCATED, END_AFTER, FAILED, SUCCESS
 
 
 class ExperimentEditorTask(EditorTask):
@@ -256,8 +257,7 @@ class ExperimentEditorTask(EditorTask):
 
     def _assemble_state_colors(self):
         colors = {}
-        for c in ('success', 'extraction', 'measurement', 'canceled', 'truncated',
-                  'failed', 'end_after', 'invalid'):
+        for c in (SUCCESS, EXTRACTION, MEASUREMENT, CANCELED, TRUNCATED, FAILED, END_AFTER, INVALID):
             v = self.application.preferences.get('pychron.experiment.{}_color'.format(c))
             colors[c] = v or '#FFFFFF'
 
