+++ pychron/experiment/automated_run/spec.py
@@ -34,7 +34,8 @@ from pychron.experiment.utilities.runid import make_rid, make_runid
 from pychron.pychron_constants import SCRIPT_KEYS, SCRIPT_NAMES, DETECTOR_IC, DURATION, EXTRACT_VALUE, EXTRACT_UNITS, \
     RAMP_RATE, RAMP_DURATION, BEAM_DIAMETER, PRECLEANUP, CLEANUP, POSTCLEANUP, DELAY_AFTER, PATTERN, LIGHT_VALUE, TRAY, \
     MASS_SPECTROMETER, POSITION, USE_CDD_WARMING, EXTRACT_DEVICE, COLLECTION_TIME_ZERO_OFFSET, WEIGHT, COMMENT, PROJECT, \
-    SAMPLE, MATERIAL, REPOSITORY_IDENTIFIER, DISABLE_BETWEEN_POSITIONS, USERNAME, NULL_STR, CRYO_TEMP
+    SAMPLE, MATERIAL, REPOSITORY_IDENTIFIER, DISABLE_BETWEEN_POSITIONS, USERNAME, NULL_STR, CRYO_TEMP, CANCELED, FAILED, \
+    TRUNCATED, SUCCESS, EXTRACTION, MEASUREMENT, INVALID, NOT_RUN, ABORTED
 
 logger = new_logger('AutomatedRunSpec')
 
@@ -48,10 +49,10 @@ class AutomatedRunSpec(HasTraits):
     spectrometer_manager = Instance('pychron.spectrometer.base_spectrometer_manager.BaseSpectrometerManager')
 
     result = Instance(AutomatedRunResult, ())
-    state = Enum('not run', 'extraction',
-                 'measurement', 'success',
-                 'failed', 'truncated', 'canceled',
-                 'invalid', 'test', 'aborted')
+    state = Enum(NOT_RUN, EXTRACTION,
+                 MEASUREMENT, SUCCESS,
+                 FAILED, TRUNCATED, CANCELED,
+                 INVALID, 'test', ABORTED)
 
     skip = Bool(False)
     end_after = Bool(False)
