+++ pychron/hardware/actuators/mcc_gp_actuator.py
@@ -1,0 +1,28 @@
+# ===============================================================================
+# Copyright 2021 ross
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ===============================================================================
+from pychron.hardware.actuators.client_gp_actuator import ClientMixin
+from pychron.hardware.actuators.gp_actuator import GPActuator
+from pychron.hardware.core.modbus import ModbusMixin
+
+
+class MCCGPActuator(GPActuator, ClientMixin):
+    def _actuate(self, obj, action):
+        pass
+
+    def get_channel_state(self, address, *args, **kw):
+        pass
+
+# ============= EOF =============================================
