+++ pychron/pipeline/tables/xlsx_table_writer.py
@@ -237,6 +237,10 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         # remove non enabled columns
         columns = [c for c in columns if c.enabled]
 
+        # remove non visible columns
+        if self._options.exclude_hidden_columns:
+            columns = [c for c in columns if c.visible]
+
         return columns
 
     def _flux_columns(self, columns):
@@ -503,28 +507,31 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         return cols
 
     def _make_human_unknowns(self, unks):
-        return self._make_sheet(unks, 'Unknowns')
+        name = self._options.human_sheet_name
+        return self._make_sheet(unks, name, 'unknown')
 
     def _make_machine_unknowns(self, unks):
-        self._make_machine_sheet(unks, 'Unknowns (Machine)')
+        name = self._options.machine_sheet_name
+        self._make_machine_sheet(unks, name)
 
     def _make_airs(self, airs):
-        self._make_sheet(airs, 'Airs')
+        self._make_sheet(airs, 'Airs', 'airs')
 
     def _make_blanks(self, blanks):
-        self._make_sheet(blanks, 'Blanks')
+        self._make_sheet(blanks, 'Blanks', 'blanks')
 
     def _make_monitors(self, monitors):
-        self._make_sheet(monitors, 'Monitors')
+        self._make_sheet(monitors, 'Monitors', 'monitors')
 
     def _make_summary_sheet(self, unks):
         self._current_row = 1
-        sh = self._workbook.add_worksheet('Summary')
+        name = self._options.summary_sheet_name
+        sh = self._workbook.add_worksheet(name)
         self._format_generic_worksheet(sh)
 
         cols = self._get_summary_columns()
         cols = [c for c in cols if c.visible]
-        self._make_title(sh, 'Summary', cols)
+        self._make_title(sh, 'Summary', cols, key='summary')
 
         fmt = self._workbook.add_format({'bottom': 1, 'align': 'center'})
         sh.set_row(self._current_row, 5)
@@ -603,7 +610,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
 
         return ngs
 
-    def _make_sheet(self, groups, name):
+    def _make_sheet(self, groups, name, key):
         self._current_row = 1
 
         worksheet = self._workbook.add_worksheet(name)
@@ -611,7 +618,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         cols = self._get_columns(name, groups)
         self._format_worksheet(worksheet, cols, (8, 2))
 
-        self._make_title(worksheet, name, cols)
+        self._make_title(worksheet, name, cols, key=key)
 
         options = self._options
         repeat_header = options.repeat_header
@@ -682,7 +689,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
 
             self._current_row += 1
 
-        self._make_notes(groups, worksheet, len(cols), name)
+        self._make_notes(groups, worksheet, len(cols), key)
         self._current_row = 1
 
         for i, c in enumerate(cols):
@@ -702,7 +709,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         cols = self._get_machine_columns(name, groups)
         self._format_worksheet(worksheet, cols, (5, 2))
 
-        self._make_title(worksheet, name, cols)
+        self._make_title(worksheet, name, cols, key='machine')
 
         repeat_header = self._options.repeat_header
 
@@ -731,19 +738,25 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
             if not c.visible:
[…]
