+++ pychron/lasers/pattern/pattern_executor.py
@@ -24,6 +24,7 @@ from threading import Event
 from numpy import polyfit, array, average, uint8, zeros_like
 from skimage.color import gray2rgb
 from skimage.draw import circle
+from skimage.draw._draw import _coords_inside_image
 from traits.api import Any, Bool
 
 from pychron.core.ui.gui import invoke_in_main_thread
@@ -392,9 +393,11 @@ class PatternExecutor(Patternable):
         per_img = zeros_like(oimg, dtype='int16')
 
         img_h, img_w = pos_img.shape
-        perimeter_circle = circle(img_h / 2, img_w / 2, pattern.perimeter_radius * pxpermm)
+        pcx, pcy = circle(img_h / 2, img_w / 2, pattern.perimeter_radius * pxpermm)
 
         color = 2 ** 15 - 1
+
+        perimeter_circle = _coords_inside_image(pcx, pcy, per_img.shape)
         per_img[perimeter_circle] = 50
         set_data('imagedata', gray2rgb(per_img.astype(uint8)))
 
