### 2021-06-16 e5abc34
fixed scatter inspection. added sclf filtering
