+++ pychron/pychron_constants.py
@@ -86,7 +86,13 @@ MSE = 'SE but if MSWD>1 use SE * sqrt(MSWD)'
 ERROR_TYPES = [MSEM, SEM, SD]
 SIG_FIGS = range(0, 15)
 STD_SIG_FIGS = ['Std', 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]
+SCHAEN2020_1 = 'Schaen 2020. (Low MSWD weighted mean)'
+SCHAEN2020_2 = 'Schaen 2020. (Weighted mean filter)'
+SCHAEN2020_3 = 'Schaen 2020. (Normality)'
+DEINO = 'Deino. Filter Extreme'
 
+CUMULATIVE = 'Cumulative'
+KERNEL = 'Kernel'
 WEIGHTED_MEAN = 'Weighted Mean'
 PLATEAU = 'Plateau'
 INTEGRATED = 'Total Integrated'
