### 2021-06-16 2ec1764
improved analysis type coloring
