### 2021-06-16 f88608b
improved setting commit author
