### 2021-06-17 8063083
updated browser view. addressed #1996
