+++ pychron/envisage/browser/sample_view.py
@@ -49,7 +49,7 @@ class GroupView(Controller):
             traits.
         """
         return {'object': self.model, 'controller': self.controller,
-                'handler': self.controller, 'pane': self.pane, 'analysis_table': self.model.analysis_table}
+                'handler': self.controller, 'pane': self.pane, 'table': self.model.table}
         return ctx
 
 
@@ -185,56 +185,56 @@ class SampleGroupView(GroupView):
 class AnalysisGroupView(GroupView):
 
     def traits_view(self):
-        analysis_tools = VGroup(HGroup(UItem('analysis_table.analysis_set',
+        analysis_tools = VGroup(HGroup(UItem('table.analysis_set',
                                              width=250,
-                                             editor=EnumEditor(name='analysis_table.analysis_set_names')),
-                                       icon_button_editor('analysis_table.refresh_analysis_set_button', 'refresh',
-                                                          enabled_when='analysis_table.items',
+                                             editor=EnumEditor(name='table.analysis_set_names')),
+                                       icon_button_editor('table.refresh_analysis_set_button', 'refresh',
+                                                          enabled_when='table.items',
                                                           tooltip='Reload selected analysis set'),
-                                       icon_button_editor('analysis_table.add_analysis_set_button', 'add',
-                                                          enabled_when='analysis_table.items',
+                                       icon_button_editor('table.add_analysis_set_button', 'add',
+                                                          enabled_when='table.items',
                                                           tooltip='Add current analyses to an analysis set'),
                                        icon_button_editor('add_analysis_group_button', 'database_add',
-                                                          enabled_when='analysis_table.items',
+                                                          enabled_when='table.items',
                                                           tooltip='Add current analyses to an analysis group')),
-                                HGroup(UItem('analysis_table.analysis_filter_parameter',
+                                HGroup(UItem('table.analysis_filter_parameter',
                                              width=-90,
-                                             editor=EnumEditor(name='analysis_table.analysis_filter_parameters')),
-                                       UItem('analysis_table.analysis_filter'),
-                                       icon_button_editor('analysis_table.scroll_to_bottom', 'arrow_down',
+                                             editor=EnumEditor(name='table.analysis_filter_parameters')),
+                                       UItem('table.analysis_filter'),
+                                       icon_button_editor('table.scroll_to_bottom', 'arrow_down',
                                                           tooltip='Scroll to bottom'),
-                                       icon_button_editor('analysis_table.scroll_to_top', 'arrow_up',
+                                       icon_button_editor('table.scroll_to_top', 'arrow_up',
                                                           tooltip='Scroll to top'),
                                        Item('use_quick_recall', label='Quick Recall')))
 
         agrp = Group(VGroup(analysis_tools,
-                            UItem('analysis_table.analyses',
+                            UItem('table.analyses',
                                   # width=0.75,
-                                  editor=myTabularEditor(adapter=self.model.analysis_table.tabular_adapter,
+                                  editor=myTabularEditor(adapter=self.model.table.tabular_adapter,
                                                          operations=['move', 'delete'],
-                                                         column_clicked='analysis_table.column_clicked',
-                                                         refresh='analysis_table.refresh_needed',
-                                                         selected='analysis_table.selected',
-                                                         dclicked='analysis_table.dclicked',
+                                                         column_clicked='table.column_clicked',
+                                                         refresh='table.refresh_needed',
+                                                         selected='table.selected',
[…]
