### 2021-06-19 #3386 @whedon
[REVIEW]: Pychron: Automated Data Collection and Reduction for Noble Gas and Ar/Ar Geochronology
**Submitting author:** @jirhiker (<a href="http://orcid.org/0000-0003-4727-7472">Jake Ross</a>)
**Repository:** https://github.com/NMGRL/pychron
**Branch with paper.md** (empty if default branch):
**Version:** v20.1
