+++ pychron/lasers/tasks/laser_preferences.py
@@ -74,6 +74,7 @@ class FusionsLaserPreferences(LaserPreferences):
     crosshairs_offset_color = Color('blue')
     crosshairs_line_width = Float(1.0)
 
+    aux_crosshairs_kind = Enum('BeamRadius', 'UserRadius', 'MaskRadius')
     aux_crosshairs_radius = Range(0.0, 10.0, 1.0)
     aux_crosshairs_offsetx = Float(0)
     aux_crosshairs_offsety = Float(0)
