+++ pychron/mv/locator.py
@@ -21,7 +21,7 @@ import time
 from operator import attrgetter
 from collections import deque
 from numpy import array, histogram, argmax, zeros, asarray, ones_like, \
-    nonzero, max, arange, argsort, invert, median, mean, zeros_like, ones, count_nonzero, unique
+    nonzero, arange, argsort, invert, median, mean, zeros_like, ones, count_nonzero, unique
 from scipy import ndimage
 from skimage.morphology import watershed
 from skimage.draw import polygon, circle, circle_perimeter, circle_perimeter_aa
@@ -58,6 +58,21 @@ def draw_circle_perimeter(frame, center_x, center_y, radius, color):
 
 def _binary_percent(src):
     return count_nonzero(src) / src.size
+
+
+def _arc_approximation_filter(src, targets, dim, tol=0.5):
+    dim = round(dim)
+    nt = []
+    dim *= 1.1
+    for target in targets:
+        if target.convexity > tol:
+            # self.info('doing arc approximation radius={}'.format(dim))
+            pts = target.poly_points
+            h, w = src.shape[0], src.shape[1]
+            args = approximate_polygon_center3(pts, dim, w, h)
+            if args:
+                nt.append(target)
+    return nt
 
 
 class Locator(Loggable):
@@ -99,7 +114,7 @@ class Locator(Loggable):
         # cw_px = ch_px = cw_px + r
         if verbose:
             self.debug('Crop: x={},y={}, cw={}, ch={}, '
-                       'width={}, height={}, ox={}, oy={}'.format(x, y, cw_px, ch_px, w, h, ox, oy))
+                       'width={}, height={}, ox={}, oy={}, pxpermm={}'.format(x, y, cw_px, ch_px, w, h, ox, oy, self.pxpermm))
         return asarray(crop(src, x, y, cw_px, ch_px))
 
     def find(self, image, frame, dim, shape='circle', confirm=False, **kw):
@@ -119,7 +134,7 @@ class Locator(Loggable):
         dx, dy = None, None
 
         st = time.time()
-        targets = self._find_targets(image, frame, dim, shape=shape, **kw)
+        targets = self._find_targets(image, frame, dim, shape=shape, do_arc_filter=self.use_arc_approximation, **kw)
         self.debug('time to find targets={:0.5f}'.format(time.time() - st))
         if targets:
             self.info('found {} potential targets'.format(len(targets)))
@@ -143,7 +158,8 @@ class Locator(Loggable):
 
                 # image.set_frame(src[:])
 
-            self._draw_center_indicator(src, size=50, shape='crosshairs', color=(255, 255, 0))
+            self._draw_center_indicator(src, size=max(10, int(src.shape[0]*0.25)),
+                                        shape='crosshairs', color=(255, 255, 0))
 
         image.refresh_needed = True
         self.info('dx={}, dy={}'.format(dx, dy))
@@ -159,7 +175,7 @@ class Locator(Loggable):
                       convexity_filter=False,
                       mask=False,
                       set_image=True, inverted=False,
-                      use_threshold_caching=False, dual_search=False):
+                      use_threshold_caching=False, dual_search=False, do_arc_filter=False, recurse=False):
         """
             use a segmentor to segment the image
         """
@@ -187,13 +203,19 @@ class Locator(Loggable):
 
         stargets = []
         sm = src.mean()
+        self.debug('recurse= {}'.format(recurse))
+        if isinstance(recurse, float):
+            othreshold = recurse+10
+            sm = max(255, othreshold+32)
+            recurse = False
 
-        othreshold = None
-        if self._cached_threshold:
-            othreshold = self._cached_threshold.pop()
+        else:
+            othreshold = None
[…]
