### 2021-06-22 d3016e1
added convex hull to arc approximation
