+++ pychron/mv/locator.py
@@ -21,7 +21,7 @@ import time
 from operator import attrgetter
 from collections import deque
 from numpy import array, histogram, argmax, zeros, asarray, ones_like, \
-    nonzero, arange, argsort, invert, median, mean, zeros_like, ones, count_nonzero, unique
+    nonzero, arange, argsort, invert, median, mean, zeros_like, ones, count_nonzero, unique, isnan
 from scipy import ndimage
 from skimage.morphology import watershed
 from skimage.draw import polygon, circle, circle_perimeter, circle_perimeter_aa
@@ -119,7 +119,7 @@ class Locator(Loggable):
                                                                               self.pxpermm))
         return asarray(crop(src, x, y, cw_px, ch_px))
 
-    def find(self, image, frame, dim, shape='circle', confirm=False, **kw):
+    def find2(self, image, frame, dim, shape='circle', confirm=False, **kw):
         """
             image is a stand alone image
             dim = float. radius or half length of a square in pixels
@@ -170,6 +170,89 @@ class Locator(Loggable):
                 dx, dy = None, None
         self.debug('total find time={:0.5f}'.format(time.time() - st))
         return dx, dy
+
+    def find(self, image, frame, dim, shape='circle', confirm=False, **kw):
+        self.alive = True
+        dx, dy = None, None
+
+        st = time.time()
+        targets = self._find_targets_bs(image, frame, dim, shape=shape, **kw)
+        self.debug('time to find targets={:0.5f}'.format(time.time() - st))
+        if targets:
+            self.info('found {} potential targets'.format(len(targets)))
+            src = image.source_frame
+            # draw targets
+            self._draw_targets(src, targets)
+
+            dx, dy = self._calculate_error(targets)
+            self._draw_center_indicator(src, size=max(10, int(src.shape[0] * 0.25)),
+                                        shape='crosshairs')
+
+        image.refresh_needed = True
+        self.info('dx={}, dy={}'.format(dx, dy))
+        if confirm:
+            if not self.confirmation_dialog('Move to position'):
+                dx, dy = None, None
+        self.debug('total find time={:0.5f}'.format(time.time() - st))
+        return dx, dy
+
+    def _find_targets_bs(self, image, frame, dim, shape, preprocess, **kw):
+        if preprocess:
+            if not isinstance(preprocess, dict):
+                preprocess = {}
+            src = self._preprocess(frame, **preprocess)
+        else:
+            src = grayspace(frame)
+
+        image.set_frame(colorspace(frame / self.pixel_depth * 255))
+
+        seg = RegionSegmenter()
+        fa = self._get_filter_target_area(shape, dim)
+
+        sm = src.mean()
+        low = sm / 2
+        high = 255 - low
+        step = 5
+
+        self.debug('src mean={}, {}, {}'.format(sm, low, high))
+
+        def find(t):
+
+            nsrc = seg.segment(src, t)
+            per = _binary_percent(nsrc)
+            if per > 0.85 or per < 0.25:
+                return []
+
+            ts = self._find_polygon_targets(nsrc)
+            ts = self._filter_targets(image, frame, dim, ts, fa, use_segmentation=False)
+            # ts = _arc_approximation_filter(frame, ts, dim)
+
+            return ts
+
+        visited = {}
+
+        def find_bs(threshold):
+            if threshold in visited:
+                return []
+
+            if threshold > high or threshold < low:
+                return []
+
+            tt = find(threshold)
+            visited[threshold] = len(tt) if tt else 0
+            if tt:
+                at = find_bs(int(threshold-step))
+                bt = find_bs(int(threshold+step))
+
+                if at:
+                    tt.extend(at)
+                if bt:
[…]
