### 2021-06-25 a2ae4ae
updated binary search
