+++ pychron/pipeline/engine.py
@@ -1170,10 +1170,10 @@ class PipelineEngine(Loggable):
                 setattr(self, lc, 0)
                 setattr(self, lr, n)
 
-        if getattr(self, lc) >= getattr(self, lr) or n == 1:
-            setattr(self, lc, 0)
-            func(editor)
-            if editor:
+            if getattr(self, lc) >= getattr(self, lr) or n == 1:
+                setattr(self, lc, 0)
+                func(editor)
+                # if editor:
                 editor.refresh_needed = True
 
     def _dclicked_changed(self, new):
