+++ app_utils/preferences/github.ini
@@ -1,0 +1,3 @@
+[pychron.github]
+organization=NMGRLDataDev
+oauth_token=
