+++ app_utils/preferences/dvc.ini
@@ -1,0 +1,12 @@
+[pychron.dvc]
+use_cache = False
+use_auto_pull = True
+update_currents_enabled = True
+max_cache_size = 100
+
+[pychron.dvc.connection]
+name = pychrondvc
+path = Database/index.sqlite3
+fav_name = Demo
+favorites = """['Demo,sqlite,,,,,True,True,/home/Pychron/demo.sqlite,NMGRLDataDev,MetaData,,1']"""
+kind = sqlite

[…]
