+++ app_utils/preferences/github.ini
@@ -1,3 +1,3 @@
 [pychron.github]
-organization=NMGRLDataDev
-oauth_token=
+organization=PychronTrainingData
+oauth_token=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
