+++ Dockerfile
@@ -18,9 +18,10 @@ RUN conda install -c dbanas chaco
 RUN pip3 install uncertainties qimage2ndarray peakutils
 
 # copy src
-COPY ./pychron ./pychron
-COPY ./launchers ./launchers
-COPY ./resources ./resources
+RUN git clone https://github.com/NMGRL/pychron
+WORKDIR pychron
+RUN git checkout dev/docker
+WORKDIR ..
 
 # copy setupfiles
 RUN mkdir /home/Pychron
@@ -29,12 +30,20 @@ COPY ./app_utils/preferences /home/Pychron/preferences
 COPY ./app_utils/demo.sqlite /home/Pychron/demo.sqlite
 COPY ./app_utils/.appdata /home/Pychron/.appdata
 
+RUN mkdir /home/Pychron/data
+RUN mkdir /home/Pychron/data/.dvc
+RUN mkdir /home/Pychron/data/.dvc/repositories
+
+RUN git clone https://github.com/PychronTrainingData/Irradiation-NM-313 /home/Pychron/data/.dvc/repositories/Irradiation-NM-313
+RUN git clone https://github.com/PychronTrainingData/Morgan01099 /home/Pychron/data/.dvc/repositories/Morgan01099
+RUN git clone https://github.com/PychronTrainingData/MetaData /home/Pychron/data/.dvc/MetaData
+
 # setup environment
 ENV QT_GRAPHICSSYSTEM="native"
 ENV QT_DEBUG_PLUGINS=0
-ENV PYTHONPATH=.
+ENV PYTHONPATH=./pychron
 ENV PYCHRON_APPNAME=pycrunch
 ENV PYCHRON_ENV=/home/Pychron
 ENV PYCHRON_USE_LOGIN=0
 
-ENTRYPOINT ["conda", "run", "--no-capture-output", "--name", "pychron3", "python", "./launchers/launcher.py"]
+ENTRYPOINT ["conda", "run", "--no-capture-output", "--name", "pychron3", "python", "./pychron/launchers/launcher.py"]
