+++ app_utils/preferences/arar_constants.ini
@@ -1,0 +1,2 @@
+[pychron.arar_constants]
+age_units = Ma
