+++ app_utils/preferences/github.ini
@@ -1,3 +1,0 @@
-[pychron.github]
-organization=PychronTrainingData
-oauth_token=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
