+++ app_utils/install.sh
@@ -1,406 +1,0 @@
-#!/usr/bin/env bash
-# This script is used to install pychron
-# First in downloads and installs miniconda
-# miniconda is used to manage the python dependencies
-# conda is used to create a new environment
-# conda and pip are used to install the dependencies
-# a Pychron support directory is created and some boilerplate support files are written
-# the pychron source code is downloaded from the available releases stored at github
-#    (i.e the source is not a git clone just a static directory)
-# if update_flag is set then clone the repository into .pychron/updates/
-# the source code is stored in the Pychron support directory
-# a launcher script is created and copied to the desktop
-
-# =========== Front Matter ==============
-
-cat << "EOF"
-  _______     _______ _    _ _____   ____  _   _
- |  __ \ \   / / ____| |  | |  __ \ / __ \| \ | |
- | |__) \ \_/ / |    | |__| | |__) | |  | |  \| |
- |  ___/ \   /| |    |  __  |  _  /| |  | | . ` |
- | |      | | | |____| |  | | | \ \| |__| | |\  |
- |_|      |_|  \_____|_|  |_|_|  \_\\____/|_| \_|
-
-
-Developer: Jake Ross (NMT)
-Date: 10-02-2016
----*---*---*---*---*---*---*---*---*---*---*---*
-Welcome to the Pychron Installer.
-
-Hit "Enter" to continue
-
----*---*---*---*---*---*---*---*---*---*---*---*
-EOF
-
-read wait
-
-cat << "EOF"
-You will be asked to provide a series of configuration values. Each value has as default value, indicated in square
-brackets e.g., [default]
-
-To except the default value hit "Enter" when prompted
-
-
-!!!WARNING!!!
-This installer is beta and not guaranteed to work. USE WITH CAUTION
-
-Hit "Enter" to continue
-EOF
-read wait
-
-if type git >/dev/null
-then
-    echo 'Git is already installed'
-else
-    echo "Git is required"
-    exit
-fi
-
-if type gcc >/dev/null
-then
-    echo 'XCode Commandline tools are already installed'
-else
-    echo "XCode Commandline tools are required"
-    xcode-select --install
-fi
-# =========== User Questions ==============
-echo "In order to submit issues to the Pychron developers you need to have a Github account"
-
-default=NMGRL
-echo -n "Github organization [$default] >> "
-read go
-[ -z "$go" ] && go=$default
-
-default=nmgrluser
-echo -n "Github user name [$default] >> "
-read gu
-[ -z "$gu" ] && gu=$default
-
-echo -n "Github password for ${gu} >> "
-read gp
-
-default=16
-echo -n "MassSpec Database Version [$default] >> "
-read dbv
-[ -z "$dbv" ] && dbv=$default
-
-default=NMGRL
-echo -n "Pychron Fork [$default] >> "
-read pychron_fork
-[ -z "$pychron_fork" ] && pychron_fork=$default
-
-default=release/py3/v18.2
[…]
