+++ app_utils/install.py
@@ -22,6 +22,487 @@ import sys
 IS_MAC = platform.system() == 'Darwin'
 IS_WINDOWS = platform.system() == 'Windows'
 
+INITIALIZATION_TXT = '''<root>
+    <globals>
+    </globals>
+    <plugins>
+        <general>
+            <plugin enabled="false">CanvasDesigner</plugin>
+            <plugin enabled="true">PyScript</plugin>
+            <plugin enabled="false">Video</plugin>
+            <plugin enabled="false">Database</plugin>
+            <plugin enabled="true">Entry</plugin>
+            <plugin enabled="false">SystemMonitor</plugin>
+            <plugin enabled="true">ArArConstants</plugin>
+            <plugin enabled="false">Loading</plugin>
+            <plugin enabled="false">LabBook</plugin>
+            <plugin enabled="false">DashboardServer</plugin>
+            <plugin enabled="false">DashboardClient</plugin>
+            <plugin enabled="true">Pipeline</plugin>
+            <plugin enabled="false">Geochron</plugin>
+            <plugin enabled="false">MediaStorage</plugin>
+            <plugin enabled="true">Update</plugin>
+            <plugin enabled="true">DVC</plugin>
+            <plugin enabled="false">MassSpec</plugin>
+            <plugin enabled="false">IGSN</plugin>
+            <plugin enabled="true">GitHub</plugin>
+            <plugin enabled="false">MachineLearning</plugin>
+            <plugin enabled="false">Image</plugin>
+            <plugin enabled="false">RemoteExperiment</plugin>
+            <plugin enabled="false">Experiment</plugin>
+            <plugin enabled="false">LabspyClient</plugin>
+            <plugin enabled="false">Classifier</plugin>
+            <plugin enabled="false">Auto</plugin>
+            <plugin enabled="false">MDD</plugin>
+            <plugin enabled="false">Query</plugin>
+            <plugin enabled="false">GitLab</plugin>
+        </general>
+        <hardware>{}
+        </hardware>
+        <social>
+        </social>
+        <data>
+            <plugin enabled="false">USGSVSCData</plugin>
+            <plugin enabled="false">WiscArData</plugin>
+            <plugin enabled="false">Sparrow</plugin>
+        </data>
+    </plugins>
+</root>
+'''
+
+HARDWARE_PLUGIN_TXT = """
+            <plugin enabled="true">ExtractionLine
+                <manager enabled="true">switch_manager
+                <device enabled="true">switch_controller</device>
+                </manager>
+            </plugin>"""
+
+DEFAULTS_TXT = '''#script defaults file
+#defines default pyscripts for the analysis types
+#this yaml file returns a dictionary of dictionaries
+#.py does not have to be included for graphic clarity.
+Bg:
+ extraction: 
+ measurement: 
+ post_equilibration:
+ post_measurement: 
+Ba:
+ extraction: air_x1
+ measurement: air
+ post_equilibration: pump_air
+ post_measurement: pump_ms
+Bc:
+ extraction: cocktail_x1
+ measurement: air
+ post_equilibration: pump_air
+ post_measurement: pump_ms
+Bu:
+ extraction:
+ measurement: background
+ post_equilibration: pump_unknown
+ post_measurement: pump_ms
+A:
+ extraction: air_x1
[…]
