+++ pychron/options/views/spectrum_views.py
@@ -74,6 +74,7 @@ class SpectrumAppearance(AppearanceSubOptions):
                       label='Fonts', show_border=True)
 
         g = VGroup(self._get_bg_group(),
+                   self._get_layout_group(),
                    self._get_padding_group(),
                    self._get_grid_group())
         return self._make_view(VGroup(g, fgrp))
