+++ pychron/mv/locator.py
@@ -14,6 +14,7 @@
 # ===============================================================================
 
 # ============= enthought library imports =======================
+from skimage.transform import rescale
 from traits.api import Float
 # ============= standard library imports ========================
 
@@ -23,10 +24,10 @@ from collections import deque, OrderedDict
 from numpy import array, histogram, argmax, zeros, asarray, ones_like, \
     nonzero, arange, argsort, invert, median, mean, zeros_like, ones, count_nonzero, unique, isnan
 from scipy import ndimage
-from skimage.morphology import watershed
+from skimage.morphology import watershed, disk
 from skimage.draw import polygon, circle, circle_perimeter, circle_perimeter_aa
 from skimage.exposure import rescale_intensity
-from skimage.filters import gaussian
+from skimage.filters import gaussian, rank
 from skimage import feature
 # ============= local library imports  ==========================
 from pychron.core.geometry.convex_hull import convex_hull
@@ -101,6 +102,9 @@ class Locator(Loggable):
         if self.step_signal:
             self.step_signal.wait()
             self.step_signal.clear()
+
+    def rescale(self, src, v):
+        return rescale(src, v, preserve_range=True)
 
     def crop(self, src, cw, ch, ox=0, oy=0, verbose=True):
 
@@ -181,6 +185,10 @@ class Locator(Loggable):
         if targets:
             self.info('found {} potential targets'.format(len(targets)))
             src = image.source_frame
+
+            self._draw_center_indicator(src, size=max(10, int(src.shape[0] * 0.25)),
+                                        shape='crosshairs')
+
             # draw targets
             self._draw_targets(src, targets)
 
@@ -196,22 +204,40 @@ class Locator(Loggable):
         self.debug('total find time={:0.5f}'.format(time.time() - st))
         return dx, dy
 
-    def _find_targets_bs(self, image, frame, dim, shape, preprocess, **kw):
+    def _find_targets_bs(self, image, frame, dim, shape, preprocess, mask=False, inverted=False,
+                         search_depth=6, min_targets=3,
+                         threshold_limiting=True,
+                         filter_targets=True,
+                         search_start=False, **kw):
         if preprocess:
             if not isinstance(preprocess, dict):
                 preprocess = {}
             src = self._preprocess(frame, **preprocess)
         else:
-            src = grayspace(frame)
+            src = grayspace(frame) * 255
 
-        image.set_frame(colorspace(frame / self.pixel_depth * 255))
+        if mask:
+            self._mask(src, mask)
 
+        if inverted:
+            src = invert(src)
+
+        # image.set_frame(colorspace(frame / self.pixel_depth * 255))
+        if image:
+            image.set_frame(colorspace(src))
         seg = RegionSegmenter()
         fa = self._get_filter_target_area(shape, dim)
 
-        sm = src.mean()
+        if search_start:
+            sm = search_start
+        else:
+            sm = src.mean()
+
         low = sm / 2
-        high = 255 - low
+        if threshold_limiting:
+            high = 255 - low
+        else:
+            high = 255
         # step = 5
 
         self.debug('src mean={}, {}, {}'.format(sm, low, high))
@@ -220,11 +246,14 @@ class Locator(Loggable):
 
             nsrc = seg.segment(src, t)
             per = _binary_percent(nsrc)
-            if per > 0.85 or per < 0.25:
+            if threshold_limiting and (per > 0.85 or per < 0.25):
                 return []
 
             ts = self._find_polygon_targets(nsrc)
-            ts = self._filter_targets(image, frame, dim, ts, fa, use_segmentation=False)
+            # self.debug('t={} polygons={}'.format(t, len(ts) if ts else 0))
+            if filter_targets:
[…]
