+++ pychron/lasers/tasks/laser_preferences.py
@@ -78,6 +78,7 @@ class FusionsLaserPreferences(LaserPreferences):
     aux_crosshairs_radius = Range(0.0, 10.0, 1.0)
     aux_crosshairs_offsetx = Float(0)
     aux_crosshairs_offsety = Float(0)
+    aux_crosshairs_offset_color = Color('red')
     aux_crosshairs_color = Color('red')
     aux_crosshairs_line_width = Float(1.0)
 
