+++ pychron/pipeline/csv_dataset_factory.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 # ===============================================================================
 import os
+from random import random
 
 from numpy import array
 from pyface.confirmation_dialog import confirm
@@ -25,6 +26,7 @@ from traitsui.api import UItem, TableEditor, HGroup, Item, VGroup, ListStrEditor
 from traitsui.extras.checkbox_column import CheckboxColumn
 from traitsui.menu import Action, Menu as MenuManager
 from traitsui.table_column import ObjectColumn
+from uncertainties import ufloat
 
 from pychron.core.csv.csv_parser import CSVColumnParser
 from pychron.core.fuzzyfinder import fuzzyfinder
@@ -39,10 +41,13 @@ from pychron.core.ui.table_editor import myTableEditor
 from pychron.envisage.icon_button_editor import icon_button_editor
 from pychron.paths import paths
 from pychron.processing.analyses.file_analysis import FileAnalysis
+from pychron.processing.isotope import Isotope
 from pychron.pychron_constants import PLUSMINUS_ONE_SIGMA
 
 HEADER = 'status', 'runid', 'age', 'age_err', 'group', 'aliquot', 'sample', 'label_name'
 SPECTRUM_HEADER = HEADER + ('k39', 'k39_err', 'rad40', 'rad40_err')
+ISOCHRON_HEADER = ('status', 'runid', 'ar40', 'ar40_err', 'ar39', 'ar39_err',
+                   'ar36', 'ar36_err', 'group', 'aliquot', 'sample', 'label_name')
 
 
 def make_line(vs, delimiter=','):
@@ -53,7 +58,6 @@ class CSVRecord(HasTraits):
     runid = Str('')
     age = CFloat
     age_err = CFloat
-
 
     group = CInt
     aliquot = CInt
@@ -80,6 +84,20 @@ class CSVSpectrumRecord(CSVRecord):
     k39_err = CFloat
     rad40 = CFloat
     rad40_err = CFloat
+
+
+class CSVIsochronRecord(CSVRecord):
+    header = ISOCHRON_HEADER
+
+    ar40 = CFloat
+    ar40_err = CFloat
+    ar39 = CFloat
+    ar39_err = CFloat
+    ar36 = CFloat
+    ar36_err = CFloat
+
+    def valid(self):
+        return self.runid and self.ar40 and self.a40_err and self.ar39 and self.ar39_err and self.ar36 and self.ar36_err
 
 
 class CSVRecordGroup(HasTraits):
@@ -248,7 +266,7 @@ e.g.
     #         gi.calculate()
 
     def _add_record_button_fired(self):
-        self.records.append(self._record_klass)
+        self.records.append(self._record_klass())
         self._make_groups()
 
     def _name_filter_changed(self, new):
@@ -399,6 +417,7 @@ e.g.
                             icon_button_editor('save_as_button', 'save_as', tooltip='Save As'),
                             icon_button_editor('clear_button', 'clear', tooltip='Clear current data'),
                             icon_button_editor('add_record_button', 'add'),
+                            icon_button_editor('test_button', 'test'),
                             icon_button_editor('open_help_button', 'help', tooltip='Show CSV formatting instructions'),
                             UItem('open_via_finder_button', tooltip='Open a csv file on your computer'),
                             # UItem('calculate_button')
@@ -435,6 +454,9 @@ e.g.
         return v
 
     def _test_button_fired(self):
+        self._load_test_data()
+
[…]
