+++ pychron/pipeline/engine.py
@@ -55,7 +55,7 @@ from pychron.pipeline.pipeline_defaults import ISOEVO, BLANKS, ICFACTOR, IDEO, S
     MASSSPEC_REDUCED, DEFINE_EQUILIBRATION, CA_CORRECTION_FACTORS, K_CORRECTION_FACTORS, \
     FLUX_VISUALIZATION, CSV_RAW_DATA_EXPORT, COMPOSITE, SIMPLE_ANALYSIS_TABLE, MASS_SPEC_FLUX, PYSCRIPT, RATIO_SERIES, \
     CSV_SPEC, ARAR_IDEO, ARAR_SPEC, ARAR_INVERSE_ISOCHRON, ARAR_SIMPLE_ANALYSIS_TABLE, RUNID_EDIT, CORRELATION_IDEO, \
-    COSMOGENIC, FLUX_EXPORT, RECENT_RUNS
+    COSMOGENIC, FLUX_EXPORT, RECENT_RUNS, CSV_INVERSE_ISOCHRON
 from pychron.pipeline.plot.editors.figure_editor import FigureEditor
 from pychron.pipeline.plot.editors.ideogram_editor import IdeogramEditor
 from pychron.pipeline.plot.editors.spectrum_editor import SpectrumEditor
@@ -877,6 +877,7 @@ class PipelineEngine(Loggable):
                      ('Series', SERIES),
                      ('Ratio Series', RATIO_SERIES),
                      ('InverseIsochron', INVERSE_ISOCHRON),
+                     ('CSV InverseIsochron', CSV_INVERSE_ISOCHRON),
                      ('XY Scatter', XY_SCATTER),
                      ('Regression', REGRESSION_SERIES),
                      ('Flux Visualization', FLUX_VISUALIZATION),
