+++ pychron/pipeline/plot/plotter/isochron.py
@@ -566,19 +566,20 @@ class InverseIsochron(Isochron):
 
         if self.options.display_inset and self.options.inset_link_status:
             plot = self.graph.plots[self.group_id]
-            for o in plot.overlays:
-                if isinstance(o, InverseIsochronLineInset):
-                    o.index.set_data(rxs)
-                    o.value.set_data(rys)
+            if plot:
+                for o in plot.overlays:
+                    if isinstance(o, InverseIsochronLineInset):
+                        o.index.set_data(rxs)
+                        o.value.set_data(rys)
 
     def update_graph_metadata(self, obj, name, old, new):
-        print('asdfsdfasdfasd', obj, name, new)
+        # print('asdfsdfasdfasd', obj, name, new)
 
         if obj:
             self._filter_metadata_changes(obj, self.sorted_analyses, self._rebuild_iso)
 
     def update_index_mapper(self, obj, name, old, new):
-        print('asdf', obj, name, new)
+        # print('asdf', obj, name, new)
         if new:
             self.update_graph_metadata(None, name, old, new)
 
