+++ pychron/graph/tools/analysis_inspector.py
@@ -19,6 +19,7 @@ from chaco.array_data_source import ArrayDataSource
 from traits.api import List, Callable
 from traitsui.menu import Action, Menu as MenuManager
 
+from pychron.core.helpers.formatting import floatfmt
 from pychron.graph.tools.point_inspector import PointInspector
 # from pychron.pipeline.plot.inspector_item import AnalysisInspectorItem
 from pychron.pychron_constants import PLUSMINUS
@@ -31,6 +32,7 @@ class AnalysisPointInspector(PointInspector):
     _selected_indices = List
     index_tag = None
     single_point = False
+    include_x = False
 
     def contextual_menu_contents(self):
         """
@@ -111,8 +113,12 @@ class AnalysisPointInspector(PointInspector):
                     for tag in ('<sub>','</sub>', '<sup>', '</sup>'):
                         name = name.replace(tag, '')
 
+                    xname = component.container.x_axis.title
+                    for tag in ('<sub>', '</sub>', '<sup>', '</sup>'):
+                        xname = xname.replace(tag, '')
+
                     y = ys[ind]
-                    # x = xs[ind]
+                    x = xs[ind]
 
                     if hasattr(component, 'yerror'):
                         try:
@@ -138,6 +144,26 @@ class AnalysisPointInspector(PointInspector):
                     info = [u'Analysis= {} UUID({})'.format(rid, analysis.display_uuid),
                             u'Tag= {}'.format(analysis.tag),
                             u'{}= {}'.format(name, y)]
+
+                    if self.include_x:
+
+                        if hasattr(component, 'xerror'):
+                            try:
+                                xerror = component.xerror
+                                if isinstance(xerror, ArrayDataSource):
+                                    xerror = xerror.get_data()
+
+                                xe = xerror[ind]
+                                pe = self.percent_error(x, xe)
+
+                                x = u'{} {}{} {}'.format(floatfmt(x), PLUSMINUS, xe, pe)
+                            except IndexError as e:
+                                print('asdf', e)
+                        else:
+                            x = floatfmt(x)
+
+                        info.append(u'{}= {}'.format(xname, x))
+
                     if analysis.comment:
                         info.insert(1, u'Comment= {}'.format(analysis.comment[:20]))
 
