+++ pychron/git_archive/author_view.py
@@ -41,6 +41,7 @@ class GitCommitAuthorView(HasTraits):
         v = okcancel_view(VGroup(Item('author', editor=ComboboxEditor(name='authornames')),
                                  Item('email')),
                           Item('remember_choice', label='Remember this choice?'),
+                          width=400,
                           title='Git Commit Author')
         return v
 
