### 2021-08-12 1db18b9
fixed xml scene loader
