+++ pychron/hardware/lakeshore/base_controller.py
@@ -17,6 +17,7 @@
 from traits.api import Enum, Float, Property, List, Int
 from traitsui.api import Item, UItem, HGroup, VGroup, Spring
 
+from pychron.core.pychron_traits import BorderVGroup
 from pychron.core.ui.lcd_editor import LCDEditor
 from pychron.graph.plot_record import PlotRecord
 from pychron.graph.stream_graph import StreamStackedGraph
@@ -170,10 +171,10 @@ class BaseLakeShoreController(CoreDevice):
             time.sleep(2)
             sp = self.read_setpoint(output, verbose=True)
             self.debug('setpoint set to={} target={}'.format(sp, v))
-            if sp==v:
+            if sp == v:
                 break
             time.sleep(1)
-            
+
         else:
             self.warning_dialog('Failed setting setpoint to {}. Got={}'.format(v, sp))
 
@@ -219,15 +220,16 @@ class BaseLakeShoreController(CoreDevice):
         return r
 
     def get_control_group(self):
-        grp = VGroup(Spring(height=10, springy=False),
-                     HGroup(Item('input_a', style='readonly', editor=LCDEditor(width=120, ndigits=6, height=30)),
-                            Item('setpoint1'),
-                            UItem('setpoint1_readback', editor=LCDEditor(width=120, height=30),
-                                  style='readonly'), Spring(width=10, springy=False)),
-                     HGroup(Item('input_b', style='readonly', editor=LCDEditor(width=120, ndigits=6, height=30)),
-                            Item('setpoint2'),
-                            UItem('setpoint2_readback', editor=LCDEditor(width=120, height=30),
-                                  style='readonly'), Spring(width=10, springy=False)))
+        grp = BorderVGroup(Spring(height=10, springy=False),
+                           HGroup(Item('input_a', style='readonly', editor=LCDEditor(width=120, ndigits=6, height=30)),
+                                  Item('setpoint1'),
+                                  UItem('setpoint1_readback', editor=LCDEditor(width=120, height=30),
+                                        style='readonly'), Spring(width=10, springy=False)),
+                           HGroup(Item('input_b', style='readonly', editor=LCDEditor(width=120, ndigits=6, height=30)),
+                                  Item('setpoint2'),
+                                  UItem('setpoint2_readback', editor=LCDEditor(width=120, height=30),
+                                        style='readonly'), Spring(width=10, springy=False)),
+                           title=self.name)
         return grp
 
     def graph_builder(self, g, **kw):
