### 2021-08-19 49372ff
added equilibration conditional
