+++ app_utils/install.py
@@ -531,7 +531,7 @@ def welcome():
 
 Developer: Jake Ross (NMT)
 Date: 10-02-2016
-Updated: 7-10-2019
+Updated: 6-11-2021
 ---*---*---*---*---*---*---*---*---*---*---*---*
 Welcome to the Pychron Installer.
 
@@ -591,6 +591,14 @@ def which(program):
             if is_exe(exe_file):
                 return exe_file
 
+def install_conda_only():
+    info_header('Install Conda Env only')
+    cfg = {}
+    vv = input('Install conda env only [n]')
+    if vv.lower() in ('y', 'yes'):
+        build_requirements(cfg)
+        install_conda(cfg)
+    return cfg
 
 def install_setupfiles_only():
     info_header('Install Setupfiles only')
@@ -682,15 +690,20 @@ def ask_config():
 def build_requirements(cfg):
     pip_reqs = ['uncertainties',
                 'peakutils',
-                'qimage2ndarray']
-    pip_git_reqs = ['git+https://github.com/enthought/chaco.git#egg=chaco',
-                    'git+https://github.com/enthought/enable.git#egg=enable']
+                'qimage2ndarray',
+                ]
 
-    conda_reqs = ['numpy', 'statsmodels', 'scikit-learn', 'PyYAML', 'yaml', 'traits=5', 'traitsui=6', 'pyface=6',
+    pip_git_reqs = []
+    # pip_git_reqs = ['git+https://github.com/enthought/chaco.git#egg=chaco',
+    #                 'git+https://github.com/enthought/enable.git#egg=enable']
+
+    conda_reqs = ['numpy', 'statsmodels', 'scikit-learn', 'PyYAML', 'yaml',
+                  # 'traits=5', 'traitsui=6', 'pyface=6',
+                  'traits', 'traitsui', 'pyface',
                   'envisage', 'sqlalchemy', 'Reportlab', 'lxml', 'xlrd', 'xlwt', 'xlsxwriter', 'requests', 'keyring',
                   'pillow', 'gitpython', 'cython', 'pytables', 'pyproj', 'pymysql', 'certifi', 'jinja2', 'swig=3',
                   cfg['qt_bindings']]
-
+    conda_other_channels = [['-c', 'dbanas', 'chaco']]
     if IS_MAC:
         conda_reqs.append('python.app')
 
@@ -700,8 +713,13 @@ def build_requirements(cfg):
     cfg['pip_requirements'] = pip_reqs
     cfg['pip_git_requirements'] = pip_git_reqs
     cfg['conda_requirements'] = conda_reqs
-
-
+    cfg['conda_other_channels'] = conda_other_channels
+    print('=========Conda Reqs==============')
+    print(' '.join(conda_reqs))
+    print('=================================')
+    print('=========PIP Reqs================')
+    print(' '.join(pip_reqs))
+    print('=================================')
 # config['pip_requirements'] = 'uncertainties peakutils qimage2ndarray'
 # config['pip_git_requirements'] =
 #
@@ -759,7 +777,7 @@ def install_conda(cfg):
 
     # create env
     env_name = cfg['conda_env_name']
-    subprocess.call(['conda', 'create', '-n', env_name, '--yes', 'python=3.7'])
+    subprocess.call(['conda', 'create', '-n', env_name, '--yes', 'python=3.8'])
 
     # install deps
     subprocess.call(['conda', 'install', '--yes',
@@ -773,6 +791,9 @@ def install_conda(cfg):
         subprocess.call([pip_path, 'install'] + cfg['pip_requirements'])
         for r in cfg['pip_git_requirements']:
             subprocess.call([pip_path, 'install', '-e', r])
+
+        for c in cfg['conda_other_channels']:
+            subprocess.call(['conda', 'install']+c)
     else:
[…]
