+++ pychron/experiment/tasks/experiment_task.py
@@ -41,7 +41,7 @@ from pychron.globals import globalv
 from pychron.lasers.laser_managers.ilaser_manager import ILaserManager
 from pychron.paths import paths
 from pychron.pipeline.plot.editors.figure_editor import FigureEditor
-from pychron.pychron_constants import SPECTROMETER_PROTOCOL, DVC_PROTOCOL, COCKTAIL, AIR, BLANK, FUSIONS_UV, STATES, \
+from pychron.pychron_constants import SPECTROMETER_PROTOCOL, DVC_PROTOCOL, COCKTAIL, AIR, BLANK, FUSIONS_UV, \
     INVALID, EXTRACTION, MEASUREMENT, CANCELED, TRUNCATED, END_AFTER, FAILED, SUCCESS
 
 
