+++ pychron/dvc/dvc.py
@@ -1759,8 +1759,9 @@ class DVC(Loggable):
             try:
                 a = DVCAnalysis(uuid, rid, expid)
             except AnalysisNotAnvailableError:
-                self.warning_dialog('Analysis {} not in repository {}. '
-                                    'You many need to pull changes'.format(rid, expid))
+                self.warning_dialog('Analysis {} not in local repository {}. '
+                                    'You may need to pull changes. If local repository is up to date you may '
+                                    'need to push changes from the data collection computer'.format(rid, expid))
                 return
 
             a.group_id = record.group_id
