+++ pychron/hardware/actuators/__init__.py
@@ -92,4 +92,5 @@ PACKAGES = dict(AgilentGPActuator='{}.agilent.agilent_gp_actuator'.format(base),
                 T4Actuator='{}.t4_actuator'.format(abase),
                 U3Actuator='{}.u3_actuator'.format(abase),
                 ProXRActuator='{}.proxr_actuator'.format(abase),
-                PLC2000GPActuator='{}.plc2000_gp_actuator'.format(abase))
+                PLC2000GPActuator='{}.plc2000_gp_actuator'.format(abase),
+                MCCGPActuator='{}.mcc_gp_actuator'.format(abase))
