+++ pychron/pipeline/plot/models/composite_model.py
@@ -33,7 +33,7 @@ class CompositeModel(FigureModel):
     # def _refresh_panels_hook(self):
     #     self.on_trait_change(self._handle_recalculate, 'panels:figures:recalculate_event')
 
-    def _make_panels(self):
+    def _make_panel_groups(self):
 
         # spo = SpectrumOptionsManager().selected_options
         # ipo = InverseIsochronOptionsManager().selected_options
@@ -42,9 +42,12 @@ class CompositeModel(FigureModel):
                             plot_options=self.plot_options.spectrum_options),
               InverseIsochronPanel(analyses=self.analyses,
                                    plot_options=self.plot_options.isochron_options)]
-        for gi in gs:
-            gi.make_figures()
-
         return gs
 
+    def _make_panels(self):
+        gs = super(CompositeModel, self)._make_panels()
+        for i, gi in enumerate(gs):
+            if gi.plot_options.auto_generate_title:
+                gi.title = gi.plot_options.generate_title(gi.analyses, i)
+        return gs
 # ============= EOF =============================================
