+++ pychron/pipeline/plot/plotter/arar_figure.py
@@ -636,14 +636,14 @@ class BaseArArFigure(SelectionFigure):
             swe = floatfmt(we, sig_figs)
 
         if self.options.index_attr in ('uF', 'Ar40/Ar36'):
-            me = u'{} {}{}'.format(sx, PLUSMINUS, swe)
+            me = u'{} {} {}'.format(sx, PLUSMINUS, swe)
         else:
             age_units = self._get_age_units()
             pe = ''
             if percent_error:
                 pe = '({})'.format(format_percent_error(x, we, include_percent_sign=True))
 
-            me = u'{} {}{}{} {}'.format(sx, PLUSMINUS, swe, pe, age_units)
+            me = u'{} {} {}{} {}'.format(sx, PLUSMINUS, swe, pe, age_units)
 
         return u'{} {} {}'.format(me, mswd, n)
 
