+++ pychron/pipeline/plot/models/regression_series_model.py
@@ -32,18 +32,8 @@ from six.moves import zip
 class RegressionSeriesModel(FigureModel):
     _panel_klass = RegressionSeriesPanel
 
-    def _make_panels(self):
+    def _make_panel_groups(self):
         gs = [self._panel_klass(analyses=[a], plot_options=self.plot_options) for a in self.analyses]
-        for gi in gs:
-            gi.make_figures()
-
-        if self.plot_options.auto_generate_title:
-            for i, gi in enumerate(gs):
-                gi.title = self.plot_options.generate_title(gi.analyses, i)
-
-        elif self.titles:
-            for ti, gi in zip(self.titles, gs):
-                gi.title = ti
 
         return gs
 
