+++ pychron/options/views/composite_views.py
@@ -13,9 +13,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
+from pychron.options.options import SubOptions
 from pychron.options.views.isochron_views import InverseIsochronMainOptions, InverseIsochronAppearance
 from pychron.options.views.spectrum_views import SpectrumMainOptions, SpectrumAppearance, DisplaySubOptions, \
     CalculationSubOptions
+
+from traitsui.api import VGroup, Item
+
+
+class LayoutSubOptions(SubOptions):
+    def traits_view(self):
+        g = VGroup(Item('orientation_layout'))
+        return self._make_view(g)
+
 
 VIEWS = {'main': SpectrumMainOptions,
          'spectrum': SpectrumMainOptions,
@@ -23,5 +33,6 @@ VIEWS = {'main': SpectrumMainOptions,
          'display(spec.)': DisplaySubOptions,
          'calculations(spec.)': CalculationSubOptions,
          'appearance(spec.)': SpectrumAppearance,
-         'appearance(iso.)': InverseIsochronAppearance}
+         'appearance(iso.)': InverseIsochronAppearance,
+         'layout': LayoutSubOptions}
 # ============= EOF =============================================
