### 2021-08-31 fd40d0f
updated sample loader
