### 2021-09-03 f15418b
fixed series inspection
