+++ pychron/extraction_line/switch_manager.py
@@ -1138,8 +1138,13 @@ class SwitchManager(Manager):
 
         qs = True
         vqs = v_elem.find('query_state')
+        if vqs is None:
+            vqs = v_elem.get('query_state')
+        else:
+            vqs = vqs.text
+
         if vqs is not None:
-            qs = to_bool(vqs.text.strip())
+            qs = to_bool(vqs.strip())
 
         parent = v_elem.find('parent')
 
