+++ pychron/canvas/canvas2D/scene/primitives/primitives.py
@@ -88,13 +88,19 @@ class Bordered(Primitive):
     use_border = True
     border_width = 2
 
-    def _get_border_color(self):
+    def _render_augmented_border(self, gc):
+        pass
+
+    def _get_border_color(self, scale=0.5):
         c = self.default_color
         if self.state:
             c = self.active_color
 
         c = self._convert_color(c)
-        c = [ci / 2. for ci in c]
+
+        if not isinstance(scale, (list, tuple)):
+            scale = (scale,) * len(c)
+        c = [ci * scale[i] for i, ci in enumerate(c)]
         if len(c) == 4:
             c[3] = 1
 
@@ -670,7 +676,7 @@ class BorderLine(Line, Bordered):
     def _render(self, gc):
         gc.save_state()
         with gc:
-            gc.set_line_width(self.width + self.border_width)
+            gc.set_line_width(self.width+self.border_width)
 
             gc.set_stroke_color(self._get_border_color())
 
