+++ pychron/canvas/canvas2D/scene/primitives/pumps.py
@@ -22,13 +22,67 @@ from pychron.canvas.canvas2D.scene.primitives.primitives import Animation
 from pychron.canvas.canvas2D.scene.primitives.rounded import RoundedRectangle, rounded_rect
 
 
-class IonPump(RoundedRectangle):
-    pass
-
-
-class Turbo(RoundedRectangle):
+class Pump(RoundedRectangle):
     use_symbol = False
 
+    def _render_base_symbol(self, gc, br):
+        gc.arc(0, 0, br, 0, 360)
+
+        for a, b in ((60, -30), (120, -150)):
+            a = math.radians(a)
+            b = math.radians(b)
+            xx = br * math.cos(a)
+            yy = br * math.sin(a)
+            gc.move_to(xx, yy)
+
+            x2 = br * math.cos(b)
+            y2 = br * math.sin(b)
+            gc.line_to(x2, y2)
+            gc.stroke_path()
+
+
+class IonPump(Pump):
+
+    def _render(self, gc):
+        super(IonPump, self)._render(gc)
+        if self.use_symbol:
+            self._render_symbol(gc)
+
+    def _render_symbol(self, gc):
+        corner_radius = self.corner_radius
+        width, height = self.get_wh()
+        x, y = self.get_xy()
+        w = width * 0.75
+        h = w
+        xx = x + (width - w) / 2
+        yy = y - h
+        with gc:
+            rounded_rect(gc, xx, yy, w, h, corner_radius)
+            self._render_border(gc, xx, yy, w, h, use_border_gaps=False)
+
+        with gc:
+            gc.set_line_width(2)
+            gc.set_stroke_color((0.1, 0.1, 0.1))
+
+            gc.translate_ctm(xx + w / 2, yy + h / 2)
+            w2 = w / 2 - 15
+            self._render_base_symbol(gc, w2 + 8)
+            gc.rotate_ctm(math.radians(30))
+
+            for i in range(3):
+                with gc:
+                    gc.rotate_ctm(math.radians(120 * i))
+                    # with gc:
+                    gc.move_to(0, 0)
+                    gc.line_to(w2-2, 0)
+                    gc.stroke_path()
+                    gc.move_to(w2 - 6, 6)
+                    gc.line_to(w2-2, 0)
+                    gc.line_to(w2 - 6, -6)
+                    gc.stroke_path()
+
+
+class Turbo(Pump):
     def _render(self, gc):
         corner_radius = self.corner_radius
         with gc:
@@ -46,7 +100,8 @@ class Turbo(RoundedRectangle):
                 yy = y - h
                 rounded_rect(gc, xx, yy, w, h, corner_radius)
                 self._render_border(gc, xx, yy, w, h, use_border_gaps=False)
-                self._draw_impeller(gc, xx, yy, w, h)
+                self._render_symbol(gc, xx, yy, w, h)
+                # self._draw_impeller(gc, xx, yy, w, h)
 
             gc.set_fill_color(self._convert_color(self.name_color))
             if self.display_name:
@@ -59,38 +114,54 @@ class Turbo(RoundedRectangle):
         # if self.animate:
         #     self._draw_impeller(gc)
 
-    def _draw_impeller(self, gc, x, y, w, h):
+    def _render_symbol(self, gc, x, y, w, h):
         with gc:
-            # x, y = self.get_xy(clear_layout_needed=False)
-            # w, h = self.get_wh()
-            # c = self._convert_color(self.default_color)
-            c = self._get_border_color()
-            gc.set_fill_color(c)
+            gc.set_line_width(2)
+            gc.set_stroke_color((0.1, 0.1, 0.1))
+            gc.translate_ctm(x + w / 2, y + h / 2)
 
-            cx = x + w / 2.
[…]
