+++ pychron/canvas/canvas2D/scene/primitives/connections.py
@@ -63,62 +63,62 @@ class Connection(ConnectionMixin, BorderLine):
 
 class RConnection(Connection):
     tag = 'rconnection'
-    # border_width = 2
-    width = 5
+    border_width = 6
+    width = 3
 
-    def _render_augmented_border(self, gc):
-        bc = self._get_border_color()
-        x, y = self.start_point.get_xy()
-        x1, y1 = self.end_point.get_xy()
-
-        i = 0
-        step = 10
-        if y1 < y or x1 < x:
-            step = -10
-
-        while 1:
-            gc.set_line_width(0)
-            gc.set_fill_color(bc)
-            # step = 10
-            with gc:
-                if x1 == x:
-                    cx = self.width+1
-                    cy = 0
-                    tx = x
-                    ty = y + (step * i)
-                else:
-                    cx = 0
-                    cy = self.width+1
-                    tx = x + (step * i)
-                    ty = y
-
-                gc.translate_ctm(tx, ty)
-                gc.arc(-cx, -cy, 4, 0, 360)
-                gc.arc(cx, cy, 4, 0, 360)
-                gc.draw_path()
-
-            i += 1
-            if x1 == x and (ty > max(y1, y) or ty < min(y1, y)):
-                break
-            elif tx > max(x1, x) or tx < min(x1, x):
-                break
+    # def _render_augmented_border(self, gc):
+    #     bc = self._get_border_color()
+    #     x, y = self.start_point.get_xy()
+    #     x1, y1 = self.end_point.get_xy()
+    #
+    #     i = 0
+    #     step = 10
+    #     if y1 < y or x1 < x:
+    #         step = -10
+    #
+    #     while 1:
+    #         gc.set_line_width(0)
+    #         gc.set_fill_color(bc)
+    #         # step = 10
+    #         with gc:
+    #             if x1 == x:
+    #                 cx = self.width+1
+    #                 cy = 0
+    #                 tx = x
+    #                 ty = y + (step * i)
+    #             else:
+    #                 cx = 0
+    #                 cy = self.width+1
+    #                 tx = x + (step * i)
+    #                 ty = y
+    #
+    #             gc.translate_ctm(tx, ty)
+    #             gc.arc(-cx, -cy, 4, 0, 360)
+    #             gc.arc(cx, cy, 4, 0, 360)
+    #             gc.draw_path()
+    #
+    #         i += 1
+    #         if x1 == x and (ty > max(y1, y) or ty < min(y1, y)):
+    #             break
+    #         elif tx > max(x1, x) or tx < min(x1, x):
+    #             break
 
     def _render(self, gc):
-        bc = self._get_border_color()
-        with gc:
-            # w, h = self.get_wh()
-            gc.set_line_width(self.width + 10)
+        # bc = self._get_border_color()
+        # with gc:
+        #     # w, h = self.get_wh()
+        #     gc.set_line_width(self.width + 10)
+        #
+        #     gc.set_stroke_color(bc)
+        #
+        #     x, y = self.start_point.get_xy()
+        #     x1, y1 = self.end_point.get_xy()
+        #     # draw border
+        #     gc.move_to(x, y)
+        #     gc.line_to(x1, y1)
+        #     gc.draw_path()
 
-            gc.set_stroke_color(bc)
-
-            x, y = self.start_point.get_xy()
-            x1, y1 = self.end_point.get_xy()
-            # draw border
-            gc.move_to(x, y)
-            gc.line_to(x1, y1)
-            gc.draw_path()
-
-        super(BorderLine, self)._render(gc)
+        super(RConnection, self)._render(gc)
 
         # draw border vertical augmentation
         self._render_augmented_border(gc)
