+++ pychron/canvas/canvas2D/scene/primitives/rounded.py
@@ -46,6 +46,7 @@ def rounded_rect(gc, x, y, width, height, corner_radius):
         gc.arc_to(x, y, x + width + corner_radius, y, corner_radius)
         gc.draw_path()
 
+
 # class Stop:
 #     def __init__(self, idx, r,g,b,a):
 #         pass
@@ -111,6 +112,7 @@ class RoundedRectangle(Rectangle, Connectable, Bordered):
                 with gc:
                     for t, c in self.connections:
                         with gc:
+                            w2 = self.border_width
                             gc.set_line_width(self.border_width + 1)
                             if isinstance(c, Elbow):
                                 p1, p2 = c.start_point, c.end_point
@@ -136,7 +138,7 @@ class RoundedRectangle(Rectangle, Connectable, Bordered):
                                         gc.line_to(xx, p2y + 5)
 
                             elif isinstance(c, BorderLine):
-                                w2 = c.width/2
+                                w2 = c.width / 2
                                 p1, p2 = c.start_point, c.end_point
                                 p2x, p2y = p2.get_xy()
                                 if p1.x == p2.x:
@@ -157,11 +159,19 @@ class RoundedRectangle(Rectangle, Connectable, Bordered):
                                     gc.line_to(xx, p2y + w2)
 
                             elif isinstance(c, Tee):
+
                                 if t == 'mid':
-                                    yy = y if c.left.y < self.y else y + height
-                                    mx = c.get_midx()
-                                    gc.move_to(mx - 5, yy)
-                                    gc.line_to(mx + 5, yy)
+                                    # tee is vertical
+                                    if abs(c.left.x - c.right.x) <= 1:
+                                        xx = x if c.left.x < self.x else x + width
+                                        yy = y + height / 2
+                                        gc.move_to(xx, yy - w2)
+                                        gc.line_to(xx, yy + w2)
+                                    else:
+                                        mx = c.get_midx()
+                                        yy = y if c.left.y < self.y else y + height
+                                        gc.move_to(mx - w2, yy)
+                                        gc.line_to(mx + w2, yy)
                                 else:
                                     gc.set_line_width(self.border_width + 2)
                                     # gc.set_stroke_color((1,0,0))
@@ -171,13 +181,13 @@ class RoundedRectangle(Rectangle, Connectable, Bordered):
                                     else:
                                         xx, yy = c.right.get_xy()
 
-                                    gc.move_to(xx, yy - 5)
-                                    gc.line_to(xx, yy + 5)
+                                    gc.move_to(xx, yy - w2)
+                                    gc.line_to(xx, yy + w2)
                             elif isinstance(c, Fork):
                                 yy = y if c.left.y < self.y else y + height
                                 mx = c.get_midx()
-                                gc.move_to(mx - 5, yy)
-                                gc.line_to(mx + 5, yy)
+                                gc.move_to(mx - w2, yy)
+                                gc.line_to(mx + w2, yy)
 
                             gc.draw_path()
 
@@ -304,5 +314,9 @@ class Getter(RoundedRectangle):
 
 
 class Gauge(RoundedRectangle):
+    pass
+
+
+class ColdFinger(RoundedRectangle):
     pass
 # ============= EOF =============================================
