+++ pychron/pyscripts/measurement_pyscript.py
@@ -99,8 +99,8 @@ class MeasurementPyScript(ValvePyScript):
     # commands
     # ===============================================================================
     @command_register
-    def data_sink(self):
-        self._automated_run_call('py_data_sink')
+    def sink_data(self, n=100):
+        self._automated_run_call('py_sink_data', n=n)
 
     @command_register
     def measurement_delay(self, duration=None, message=None):
