+++ pychron/pyscripts/measurement_pyscript.py
@@ -98,11 +98,13 @@ class MeasurementPyScript(ValvePyScript):
     # ===============================================================================
     # commands
     # ===============================================================================
+    @verbose_skip
     @command_register
     def sink_data(self, n=100, calc_time=False):
         self._automated_run_call('py_sink_data', n=n,
                                  is_syntax_testing=self.testing_syntax or calc_time)
 
+    @verbose_skip
     @command_register
     def measurement_delay(self, duration=None, message=None):
         if duration:
