+++ pychron/pyscripts/measurement_pyscript.py
@@ -100,9 +100,17 @@ class MeasurementPyScript(ValvePyScript):
     # ===============================================================================
     @verbose_skip
     @command_register
-    def sink_data(self, n=100, calc_time=False):
-        self._automated_run_call('py_sink_data', n=n,
-                                 is_syntax_testing=self.testing_syntax or calc_time)
+    def sink_data(self, n=100, delay=1, calc_time=False):
+        """
+
+        @param n: number of measurements
+        @param period: delay between measurements
+        @param calc_time:
+        @return:
+        """
+        if calc_time:
+            return n*delay
+        self._automated_run_call('py_sink_data', n=n, delay=delay)
 
     @verbose_skip
     @command_register
