### 2021-09-30 #3386 @richardjgowers
I’ll try to find another reviewer, perhaps someone new is available now.

On Thu, Sep 30, 2021 at 19:47, Daniel S. Katz ***@***.***>
wrote:

> What's the next step? If we can't find another reviewer, we may have to
> mark it as withdrawn. But in any case, we shouldn't just leave it hanging.
>
