### 2021-10-01 #2919 @bbucior
@richardjgowers -- sorry, I'm in a similar boat as Andrew. The only lead that comes to mind is the team over at https://github.com/numat since they've written some other data collection software.
