### 2021-10-01 6906f3e
addressed item 2 of issue #2019
