+++ pychron/options/views/composite_views.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from pychron.options.options import SubOptions
+from pychron.options.options import SubOptions, TitleSubOptions
 from pychron.options.views.isochron_views import InverseIsochronMainOptions, InverseIsochronAppearance
 from pychron.options.views.spectrum_views import SpectrumMainOptions, SpectrumAppearance, DisplaySubOptions, \
     CalculationSubOptions
@@ -27,6 +27,9 @@ class LayoutSubOptions(SubOptions):
         return self._make_view(g)
 
 
+# class TitleSubOptions(SubOptions):
+#     pass
+
 VIEWS = {'main': SpectrumMainOptions,
          'spectrum': SpectrumMainOptions,
          'isochron': InverseIsochronMainOptions,
@@ -34,5 +37,6 @@ VIEWS = {'main': SpectrumMainOptions,
          'calculations(spec.)': CalculationSubOptions,
          'appearance(spec.)': SpectrumAppearance,
          'appearance(iso.)': InverseIsochronAppearance,
-         'layout': LayoutSubOptions}
+         'layout': LayoutSubOptions,
+         'title': TitleSubOptions}
 # ============= EOF =============================================
