+++ pychron/pipeline/plot/models/composite_model.py
@@ -24,20 +24,7 @@ from pychron.pipeline.plot.panels.spectrum_panel import SpectrumPanel
 
 class CompositeModel(FigureModel):
 
-    # @on_trait_change('panels:figures:recalculate_event')
-    # def _handle_recalculate(self):
-    #     print('recalads')
-    #     for p in self.panels:
-    #         p.make_graph()
-
-    # def _refresh_panels_hook(self):
-    #     self.on_trait_change(self._handle_recalculate, 'panels:figures:recalculate_event')
-
     def _make_panel_groups(self):
-
-        # spo = SpectrumOptionsManager().selected_options
-        # ipo = InverseIsochronOptionsManager().selected_options
-
         gs = [SpectrumPanel(analyses=self.analyses,
                             plot_options=self.plot_options.spectrum_options),
               InverseIsochronPanel(analyses=self.analyses,
@@ -46,8 +33,14 @@ class CompositeModel(FigureModel):
 
     def _make_panels(self):
         gs = super(CompositeModel, self)._make_panels()
-        for i, gi in enumerate(gs):
-            if gi.plot_options.auto_generate_title:
-                gi.title = gi.plot_options.generate_title(gi.analyses, i)
+        if self.plot_options.auto_generate_title:
+            for gi in gs:
+                gi.title = ''
+
+        else:
+            for i, gi in enumerate(gs):
+                if gi.plot_options.auto_generate_title:
+                    gi.title = gi.plot_options.generate_title(gi.analyses, i)
+
         return gs
 # ============= EOF =============================================
