+++ pychron/spectrometer/thermo/spectrometer/helix.py
@@ -36,10 +36,10 @@ class HelixSpectrometer(ThermoSpectrometer):
 
     def hardware_names(self):
         hn = super(HelixSpectrometer, self).hardware_names()
-        hn['flatapole'] = 'Flatapole'
+        hn['flatapole'] = 'DAC_1_0_(Flata-Pole)'
         hn['rotation_quad'] = 'RotationQuad'
-        hn['vertical_deflection_n'] = 'VerticalDeflection N Set'
-        hn['vertical_deflection_s'] = 'VerticalDeflection S Set'
+        hn['vertical_deflection_n'] = 'DAC_0_0_(Pole-N)'
+        hn['vertical_deflection_s'] = 'DAC_0_4_(Pole-S)'
         hn['horizontal_symmetry'] = 'Horizontal Symmetry Set'
         hn['extraction_focus'] = 'Extraction Focus Set'
         hn['extraction_symmetry'] = 'Extraction Symmetry Set'
