### 2021-10-04 caee3ea
fixed set_cryo device_name
