### 2021-10-05 67c3667
functioning data stream
