+++ pychron/stage/maps/laser_stage_map.py
@@ -235,217 +235,4 @@ class UVLaserStageMap(LaserStageMap):
 
         return pos
 
-        # ============= EOF =============================================
-        #     def interpolate_noncorrected(self):
-        #         self.info('iteratively fill in non corrected holes')
-        #         n = len(self.sample_holes)
-        #         for i in range(2):
-        #             self._interpolate_noncorrected()
-        #
-        #             g = len([h for h in self.sample_holes
-        #                      if h.has_correction()])
-        #
-        #             if g == n:
-        #                 break
-        #             self.info('iteration {}, total={}'.format(i + 1, g))
-        #         if g < n:
-        #             self.info('{} holes remain noncorrected'.format(n - g))
-        #         else:
-        #             self.info('all holes now corrected')
-        #
-        #     def get_interpolated_position(self, holenum, cpos=None, rotation=None):
-        #         self.cpos = cpos
-        #         self.rotation = rotation
-        #
-        #         h = self.get_hole(holenum)
-        #         if h is not None:
-        #             nxs = []
-        #             nys = []
-        #             iholes = []
-        #             n = 3
-        #             for sd in range(n):
-        #                 xi, yi, hi = self._calculated_interpolated_position(h, sd + 1)
-        #                 # do simple weighting by distance
-        #                 w = (n - sd)
-        #                 nxs += xi * w
-        #                 nys += yi * w
-        #                 iholes += hi
-        #
-        #             if nxs and nys:
-        #                 nx, ny = (sum(nxs) / max(1, len(nxs)),
-        #                           sum(nys) / max(1, len(nys)))
-        #
-        #                 # verify within tolerance
-        #                 tol = h.dimension * 0.85
-        #
-        #                 hx, hy = self.map_to_calibration(h.nominal_position)
-        #                 if abs(nx - hx) < tol and abs(ny - hy) < tol:
-        #                     h.interpolated = True
-        #                     h.corrected = True
-        #                     h.interpolation_holes = set(iholes)
-        #
-        #                     h.x_cor = nx
-        #                     h.y_cor = ny
-        #
-        #                     return nx, ny
-        # def _interpolate_noncorrected(self):
-        #         self.sample_holes.reverse()
-        #         for h in self.sample_holes:
-        #             self._calculated_interpolated_position(h)
-        #
-        #     def _calculated_interpolated_position(self, h, search_distance):
-        #         """
-        #             search distance is a scalar in hole units. it defines how many
-        #             holes away to
-        #         """
-        #
-        #         spacing = search_distance * abs(self.sample_holes[0].x - self.sample_holes[1].x)
-        #         #         debug_hole = '18'
-        #         nxs = []
-        #         nys = []
-        #         iholes = []
-        #
-        #         if  not h.has_correction():
-        #             # this hole does not have a correction value
-        #             found = []
-        #             # get the cardinal holes and corner holes
-        #             for rx, ry in [(0, 1),
-        #                            (-1, 0), (1, 0),
-        #                            (0, -1),
-        #
-        #                            (-1, 1), (1, 1),
-        #                            (-1, -1), (1, -1)]:
-        #
-        #                 x = h.x + rx * spacing
-        #                 y = h.y + ry * spacing
-        #
-        #                 ihole = self._get_hole_by_position(x, y)
-        #
-        #                 if ihole == h:
-        #                     ihole = None
-        #
-        #                 fo = None
-        #                 if ihole is not None:
-        #                     if ihole.has_correction():
-        #                         fo = ihole
[…]
