### 2021-10-06 8a97d8c
addressed wiscar issue with slope
