### 2021-10-06 73bc716
addressed wiscar issue with slope
