### 2021-10-08 cc68c16
jan hot fixes. fixed loading conditionals
