### 2021-10-08 822ee62
jan hot fixes. fixed equilibration threading issues
