### 2021-10-09 3735d69
jan hot fixes. fixed threading issues
