### 2021-10-09 b11fb2d
jan hot fixes. fixed threading issues
