+++ pychron/pychron_constants.py
@@ -334,6 +334,7 @@ SAMPLE_METADATA = (SAMPLE,
                    MATERIAL,
                    'grainsize',
                    PROJECT,
+                   'note',
                    'principal_investigator',
                    'latitude',
                    'longitude',
