+++ pychron/sparrow/sparrow_client.py
@@ -1,0 +1,194 @@
+# ===============================================================================
+# Copyright 2021 ross
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ===============================================================================
+import json
+from datetime import datetime
+from json import JSONDecodeError
+from pprint import pformat
+
+import requests
+from requests.exceptions import InvalidSchema, ConnectTimeout, ConnectionError
+from traits.api import Str, Int
+from apptools.preferences.preference_binding import bind_preference
+
+from pychron.core.pychron_traits import URLStr
+from pychron.experiment.utilities.identifier import is_step_heat
+from pychron.loggable import Loggable
+
+
+def sample_payload(sm):
+    return {'name': sm.get('sample', 'NoSample'),
+            'note': sm.get('note'),
+            'project': [{'name': sm.get('project', 'NoProject')}],
+            'material': sm.get('material', 'NoMaterial'),
+            'location': {'type': 'Point',
+                         'coordinates': [float(sm.get('longitude', 0)),
+                                         float(sm.get('latitude', 0))]}
+            }
+
+
+def instrument_payload(cm):
+    return {'name': cm.get('instrument', 'NoInstrument')}
+
+
+def datum_factory(p):
+    return {'type': {'parameter': p['kind'], 'unit': p.get('unit', '')},
+            'value': p['value'],
+            'error': p['error']}
+
+
+def analyses_payload(ans, preferred_kinds):
+    analyses = [{'is_bad': ai.get('tag', 'ok').lower() == 'invalid',
+                 'analysis_type': 'Step Heat' if is_step_heat(ai.get('record_id')) else 'Fusion',
+                 'analysis_name': ai.get('record_id'),
+                 'uuid': ai.get('uuid'),
+                 'datum': [{'type': {'parameter': 'age', 'unit': ai.get('age_units', 'Ma')},
+                            'value': ai.get('age', 0),
+                            'error': ai.get('age_err', 0)},
+                           {'type': {'parameter': 'kca', 'unit': ''},
+                            'value': ai.get('kca', 0),
+                            'error': ai.get('kca_err', 0)}
+                           ]
+                 }
+                for ai in ans]
+
+    preferreds = [{'analysis_name': 'preferred',
+                   'analysis_type': 'Preferred',
+                   'datum': [datum_factory(p) for p in preferred_kinds]}]
+
+    pref = next((p for p in preferred_kinds if p['attr'] == 'age'), {})
+    page = pref.get('value', 0)
+    perror = pref.get('error', 0)
+    punit = pref.get('unit', 'Ma')
+    preferred_age = {'analysis_name': 'PreferredAge',
+                     'anaylsis_type': 'Preferred',
+                     'is_interpreted': True,
+                     'datum': [{'type': {'parameter': 'age', 'unit': punit},
+                                'value': page,
+                                'error': perror}]}
+    analyses.extend(preferreds)
+    analyses.append(preferred_age)
+    return analyses
+
+
+class SparrowClient(Loggable):
+    """ Thin http client used to interact with Sparrow API"""
+    username = Str
+    password = Str
+    host = URLStr
+    token = Str
+
+    def __init__(self, bind=True, *args, **kw):
+        super().__init__(*args, **kw)
[…]
