+++ .github/workflows/format_code.yml
@@ -1,0 +1,21 @@
+name: Format code
+on:
+  pull_request:
+    branches: [dev/dr, dev/dc]
+  push:
+    branches: [dev/dr, dev/dc]
+jobs:
+  format:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          ref: ${{ github.head_ref }}
+      - uses: rickstaa/action-black@v1
+        with:
+          black_args: "--extend-exclude [sandbox|docs|dist|zobs|test|alembic_dvc] pychron/**/*.py "
+      - name: Commit changes
+        uses: stefanzweifel/git-auto-commit-action@v4.1.2
+        with:
+          commit_message: Formatting changes
+          branch: ${{ github.head_ref }}
