+++ pychron/applications/pysampleprep.py
@@ -19,13 +19,15 @@
 # from traitsui.api import View, Item
 from __future__ import absolute_import
 from pychron.applications.pychron_application import PychronApplication
+
 # ============= standard library imports ========================
 # ============= local library imports  ==========================
 
 
 class PySamplePrep(PychronApplication):
-    id = 'pychron.sampleprep.application'
-    name = 'pySamplePrep'
-    shortname = 'sampleprep'
+    id = "pychron.sampleprep.application"
+    name = "pySamplePrep"
+    shortname = "sampleprep"
+
 
 # ============= EOF =============================================
