### 2021-10-14 eb5521b
added install action
