+++ .github/workflows/install.yml
@@ -12,39 +12,39 @@ on:
 jobs:
   job-1:
     name: Pychron Install (${{ matrix.python-version }}, ${{ matrix.os }})
-      runs-on: "macos-latest"
-#      strategy:
-#        fail-fast: false
-#        matrix:
-#          os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-#          python-version: [ "3.9",]
-      steps:
-        - uses: conda-incubator/setup-miniconda@v2
-          with:
-            auto-update-conda: true
-            python-version: ${{ matrix.python-version }}
-        - uses: actions/checkout@v2
-  #      - name: Set up Python 3.9
-  #        uses: actions/setup-python@v2
-  #        with:
-  #          python-version: 3.9
-        - name: Install Conda dependencies
-          run: |
-            conda install numpy statsmodels scikit-learn PyYAML \
-            yaml traits traitsui pyface envisage sqlalchemy \
-            Reportlab lxml xlrd xlwt xlsxwriter requests keyring \
-            pyparsing pillow gitpython pytables pyproj pymysql \
-            certifi jinja2 swig=3 cython importlib_resources
-        - name: Install dependencies
-          run: |
-            python -m pip install --upgrade pip
-            pip install flake8 pytest uncertainties qdimage2ndarray peakutils chaco
-        - name: Lint with flake8
-          run: |
-            # stop the build if there are Python syntax errors or undefined names
-            flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-            # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-            flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-        - name: Test with pytest
-          run: |
-            pytest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
+        python-version: [ "3.9",]
+    steps:
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          auto-update-conda: true
+          python-version: ${{ matrix.python-version }}
+      - uses: actions/checkout@v2
+#      - name: Set up Python 3.9
+#        uses: actions/setup-python@v2
+#        with:
+#          python-version: 3.9
+      - name: Install Conda dependencies
+        run: |
+          conda install numpy statsmodels scikit-learn PyYAML \
+          yaml traits traitsui pyface envisage sqlalchemy \
+          Reportlab lxml xlrd xlwt xlsxwriter requests keyring \
+          pyparsing pillow gitpython pytables pyproj pymysql \
+          certifi jinja2 swig=3 cython importlib_resources
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install flake8 pytest uncertainties qdimage2ndarray peakutils chaco
+      - name: Lint with flake8
+        run: |
+          # stop the build if there are Python syntax errors or undefined names
+          flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+          flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+      - name: Test with pytest
+        run: |
+          pytest
