+++ .github/workflows/install.yml
@@ -38,7 +38,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install flake8 pytest uncertainties qdimage2ndarray peakutils chaco
+          pip install flake8 pytest uncertainties qimage2ndarray peakutils chaco
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
