+++ .github/workflows/install.yml
@@ -30,9 +30,7 @@ jobs:
 #          python-version: 3.9
       - name: Install Conda dependencies
         run: |
-          conda install -n test -c conda-forge gcc numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface
-          envisage
-          sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
+          conda install -n test -c conda-forge gcc numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface envisage sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
       - name: Install dependencies
         run: |
           unset PYTHONPATH
