+++ .github/workflows/install.yml
@@ -31,6 +31,9 @@ jobs:
 #        uses: actions/setup-python@v2
 #        with:
 #          python-version: 3.9
+      - name: Install swig
+        run: |
+          apt install swig
       - name: Install Conda dependencies
         run: |
           conda install -n test -c conda-forge numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface envisage sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
