+++ .github/workflows/install.yml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
 #        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-        os: [ "ubuntu-latest", ]
+        os: [ "ubuntu-latest", "macos-latest"]
         python-version: [ "3.9",]
     steps:
       - uses: conda-incubator/setup-miniconda@v2
@@ -37,8 +37,8 @@ jobs:
       - name: Install dependencies
         run: |
           unset PYTHONPATH
-          $CONDA/envs/test/bin/python -m pip install --upgrade pip
-          $CONDA/envs/test/bin/python -m pip install --upgrade-strategy only-if-needed flake8 pytest chaco
+          $CONDA/envs/test/bin/pip3 install --upgrade pip
+          $CONDA/envs/test/bin/pip3 install --upgrade-strategy only-if-needed flake8 pytest chaco
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
