+++ .github/workflows/install.yml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
 #        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-        os: [ "ubuntu-latest", "macos-latest"]
+        os: [ "macos-latest"]
         python-version: [ "3.9",]
     steps:
       - uses: conda-incubator/setup-miniconda@v2
@@ -36,8 +36,7 @@ jobs:
 #          sudo apt-get install swig=3.0.\*
       - name: Install Conda dependencies
         run: |
-          conda install -n test -c conda-forge pip numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface envisage sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
-          $CONDA/envs/test/bin/pip3 install --upgrade-strategy only-if-needed flake8 pytest
+          conda install -n test -c conda-forge pip flake8 pytest numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface envisage sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
           $CONDA/envs/test/bin/pip3 install -e git+https://github.com/enthought/enable.git#egg=enable
           $CONDA/envs/test/bin/pip3 install -e git+https://github.com/enthought/chaco.git#egg=chaco
       - name: Lint with flake8
