+++ app_utils/install.py
@@ -454,6 +454,20 @@ EXPERIMENT_DEFAULTS = """columns:
   - Comment
 
 """
+EDM_LAUNCHER_SH = """#!/bin/bash
+export GITHUB_ORGANIZATION={github_org:}
+export GITHUB_TOKEN={github_token:}
+export MassSpecDBVersion={massspec_db_version:}
+export APPLICATION_ID={app_id:}
+
+export PYCHRON_APPNAME={app_name:}
+export PYCHRON_DATABASE_UPDATE={update_db:}
+export PYCHRON_ALEMBIC_URL={alembic_url:}
+export PYCHRON_USE_LOGIN=0
+
+export PYTHONPATH={pychron_path:}
+
+{EDM_BIN:}/python {pychron_path:}/launchers/launcher.py"""
 
 LAUNCHER_SH = """#!/bin/bash
 export GITHUB_ORGANIZATION={github_org:}
@@ -509,6 +523,7 @@ else:
     GIT = 'git'
 
 HOME = os.path.expanduser('~')
+EDM_BIN = os.path.join(HOME, '.edm', 'envs', 'edm', 'bin')
 
 
 def info_header(msg):
@@ -592,6 +607,24 @@ def which(program):
                 return exe_file
 
 
+def install_edm_only():
+    info_header('Install edm env only')
+    cfg = {'qt_bindings': 'pyqt=5',
+           'qt_api': 'pyqt5',
+           'use_edm': True,
+           'install_gis_plugin': False,
+           # 'conda_distro': os.path.join(os.path.expanduser('~'), 'miniconda3'),
+           # 'edm_env_name': 'pychron-stable',
+           # 'python_version': '3.9'
+           }
+    if yes('Install edm env only [y]/n '):
+        # ask(cfg, 'conda_distro', 'Conda Distro Path')
+        # ask(cfg, 'conda_env_name', 'Conda environment name')
+        # ask(cfg, 'install_gis_plugin', 'Install GIS Plugin')
+        build_requirements(cfg)
+        return cfg
+
+
 def install_conda_only():
     info_header('Install conda env only')
     cfg = {'qt_bindings': 'pyqt=5',
@@ -611,8 +644,7 @@ def install_conda_only():
 def install_setupfiles_only():
     info_header('Install Setupfiles only')
     cfg = {}
-    vv = input('Install setup files only [n]')
-    if vv.lower() in ('y', 'yes'):
+    if no('Install setup files only y/[n]'):
         cfg['install_exp_setupfiles'] = True
         cfg['pychron_data_dir'] = 'PychronUF'
         cfg['include_hardware_plugins'] = True
@@ -626,6 +658,20 @@ def ask(cfg, key, msg):
     if not vv:
         vv = default
     cfg[key] = vv
+
+
+def yes(msg):
+    if not msg.endswith(' '):
+        msg = '{} '.format(msg)
+
+    return input(msg) in ('', 'y', 'yes', 'Yes', 'YES')
+
+
+def no(msg):
+    if not msg.endswith(' '):
+        msg = '{} '.format(msg)
+
+    return input(msg).lower() in ('y', 'yes')
 
 
 def ask_config():
@@ -652,7 +698,8 @@ def ask_config():
               'alembic_url': 'mysql+pymysql://<user>:<pwd>@<host>/<db>',
               'install_gis_plugin': False,
               'install_exp_setupfiles': False,
-              'include_hardware_plugins': False
+              'include_hardware_plugins': False,
+              'use_edm': True
               }
 
     ask(config, 'use_all_defaults', 'Use all defaults')
@@ -664,6 +711,10 @@ def ask_config():
         ask(config, 'branch', 'Pychron Branch')
[…]
