### 2021-10-15 413dd08
added install_edm action, fixed sparrow client
