+++ .github/workflows/install_edm.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Python application
+name: Python Install via EDM
 
 on:
   push:
@@ -35,7 +35,7 @@ jobs:
           scipy sqlalchemy traits xlrd xlsxwriter xlwt
       - name: Install PIP dependencies
         run: >
-          ~/.edm/envs/edm/bin python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
+          ~/.edm/envs/edm/bin/python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
 #      - name: Lint with flake8
 #        run: |
 #          # stop the build if there are Python syntax errors or undefined names
