+++ .github/workflows/install_edm.yml
@@ -40,9 +40,9 @@ jobs:
           python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
       - name: Lint with flake8
         run: |
-#          # stop the build if there are Python syntax errors or undefined names
+#          stop the build if there are Python syntax errors or undefined names
           flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+#          exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         run: |
