### 2021-10-15 941d567
fixed install_edm
