### 2021-10-15 ef2c768
code formatting. flake8 updates
