+++ pychron/hardware/mikron_pyrometer.py
@@ -278,7 +278,6 @@ class MikronGA140Pyrometer(CoreDevice):
         )
         return cg
 
-
 # ============= EOF =============================================
 
 #    def scan(self, *args):
