+++ pychron/data_mapper/__init__.py
@@ -81,5 +81,4 @@ def do_import_analyses(dvc, sources):
     info = open_view(view)
     return info.result
 
-
 # ============= EOF =============================================
