+++ pychron/external_pipette/protocol.py
@@ -18,6 +18,7 @@
 from __future__ import absolute_import
 from traits.api import Interface
 
+
 # ============= standard library imports ========================
 # ============= local library imports  ==========================
 class IPipetteManager(Interface):
@@ -38,6 +39,5 @@ class IPipetteManager(Interface):
 
     def set_extract_state(self, state, *args, **kw):
         pass
-
 
 # ============= EOF =============================================
