+++ pychron/furnace/base_furnace_manager.py
@@ -73,4 +73,5 @@ class BaseFurnaceManager(StreamGraphManager):
         self.info("testing connection")
         return self.test_furnace_api()
 
+
 # ============= EOF =============================================
