+++ pychron/options/icfactor.py
@@ -73,4 +73,5 @@ class ICFactorOptions(RatioSeriesOptions):
         self.aux_plots = pp
         self.selected = []
 
+
 # ============= EOF =============================================
