+++ pychron/processing/k3739_edit.py
@@ -120,4 +120,5 @@ class K3739EditView(Controller):
         )
         return v
 
+
 # ============= EOF =============================================
