+++ pychron/envisage/pychron_run.py
@@ -284,4 +284,5 @@ def launch(klass):
         # finally:
         #     sys.exit(0)
 
+
 # ============= EOF ====================================
