+++ pychron/hardware/thermo_spectrometer_controller.py
@@ -32,4 +32,5 @@ class ArgusController(CoreDevice):
 class HelixController(CoreDevice):
     pass
 
+
 # ============= EOF =============================================
