### 2021-10-15 f89a9f2
fixed test_suite discovery
