+++ .github/workflows/install_edm.yml
@@ -40,7 +40,7 @@ jobs:
           ~/.edm/envs/edm/bin/python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
       - name: Test with pytest
         run: |
-          ~/.edm/envs/edm/bin/pytest pychron/test_suite.py
+          ~/.edm/envs/edm/bin/pytest ./test_suite.py
       - name: Lint with flake8
         run: |
           ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
