### 2021-10-15 6645401
fixed test_suite discovery
