### 2021-10-15 e7fff72
fixed test_suite discovery
