### 2021-10-15 6486f1a
fixed test_suite discovery
