### 2021-10-15 347ad8f
fixed test_suite discovery
