### 2021-10-15 16cd699
fixed test_suite discovery
