### 2021-10-15 cedf596
fixed test_suite discovery
