### 2021-10-15 841f2e9
fixed test_suite discovery
