+++ .github/workflows/install_edm.yml
@@ -41,7 +41,7 @@ jobs:
 
       - name: Test with pytest
         run: |
-          ~/.edm/envs/edm/bin/pytest test_suite.py
+          ~/.edm/envs/edm/bin/pytest
       - name: Lint with flake8
         run: |
           ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
