### 2021-10-15 9ebc42b
fixed test_suite discovery
