+++ .github/workflows/install_edm.yml
@@ -39,9 +39,9 @@ jobs:
         run: >
           ~/.edm/envs/edm/bin/python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
 
-      - name: Test with pytest
+      - name: Test with unittest
         run: |
-          ~/.edm/envs/edm/bin/pytest
+          ~/.edm/envs/edm/bin/python test_suite.py
       - name: Lint with flake8
         run: |
           ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
