### 2021-10-15 e360254
fixed test_suite discovery
