### 2021-10-16 a65aaaa
fixed test_suite discovery
