### 2021-10-16 62fefd0
fixed test_suite discovery
