### 2021-10-16 0a6bd1e
fixed test_suite discovery
