### 2021-10-16 63f0a05
fixed test_suite discovery
