+++ .github/workflows/install_edm.yml
@@ -40,10 +40,10 @@ jobs:
           ~/.edm/envs/edm/bin/python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
       - name: set pythonpath
         run: |
-          echo "PYTHONPATH=/home/runner/work/pychron/" >> $GITHUB_ENV
-      - name: Test with unittest
+          echo "PYTHONPATH=/home/runner/work/pychron/pychron" >> $GITHUB_ENV
+      - name: Test with pytest
         run: |
-          ~/.edm/envs/edm/bin/python pychron/test_suite.py
+          ~/.edm/envs/edm/bin/python -m pytest pychron/test_suite.py
       - name: Lint with flake8
         run: |
           ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
