### 2021-10-16 526dda5
fixed test_suite discovery
