+++ .github/workflows/install_edm.yml
@@ -38,12 +38,9 @@ jobs:
       - name: Install PIP dependencies
         run: >
           ~/.edm/envs/edm/bin/python -m pip install --no-dependencies uncertainties statsmodels qimage2ndarray peakutils
-      - name: set pythonpath
-        run: |
-          echo "PYTHONPATH=/home/runner/work/pychron/pychron" >> $GITHUB_ENV
       - name: Test with pytest
         run: |
-          ~/.edm/envs/edm/bin/python -m pytest pychron/test_suite.py
+          PYTHONPATH=pychron/ ~/.edm/envs/edm/bin/python -m pytest pychron/test_suite.py
       - name: Lint with flake8
         run: |
           ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
