### 2021-10-16 0a46c87
fixed test_suite discovery
