### 2021-10-16 00fc549
fixed test_suite discovery
