### 2021-10-16 2cdf57a
simplified test suite
