+++ pychron/test_suite.py
@@ -2,152 +2,152 @@ import os
 import unittest
 
 from pychron.core.helpers.tests.floatfmt import SigFigStdFmtTestCase
-from pychron.core.stats.tests.mswd_tests import MSWDTestCase
-from pychron.pyscripts.tests.extraction_script import WaitForTestCase
-from pychron.canvas.canvas2D.tests.calibration_item import CalibrationObjectTestCase
-
-# Core
-from pychron.core.tests.spell_correct import SpellCorrectTestCase
-from pychron.core.tests.filtering_tests import FilteringTestCase
-from pychron.core.stats.tests.peak_detection_test import MultiPeakDetectionTestCase
-from pychron.core.helpers.tests.floatfmt import FloatfmtTestCase
-from pychron.core.helpers.tests.strtools import CamelCaseTestCase
-from pychron.core.xml.tests.xml_parser import XMLParserTestCase
-from pychron.core.regression.tests.regression import OLSRegressionTest, MeanRegressionTest, \
-    FilterOLSRegressionTest, OLSRegressionTest2, TruncateRegressionTest
-from pychron.core.tests.alpha_tests import AlphaTestCase
-
-# DataMapper
-from pychron.data_mapper.tests.usgs_vsc_file_source import USGSVSCFileSourceUnittest, \
-    USGSVSCIrradiationSourceUnittest
-from pychron.data_mapper.tests.nmgrl_legacy_source import NMGRLLegacySourceUnittest
-
-# Experiment
-from pychron.experiment.tests.repository_identifier import ExperimentIdentifierTestCase
-from pychron.experiment.tests.peak_hop_parse import PeakHopYamlCase1
-from pychron.experiment.tests.peak_hop_parse import PeakHopYamlCase2
-from pychron.experiment.tests.backup import BackupTestCase
-from pychron.experiment.tests.peak_hop_parse import PeakHopTxtCase
-from pychron.experiment.tests.duration_tracker import DurationTrackerTestCase
-from pychron.experiment.tests.frequency_test import FrequencyTestCase, FrequencyTemplateTestCase
-from pychron.experiment.tests.position_regex_test import XYTestCase
-from pychron.experiment.tests.renumber_aliquot_test import RenumberAliquotTestCase
-from pychron.experiment.tests.conditionals import ConditionalsTestCase, ParseConditionalsTestCase
-from pychron.experiment.tests.identifier import IdentifierTestCase
-from pychron.experiment.tests.comment_template import CommentTemplaterTestCase
-
-# ExternalPipette
-from pychron.external_pipette.tests.external_pipette import ExternalPipetteTestCase
-
-# Processing
-from pychron.processing.tests.plateau import PlateauTestCase
-from pychron.processing.tests.ratio import RatioTestCase
-from pychron.processing.tests.age_converter import AgeConverterTestCase
-
-# Pyscripts
+# from pychron.core.stats.tests.mswd_tests import MSWDTestCase
 # from pychron.pyscripts.tests.extraction_script import WaitForTestCase
-from pychron.pyscripts.tests.measurement_pyscript import InterpolationTestCase, DocstrContextTestCase
-
-# Spectrometer
-from pychron.spectrometer.tests.mftable import MFTableTestCase, DiscreteMFTableTestCase
-from pychron.spectrometer.tests.integration_time import IntegrationTimeTestCase
-
-from pychron.stage.tests.stage_map import StageMapTestCase, TransformTestCase
-
-os.environ['MassSpecDBVersion'] = '16'
-
-from pychron.paths import paths
-
-paths.build('_dev')
-
-use_logger = False
-
-
-def suite():
[…]
