### 2021-10-16 5445d2b
simplified test suite
