+++ .github/workflows/install_edm.yml
@@ -10,6 +10,21 @@ on:
     branches: [ dev/dr ]
 
 jobs:
+  flake:
+    name: Flake8 the source
+    strategy:
+      matrix:
+        os: [ macos-latest, ubuntu-latest, windows-latest ]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+        with:
+          ref: ${{ github.head_ref }}
+      - name: Lint with flake8
+        run: |
+          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
+          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --ignore E501
   build:
     name:  Pychron Install via EDM (${{ matrix.python-version }}, ${{ matrix.os }})
     strategy:
@@ -41,10 +56,6 @@ jobs:
       - name: Test with pytest
         run: |
           ~/.edm/envs/edm/bin/python -m unittest pychron/test_suite.py
-      - name: Lint with flake8
-        run: |
-          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
-          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --ignore E501
 
 
 #
