### 2021-10-16 705b10b
simplified test suite
