+++ .github/workflows/install_edm.yml
@@ -6,25 +6,55 @@ name: Pychron Install via EDM with pytest suite
 on:
   push:
     branches: [ dev/dr ]
+    paths:
+      - '*.py'
   pull_request:
     branches: [ dev/dr ]
+    paths:
+      - '*.py'
 
 jobs:
-  flake:
-    name: Flake8 the source
-    strategy:
-      matrix:
-        os: [ macos-latest, ubuntu-latest, windows-latest ]
-    runs-on: ${{ matrix.os }}
+  flake8_py3:
+    runs-on: ubuntu-latest
     steps:
-      - name: Checkout
-        uses: actions/checkout@v2
+      - name: Setup Python
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.7.4
+          architecture: x64
+      - name: Checkout pychron
+        uses: actions/checkout@master
         with:
           ref: ${{ github.head_ref }}
-      - name: Lint with flake8
-        run: |
-          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
-          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --ignore E501
+      - name: Install flake8
+        run: pip install flake8
+      - name: Run flake8
+        uses: py-actions/flake8@v1
+          with:
+            ignore: "F401"
+            exclude: "docs,sandbox,tests,alembic_dvc,zobs"
+            max-line-length: "100"
+#        uses: suo/flake8-github-action@releases/v1
+#        with:
+#          checkName: 'flake8_py3'   # NOTE: this needs to be the same as the job name
+#        env:
+#          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+#  flake:
+#    name: Flake8 the source
+#    strategy:
+#      matrix:
+#        os: [ macos-latest, ubuntu-latest, windows-latest ]
+#    runs-on: ${{ matrix.os }}
+#    steps:
+#      - name: Checkout
+#        uses: actions/checkout@v2
+#        with:
+#          ref: ${{ github.head_ref }}
+#      - name: Lint with flake8
+#        run: |
+#          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
+#          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --ignore E501
   build:
     name:  Pychron Install via EDM (${{ matrix.python-version }}, ${{ matrix.os }})
     strategy:
