### 2021-10-16 7fe7534
simplified test suite
