+++ .github/workflows/install_edm.yml
@@ -30,10 +30,10 @@ jobs:
         run: pip install flake8
       - name: Run flake8
         uses: py-actions/flake8@v1
-          with:
-            ignore: "F401"
-            exclude: "docs,sandbox,tests,alembic_dvc,zobs"
-            max-line-length: "100"
+        with:
+          ignore: "F401"
+          exclude: "docs,sandbox,tests,alembic_dvc,zobs"
+          max-line-length: "100"
 #        uses: suo/flake8-github-action@releases/v1
 #        with:
 #          checkName: 'flake8_py3'   # NOTE: this needs to be the same as the job name
