### 2021-10-16 51a3897
simplified test suite
