+++ .github/workflows/install_edm.yml
@@ -17,44 +17,18 @@ jobs:
   flake8_py3:
     runs-on: ubuntu-latest
     steps:
-      - name: Setup Python
+      - name: Check out source repository
+        uses: actions/checkout@v2
+      - name: Set up Python environment
         uses: actions/setup-python@v1
         with:
-          python-version: 3.7.4
-          architecture: x64
-      - name: Checkout pychron
-        uses: actions/checkout@master
-        with:
-          ref: ${{ github.head_ref }}
-      - name: Install flake8
-        run: pip install flake8
+          python-version: "3.8"
       - name: Run flake8
         uses: py-actions/flake8@v1
         with:
           ignore: "F401"
           exclude: "docs,sandbox,tests,alembic_dvc,zobs"
           max-line-length: "100"
-#        uses: suo/flake8-github-action@releases/v1
-#        with:
-#          checkName: 'flake8_py3'   # NOTE: this needs to be the same as the job name
-#        env:
-#          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-#  flake:
-#    name: Flake8 the source
-#    strategy:
-#      matrix:
-#        os: [ macos-latest, ubuntu-latest, windows-latest ]
-#    runs-on: ${{ matrix.os }}
-#    steps:
-#      - name: Checkout
-#        uses: actions/checkout@v2
-#        with:
-#          ref: ${{ github.head_ref }}
-#      - name: Lint with flake8
-#        run: |
-#          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --select=E9,F63,F7,F82 --show-source --statistics
-#          ~/.edm/envs/edm/bin/flake8 . --extend-exclude=docs/,zobs/,sandbox/,test/,alembic_dvc --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --ignore E501
   build:
     name:  Pychron Install via EDM (${{ matrix.python-version }}, ${{ matrix.os }})
     strategy:
@@ -88,45 +62,3 @@ jobs:
           ~/.edm/envs/edm/bin/python -m unittest pychron/test_suite.py
 
 
-#
-#jobs:
-#  job-1:
-#    name: Pychron Install (${{ matrix.python-version }}, ${{ matrix.os }})
-#    runs-on: ${{ matrix.os }}
-#
-#
-#    strategy:
-#      fail-fast: false
-#      matrix:
-##        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-#        os: [ "macos-latest"]
-#        python-version: [ "3.9",]
-#    steps:
-#      - uses: conda-incubator/setup-miniconda@v2
-#        with:
-#          activate-environment: test
-#          auto-activate-base: false
-#          auto-update-conda: true
-#          python-version: ${{ matrix.python-version }}
-#      - uses: actions/checkout@v2
-##      - name: Set up Python 3.9
-##        uses: actions/setup-python@v2
-##        with:
-##          python-version: 3.9
-##      - name: Install swig 3
-##        run: |
-##          sudo apt-get install swig=3.0.\*
-#      - name: Install Conda dependencies
-#        run: |
-#          conda install -n test -c conda-forge pip flake8 pytest numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface envisage sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
-#          $CONDA/envs/test/bin/pip3 install -e git+https://github.com/enthought/enable.git#egg=enable
[…]
