### 2021-10-16 805d7e4
simplified test suite
