### 2021-10-16 aba0256
simplified test suite
