+++ pychron/test_suite.py
@@ -15,7 +15,7 @@ from pychron.pyscripts.tests.extraction_script import WaitForTestCase
 # from pychron.core.xml.tests.xml_parser import XMLParserTestCase
 # from pychron.core.regression.tests.regression import OLSRegressionTest, MeanRegressionTest, \
 #     FilterOLSRegressionTest, OLSRegressionTest2, TruncateRegressionTest
-# from pychron.core.tests.alpha_tests import AlphaTestCase
+from pychron.core.tests.alpha_tests import AlphaTestCase
 #
 # # DataMapper
 # from pychron.data_mapper.tests.usgs_vsc_file_source import USGSVSCFileSourceUnittest, \
@@ -82,7 +82,7 @@ def suite():
         # CalibrationObjectTestCase,
 
         # Core
-        # AlphaTestCase,
+        AlphaTestCase,
         # SpellCorrectTestCase,
         # FilteringTestCase,
         # MultiPeakDetectionTestCase,
