### 2021-10-16 290dba1
simplified test suite
