### 2021-10-16 c5ca149
simplified test suite
