### 2021-10-16 ee98f4b
simplified test suite
