+++ pychron/test_suite.py
@@ -20,6 +20,7 @@ from pychron.core.helpers.tests.strtools import CamelCaseTestCase
 #     FilterOLSRegressionTest, OLSRegressionTest2, TruncateRegressionTest
 from pychron.core.tests.alpha_tests import AlphaTestCase
 
+
 #
 # # DataMapper
 # from pychron.data_mapper.tests.usgs_vsc_file_source import USGSVSCFileSourceUnittest, \
