+++ pychron/core/helpers/fits.py
@@ -51,12 +51,16 @@ def convert_fit(f):
 
         if f in FITS:
             f = FITS.index(f) + 1
-        elif f in (
-            "average",
-            "weighted mean",
-            EXPONENTIAL,
-            AUTO_LINEAR_PARABOLIC.lower(),
-        ) or f.startswith("custom:"):
+        elif (
+            f
+            in (
+                "average",
+                "weighted mean",
+                EXPONENTIAL,
+                AUTO_LINEAR_PARABOLIC.lower(),
+            )
+            or f.startswith("custom:")
+        ):
             if not err:
                 err = "SEM" if "sem" in f else "SD"
         else:
