+++ pychron/pychron_constants.py
@@ -24,234 +24,288 @@ from pychron.core.helpers.formatting import floatfmt
 from pychron.core.yaml import yload
 from pychron.paths import paths
 
-IS_WINDOWS = platform.system() == 'Windows'
+IS_WINDOWS = platform.system() == "Windows"
 
 STARTUP_MESSAGE_POSITION = (100, 300)
 
-SPECTROMETER_PROTOCOL = 'pychron.spectrometer.base_spectrometer_manager.BaseSpectrometerManager'
-ION_OPTICS_PROTOCOL = 'pychron.spectrometer.ion_optics_manager.IonOpticsManager'
-SCAN_PROTOCOL = 'pychron.spectrometer.scan_manager.ScanManager'
-EL_PROTOCOL = 'pychron.extraction_line.extraction_line_manager.ExtractionLineManager'
-DVC_PROTOCOL = 'pychron.dvc.dvc.DVC'
-FURNACE_PROTOCOL = 'pychron.furnace.furnace_manager.BaseFurnaceManager'
-ILASER_PROTOCOL = 'pychron.lasers.laser_managers.ilaser_manager.ILaserManager'
-IFURNACE_PROTOCOL = 'pychron.furnace.ifurnace_manager.IFurnaceManager'
-IPIPETTE_PROTOCOL = 'pychron.external_pipette.protocol.IPipetteManager'
-CRYO_PROTOCOL = 'pychron.extraction_line.cryo_manager.CryoManager'
+SPECTROMETER_PROTOCOL = (
+    "pychron.spectrometer.base_spectrometer_manager.BaseSpectrometerManager"
+)
+ION_OPTICS_PROTOCOL = "pychron.spectrometer.ion_optics_manager.IonOpticsManager"
+SCAN_PROTOCOL = "pychron.spectrometer.scan_manager.ScanManager"
+EL_PROTOCOL = "pychron.extraction_line.extraction_line_manager.ExtractionLineManager"
+DVC_PROTOCOL = "pychron.dvc.dvc.DVC"
+FURNACE_PROTOCOL = "pychron.furnace.furnace_manager.BaseFurnaceManager"
+ILASER_PROTOCOL = "pychron.lasers.laser_managers.ilaser_manager.ILaserManager"
+IFURNACE_PROTOCOL = "pychron.furnace.ifurnace_manager.IFurnaceManager"
+IPIPETTE_PROTOCOL = "pychron.external_pipette.protocol.IPipetteManager"
+CRYO_PROTOCOL = "pychron.extraction_line.cryo_manager.CryoManager"
 
-TTF_FONTS = ['Andale Mono', 'Arial', 'Arial Black', 'Arial Unicode MS',
-             'Calibri', 'Cambria', 'Comic Sans MS', 'Consolas', 'Courier New',
-             'Georgia',
-             'Helvetica',
-             'Impact',
-             'Trebuchet MS',
-             'Verdana']
-
+TTF_FONTS = [
+    "Andale Mono",
+    "Arial",
+    "Arial Black",
+    "Arial Unicode MS",
+    "Calibri",
+    "Cambria",
+    "Comic Sans MS",
+    "Consolas",
+    "Courier New",
[…]
