+++ .github/workflows/format_code.yml
@@ -27,8 +27,8 @@ jobs:
           ref: ${{ github.head_ref }}
       - uses: psf/black@stable
         with:
-          options: " --verbose --extend-exclude [sandbox|docs|dist|zobs|test|alembic_dvc]"
-          src: "pychron/**/*.py"
+          options: "--verbose --extend-exclude [sandbox|docs|dist|zobs|test|alembic_dvc]"
+          src: "pychron"
 #      - uses: rickstaa/action-black@v1
 #        with:
 #          black_args: " pychron/**/*.py "
