+++ pyproject.toml
@@ -1,3 +1,6 @@
 [tool.black]
-extend-exclude= "docs,sandbox,test,alembic_dvc,zobs"
+include = '\.py?$'
+extend-exclude= "^/[docs|sandbox|test|alembic_dvc|zobs]"
 max-line-length= "100"
+verbose = "true"
+# check = "true"
