+++ .github/workflows/format_code.yml
@@ -26,9 +26,9 @@ jobs:
         with:
           ref: ${{ github.head_ref }}
       - uses: psf/black@stable
-        with:
-          options: "--verbose --extend-exclude [sandbox|docs|dist|zobs|test|alembic_dvc]"
-          src: "pychron"
+#        with:
+#          options: "--verbose --extend-exclude [sandbox|docs|dist|zobs|test|alembic_dvc]"
+#          src: "pychron/"
 #      - uses: rickstaa/action-black@v1
 #        with:
 #          black_args: " pychron/**/*.py "
