+++ pychron/pipeline/plot/plotter/references_series.py
@@ -47,19 +47,21 @@ def calc_limits(ys, ye, n):
 
 def unzip_data(data):
     try:
-        return array([nominal_value(ri) for ri in data]), array([std_dev(ri) for ri in data])
+        return array([nominal_value(ri) for ri in data]), array(
+            [std_dev(ri) for ri in data]
+        )
     except ValueError as e:
         print(e)
 
 
 class ReferencesSeries(BaseSeries):
     references = List
-    sorted_references = Property(depends_on='references')
+    sorted_references = Property(depends_on="references")
     show_current = True
     rxs = Array
-    references_name = 'References'
-    xtitle = 'Time (hrs)'
-    _normalization_factor = 3600.
+    references_name = "References"
+    xtitle = "Time (hrs)"
+    _normalization_factor = 3600.0
 
     def set_interpolated_values(self, iso, reg, fit):
         mi, ma = self._get_min_max()
@@ -97,27 +99,31 @@ class ReferencesSeries(BaseSeries):
                     graph.add_statistics(plotid=i)
 
             mi, ma = self._get_min_max()
-            self.xmi, self.xma = (mi - ma) / 3600., 0
-            self.xpad = '0.1'
+            self.xmi, self.xma = (mi - ma) / 3600.0, 0
+            self.xpad = "0.1"
 
-            legend = ExplicitLegend(plots=self.graph.plots[0].plots,
-                                    labels=[('plot1', self.references_name),
-                                            ('data0', self.references_name),
-                                            ('plot0', 'Unk. Current'),
-                                            ('Unknowns-predicted0', 'Unk. Predicted')])
+            legend = ExplicitLegend(
+                plots=self.graph.plots[0].plots,
+                labels=[
+                    ("plot1", self.references_name),
+                    ("data0", self.references_name),
+                    ("plot0", "Unk. Current"),
+                    ("Unknowns-predicted0", "Unk. Predicted"),
+                ],
+            )
             self.graph.plots[-1].overlays.append(legend)
 
     # private
-    @on_trait_change('graph:regression_results')
+    @on_trait_change("graph:regression_results")
     def _update_regression(self, new):
-        key = 'Unknowns-predicted{}'
+        key = "Unknowns-predicted{}"
         key = key.format(0)
         for plotobj, reg in new:
             if isinstance(reg, BaseRegressor):
 
                 excluded = reg.get_excluded()
                 for i, r in enumerate(self.sorted_references):
-                    r.set_temp_status('omit' if i in excluded else 'ok')
+                    r.set_temp_status("omit" if i in excluded else "ok")
 
                 self._set_values(plotobj, reg, key)
 
@@ -135,14 +141,19 @@ class ReferencesSeries(BaseSeries):
     def _calc_limits(self, ys, ye):
         return calc_limits(ys, ye, self.options.nsigma)
 
-    def _add_plot_label(self, pid, po, overlay_position='inside top', hjustify='left', **kw):
+    def _add_plot_label(
+        self, pid, po, overlay_position="inside top", hjustify="left", **kw
+    ):
         txt = self._get_plot_label_text(po)
         if txt:
             comp = self.graph.plots[pid]
-            pl = OffsetPlotLabel(txt,
-                                 component=comp,
-                                 overlay_position=overlay_position, hjustify=hjustify,
-                                 **kw)
+            pl = OffsetPlotLabel(
+                txt,
+                component=comp,
+                overlay_position=overlay_position,
+                hjustify=hjustify,
+                **kw
+            )
             comp.overlays.append(pl)
 
     def _get_plot_label_text(self, po):
[…]
