+++ README.md
@@ -1,7 +1,7 @@
 Pychron
 ========
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3237834.svg)](https://doi.org/10.5281/zenodo.3237834)
-[![Format code](https://github.com/NMGRL/pychron/actions/workflows/format_code.yml/badge.svg?branch=dev%2Fdr)](https://github.com/NMGRL/pychron/actions/workflows/format_code.yml)
+[![Format code](https://github.com/NMGRL/pychron/actions/workflows/format_code.yml/badge.svg?branch=develop)](https://github.com/NMGRL/pychron/actions/workflows/format_code.yml)
 
 [Changes](CHANGELOG.md)
 
