+++ pychron/hardware/core/scanable_device.py
@@ -16,10 +16,12 @@
 # ============= enthought library imports =======================
 from traits.api import Event, Property, Any, Bool, Float, Str, Instance, List
 from traitsui.api import HGroup, VGroup, Item, spring, ButtonEditor
+
 # ============= standard library imports ========================
 import os
 import time
 from threading import Lock
+
 # ============= local library imports  ==========================
 from pychron.core.helpers.datetime_tools import generate_datetimestamp
 from pychron.database.data_warehouse import DataWarehouse
@@ -32,7 +34,7 @@ from pychron.paths import paths
 
 class ScanableDevice(ViewableDevice):
     scan_button = Event
-    scan_label = Property(depends_on='_scanning')
+    scan_label = Property(depends_on="_scanning")
 
     alarms = List(Alarm)
 
@@ -42,7 +44,7 @@ class ScanableDevice(ViewableDevice):
     timer = None
     scan_period = Float(1000, enter_set=True, auto_set=False)
     scan_width = Float(5, enter_set=True, auto_set=False)
-    scan_units = 'ms'
+    scan_units = "ms"
     record_scan_data = Bool(False)
     graph_scan_data = Bool(False)
     scan_path = Str
@@ -50,7 +52,7 @@ class ScanableDevice(ViewableDevice):
     scan_root = Str
     scan_name = Str
 
-    graph = Instance('pychron.graph.graph.Graph')
+    graph = Instance("pychron.graph.graph.Graph")
     graph_ytitle = Str
     graph_klass = None
 
@@ -71,32 +73,45 @@ class ScanableDevice(ViewableDevice):
     # streamin interface
     # ===============================================================================
     def setup_scan(self):
-        self.debug('setup scan')
+        self.debug("setup scan")
         # should get scan settings from the config file not the initialization.xml
         config = self.get_configuration()
-        if config.has_section('Scan'):
-            enabled = self.config_get(config, 'Scan', 'enabled', cast='boolean', optional=True, default=True)
+        if config.has_section("Scan"):
+            enabled = self.config_get(
+                config, "Scan", "enabled", cast="boolean", optional=True, default=True
+            )
             self.is_scanable = enabled
             if enabled:
-                self.set_attribute(config, 'auto_start', 'Scan', 'auto_start', cast='boolean', default=False)
-                self.set_attribute(config, 'scan_period', 'Scan', 'period', cast='float')
-                self.set_attribute(config, 'scan_width', 'Scan', 'width', cast='float')
-                self.set_attribute(config, 'scan_units', 'Scan', 'units')
-                self.set_attribute(config, 'record_scan_data', 'Scan', 'record', cast='boolean')
-                self.set_attribute(config, 'graph_scan_data', 'Scan', 'graph', cast='boolean')
+                self.set_attribute(
+                    config,
+                    "auto_start",
+                    "Scan",
+                    "auto_start",
+                    cast="boolean",
+                    default=False,
+                )
+                self.set_attribute(
+                    config, "scan_period", "Scan", "period", cast="float"
+                )
+                self.set_attribute(config, "scan_width", "Scan", "width", cast="float")
+                self.set_attribute(config, "scan_units", "Scan", "units")
+                self.set_attribute(
+                    config, "record_scan_data", "Scan", "record", cast="boolean"
+                )
+                self.set_attribute(
+                    config, "graph_scan_data", "Scan", "graph", cast="boolean"
+                )
 
-                func = self.config_get(config, 'Scan', 'function', optional=True)
+                func = self.config_get(config, "Scan", "function", optional=True)
                 if func:
                     self.scan_func = func
 
     def setup_alarms(self):
-        self.debug('setup alarms')
+        self.debug("setup alarms")
[…]
