+++ pychron/dvc/dvc_persister.py
@@ -607,9 +607,12 @@ class DVCPersister(BasePersister):
                 (d for d in per_spec.active_detectors if d.name == iso.detector), None
             )
 
-            isod = {'detector': iso.detector, 'name': iso.name,
-                    'units': detector.units,
-                    'serial_id': detector.serial_id if detector else '00000'}
+            isod = {
+                "detector": iso.detector,
+                "name": iso.name,
+                "units": detector.units,
+                "serial_id": detector.serial_id if detector else "00000",
+            }
 
             if clf is not None:
                 klass, prob = clf.predict_isotope(iso)
