+++ pychron/core/wait/wait_control.py
@@ -126,6 +126,7 @@ class WaitControl(Loggable):
 
     def pause(self):
         self._paused = True
+
     # ===============================================================================
     # private
     # ===============================================================================

[…]
