+++ pychron/experiment/tasks/experiment_panes.py
@@ -17,8 +17,21 @@
 # ============= enthought library imports =======================
 from pyface.tasks.traits_dock_pane import TraitsDockPane
 from traits.api import Color, Instance, DelegatesTo, List, Any, Property, Button, Event
-from traitsui.api import View, Item, UItem, VGroup, HGroup, spring, \
-    Group, Spring, Label, VSplit, UReadonly, ListEditor, Readonly
+from traitsui.api import (
+    View,
+    Item,
+    UItem,
+    VGroup,
+    HGroup,
+    spring,
+    Group,
+    Spring,
+    Label,
+    VSplit,
+    UReadonly,
+    ListEditor,
+    Readonly,
+)
 from traitsui.editors.api import TableEditor, EnumEditor
 from traitsui.table_column import ObjectColumn
 
@@ -31,9 +44,18 @@ from pychron.core.ui.led_editor import LEDEditor
 from pychron.envisage.icon_button_editor import icon_button_editor
 from pychron.experiment.plot_panel import PlotPanel
 from pychron.experiment.utilities.identifier import SPECIAL_NAMES
-from pychron.pychron_constants import MEASUREMENT_COLOR, EXTRACTION_COLOR, \
-    NOT_EXECUTABLE_COLOR, SKIP_COLOR, SUCCESS_COLOR, CANCELED_COLOR, \
-    TRUNCATED_COLOR, FAILED_COLOR, END_AFTER_COLOR, SPECIAL_IDENTIFIER
+from pychron.pychron_constants import (
+    MEASUREMENT_COLOR,
+    EXTRACTION_COLOR,
+    NOT_EXECUTABLE_COLOR,
+    SKIP_COLOR,
+    SUCCESS_COLOR,
+    CANCELED_COLOR,
+    TRUNCATED_COLOR,
+    FAILED_COLOR,
+    END_AFTER_COLOR,
+    SPECIAL_IDENTIFIER,
+)
 
 
 # ===============================================================================
@@ -46,11 +68,11 @@ def spacer(w):
 
 
 def queue_factory_name(name):
-    return 'object.queue_factory.{}'.format(name)
+    return "object.queue_factory.{}".format(name)
 
 
 def run_factory_name(name):
-    return 'object.run_factory.{}'.format(name)
+    return "object.run_factory.{}".format(name)
 
 
 def queue_factory_item(name, **kw):
@@ -58,7 +80,7 @@ def queue_factory_item(name, **kw):
 
 
 def queue_factory_uitem(name, **kw):
-    kw['show_label'] = False
+    kw["show_label"] = False
     return queue_factory_item(name, **kw)
 
 
@@ -67,264 +89,410 @@ def run_factory_item(name, **kw):
 
 
 def run_factory_uitem(name, **kw):
-    kw['show_label'] = False
+    kw["show_label"] = False
     return run_factory_item(name, **kw)
 
 
 class ExperimentFactoryPane(TraitsDockPane):
-    id = 'pychron.experiment.factory'
-    name = 'Experiment Editor'
-    info_label = Property(depends_on='model.run_factory.info_label')
+    id = "pychron.experiment.factory"
+    name = "Experiment Editor"
+    info_label = Property(depends_on="model.run_factory.info_label")
 
     def _get_info_label(self):
-        return '<font color="green"><b>{}</b></font>'.format(self.model.run_factory.info_label)
+        return '<font color="green"><b>{}</b></font>'.format(
+            self.model.run_factory.info_label
+        )
 
     def traits_view(self):
-        add_button = icon_button_editor('add_button', 'add',
-                                        # enabled_when='ok_add',
-                                        tooltip='Add run')
+        add_button = icon_button_editor(
+            "add_button",
+            "add",
[…]
