+++ pychron/experiment/tasks/experiment_task.py
@@ -32,27 +32,52 @@ from pychron.envisage.view_util import open_view
 from pychron.experiment.experiment_launch_history import update_launch_history
 from pychron.experiment.experimentor import Experimentor
 from pychron.experiment.queue.base_queue import extract_meta
-from pychron.experiment.tasks.experiment_editor import ExperimentEditor, UVExperimentEditor
-from pychron.experiment.tasks.experiment_panes import ExperimentFactoryPane, StatsPane, \
-    ControlsPane, IsotopeEvolutionPane, ConnectionStatusPane, LoggerPane, ExplanationPane, ConditionalsPane
+from pychron.experiment.tasks.experiment_editor import (
+    ExperimentEditor,
+    UVExperimentEditor,
+)
+from pychron.experiment.tasks.experiment_panes import (
+    ExperimentFactoryPane,
+    StatsPane,
+    ControlsPane,
+    IsotopeEvolutionPane,
+    ConnectionStatusPane,
+    LoggerPane,
+    ExplanationPane,
+    ConditionalsPane,
+)
 from pychron.experiment.utilities.identifier import convert_extract_device, is_special
 from pychron.experiment.utilities.save_dialog import ExperimentSaveDialog
 from pychron.globals import globalv
 from pychron.lasers.laser_managers.ilaser_manager import ILaserManager
 from pychron.paths import paths
 from pychron.pipeline.plot.editors.figure_editor import FigureEditor
-from pychron.pychron_constants import SPECTROMETER_PROTOCOL, DVC_PROTOCOL, COCKTAIL, AIR, BLANK, FUSIONS_UV, \
-    INVALID, EXTRACTION, MEASUREMENT, CANCELED, TRUNCATED, END_AFTER, FAILED, SUCCESS
+from pychron.pychron_constants import (
+    SPECTROMETER_PROTOCOL,
+    DVC_PROTOCOL,
+    COCKTAIL,
+    AIR,
+    BLANK,
+    FUSIONS_UV,
+    INVALID,
+    EXTRACTION,
+    MEASUREMENT,
+    CANCELED,
+    TRUNCATED,
+    END_AFTER,
+    FAILED,
+    SUCCESS,
+)
 
 
 class ExperimentEditorTask(EditorTask):
-    id = 'pychron.experiment.task'
-    name = 'Experiment'
+    id = "pychron.experiment.task"
+    name = "Experiment"
 
-    default_open_action = 'open files'
-    wildcard = '*.txt'
+    default_open_action = "open files"
+    wildcard = "*.txt"
 
-    loading_manager = Instance('pychron.loading.loading_manager.LoadingManager')
+    loading_manager = Instance("pychron.loading.loading_manager.LoadingManager")
 
     last_experiment_changed = Event
 
@@ -66,8 +91,8 @@ class ExperimentEditorTask(EditorTask):
     isotope_evolution_pane = Instance(IsotopeEvolutionPane)
     experiment_factory_pane = Instance(ExperimentFactoryPane)
 
-    load_pane = Instance('pychron.loading.tasks.panes.LoadDockPane')
-    load_table_pane = Instance('pychron.loading.tasks.panes.LoadTablePane')
+    load_pane = Instance("pychron.loading.tasks.panes.LoadDockPane")
+    load_table_pane = Instance("pychron.loading.tasks.panes.LoadTablePane")
     laser_control_client_pane = None
 
     events = None
@@ -76,9 +101,9 @@ class ExperimentEditorTask(EditorTask):
     deactivations = None
 
     def save_as_current_experiment(self):
-        self.debug('save as current experiment')
+        self.debug("save as current experiment")
         if self.has_active_editor():
-            path = os.path.join(paths.experiment_dir, 'CurrentExperiment.txt')
[…]
