+++ README.md
@@ -16,7 +16,11 @@ Pychron
 What is Pychron
 ===============
 
-Pychron is a set of applications for the collection and processing of noble gas mass spectrometry data. Pychron is developed at the New Mexico Geochronology Research Laboratory at New Mexico Tech. Components of pychron are used within multiple research domains, but mainly for Ar-Ar geochronology and thermochronology. Pychron's main applications are pyValve, pyLaser, pyExperiment and pyView. Additional components include RemoteControlServer.cs and Bakedpy.
+Pychron is a set of applications for the collection and processing of noble gas mass spectrometry data. Pychron is 
+developed at the New Mexico Geochronology Research Laboratory at New Mexico Tech. Components of pychron are used 
+within multiple research domains, but mainly for Ar-Ar geochronology and thermochronology. Pychron's main 
+applications are pyValve, pyLaser, pyExperiment and pyCrunch. Additional components include RemoteControlServer.cs and 
+Bakedpy.
 
 Pychron aims to augment and replace the current widely used program Mass Spec by Alan Deino of Berkeley Geochronology Center
 
@@ -45,6 +49,16 @@ Installation of Pychron at other laboratories is ongoing. Current interested lab
 Additionally, Remote Control Server, a script made by the pychron developers, is used extensively 
 by the international community to interface third-party software with Thermo Scientific's Mass Spectrometer control software.
 
+
+pyExperiment
+--------------
+Write and run a set of automated analyses. Allows NMGRL to operate continuously. only limited by size of analysis chamber.
+
+pyCrunch
+-------
+Display, process and publish Ar-Ar geochronology and thermochonology data. Export publication ready PDF tables and figures. Export Excel, CSV, and XML data tables. Store and search for figures in database.  
+
+
 pyValve
 -----------
 Used to control and monitor a noble gas extraction line a.k.a prep system. Displays a graphical interface for user to interact with. A RPC interface is also provided enabling control of the prep system by other applications.
@@ -53,14 +67,6 @@ pyLaser
 ----------
 Configure for multiple types of lasers. Currently compatible with Photon machines Fusions CO2, 810 diode and ATLEX UV lasers. Watlow or Eurotherm interface for PID control. Machine vision
 for laser auto targeting and modulated degassing.
-
-pyExperiment
---------------
-Write and run a set of automated analyses. Allows NMGRL to operate continuously. only limited by size of analysis chamber.
-
-pyView
--------
-Display, process and publish Ar-Ar geochronology and thermochonology data. Export publication ready PDF tables and figures. Export Excel, CSV, and XML data tables. Store and search for figures in database.  
 
 furPi
 -------
@@ -74,3 +80,9 @@ To turn off App Nap system wide use
 
     
     defaults write NSGlobalDomain NSAppSleepDisabled -bool YES
+
+
+# Citing Pychron 
+Are you using pychron for data collection and/or data reduction of published data please cite it by including as much
+information as possible from the following: *Jake Ross. (2019). NMGRL/pychron v18.2 (v18.2). Zenodo. https://doi.
+org/10.5281/zenodo.3237834*
