+++ README.md
@@ -16,7 +16,11 @@ Pychron
 What is Pychron
 ===============
 
-Pychron is a set of applications for the collection and processing of noble gas mass spectrometry data. Pychron is developed at the New Mexico Geochronology Research Laboratory at New Mexico Tech. Components of pychron are used within multiple research domains, but mainly for Ar-Ar geochronology and thermochronology. Pychron's main applications are pyValve, pyLaser, pyExperiment and pyView. Additional components include RemoteControlServer.cs and Bakedpy.
+Pychron is a set of applications for the collection and processing of noble gas mass spectrometry data. Pychron is 
+developed at the New Mexico Geochronology Research Laboratory at New Mexico Tech. Components of pychron are used 
+within multiple research domains, but mainly for Ar-Ar geochronology and thermochronology. Pychron's main 
+applications are pyValve, pyLaser, pyExperiment and pyCrunch. Additional components include RemoteControlServer.cs and 
+Bakedpy.
 
 Pychron aims to augment and replace the current widely used program Mass Spec by Alan Deino of Berkeley Geochronology Center
 
@@ -66,7 +70,7 @@ pyExperiment
 --------------
 Write and run a set of automated analyses. Allows NMGRL to operate continuously. only limited by size of analysis chamber.
 
-pyView
+pyCrunch
 -------
 Display, process and publish Ar-Ar geochronology and thermochonology data. Export publication ready PDF tables and figures. Export Excel, CSV, and XML data tables. Store and search for figures in database.  
 
