+++ pychron/test_suite.py
@@ -19,45 +19,21 @@ from pychron.core.xml.tests.xml_parser import XMLParserTestCase
 from pychron.core.regression.tests.regression import OLSRegressionTest, MeanRegressionTest, \
     FilterOLSRegressionTest, OLSRegressionTest2, TruncateRegressionTest
 from pychron.core.tests.alpha_tests import AlphaTestCase
-
-
-#
-# # DataMapper
-# from pychron.data_mapper.tests.usgs_vsc_file_source import USGSVSCFileSourceUnittest, \
-#     USGSVSCIrradiationSourceUnittest
-# from pychron.data_mapper.tests.nmgrl_legacy_source import NMGRLLegacySourceUnittest
-#
-# # Experiment
-# from pychron.experiment.tests.repository_identifier import ExperimentIdentifierTestCase
-# from pychron.experiment.tests.peak_hop_parse import PeakHopYamlCase1
-# from pychron.experiment.tests.peak_hop_parse import PeakHopYamlCase2
-# from pychron.experiment.tests.backup import BackupTestCase
-# from pychron.experiment.tests.peak_hop_parse import PeakHopTxtCase
-# from pychron.experiment.tests.duration_tracker import DurationTrackerTestCase
-# from pychron.experiment.tests.frequency_test import FrequencyTestCase, FrequencyTemplateTestCase
-# from pychron.experiment.tests.position_regex_test import XYTestCase
-# from pychron.experiment.tests.renumber_aliquot_test import RenumberAliquotTestCase
-# from pychron.experiment.tests.conditionals import ConditionalsTestCase, ParseConditionalsTestCase
-# from pychron.experiment.tests.identifier import IdentifierTestCase
-# from pychron.experiment.tests.comment_template import CommentTemplaterTestCase
-#
-# # ExternalPipette
-# from pychron.external_pipette.tests.external_pipette import ExternalPipetteTestCase
-#
-# # Processing
-# from pychron.processing.tests.plateau import PlateauTestCase
+from pychron.experiment.tests.backup import BackupTestCase
+from pychron.experiment.tests.comment_template import CommentTemplaterTestCase
+from pychron.experiment.tests.conditionals import ConditionalsTestCase, ParseConditionalsTestCase
+from pychron.experiment.tests.duration_tracker import DurationTrackerTestCase
+from pychron.experiment.tests.frequency_test import FrequencyTestCase, FrequencyTemplateTestCase
+from pychron.experiment.tests.identifier import IdentifierTestCase
+from pychron.experiment.tests.peak_hop_parse import PeakHopYamlCase1, PeakHopTxtCase
+from pychron.experiment.tests.peak_hop_parse import PeakHopYamlCase2
+from pychron.experiment.tests.position_regex_test import XYTestCase
+from pychron.experiment.tests.renumber_aliquot_test import RenumberAliquotTestCase
+from pychron.external_pipette.tests.external_pipette import ExternalPipetteTestCase
+from pychron.processing.tests.age_converter import AgeConverterTestCase
+from pychron.processing.tests.plateau import PlateauTestCase
 from pychron.processing.tests.ratio import RatioTestCase
-# from pychron.processing.tests.age_converter import AgeConverterTestCase
-#
-# # Pyscripts
-# # from pychron.pyscripts.tests.extraction_script import WaitForTestCase
-# from pychron.pyscripts.tests.measurement_pyscript import InterpolationTestCase, DocstrContextTestCase
-#
-# # Spectrometer
-# from pychron.spectrometer.tests.mftable import MFTableTestCase, DiscreteMFTableTestCase
[…]
