+++ pychron/git/tasks/base_git_plugin.py
@@ -42,7 +42,8 @@ class BaseGitPlugin(BasePlugin):
             )
         try:
             self.debug("checking for gh cli")
-            subprocess.call(["gh", "--version"])
+            subprocess.call(["gh", "--version"],
+                            stdout=subprocess.DEVNULL,)
             self.debug("github authentication handled by gh")
             return
         except FileNotFoundError:
