+++ pychron/spectrometer/thermo/spectrometer/__init__.py
@@ -16,7 +16,30 @@
 
 # ============= enthought library imports =======================
 # ============= standard library imports ========================
+from numpy import array, argmin
 # ============= local library imports  ==========================
+from pychron.pychron_constants import QTEGRA_INTEGRATION_TIMES
 
+
+def normalize_integration_time(it):
+    """
+    find the integration time closest to "it"
+    """
+    try:
+        x = array(QTEGRA_INTEGRATION_TIMES)
+        return x[argmin(abs(x - it))]
+    except TypeError:
+        return 1.0
+
+
+def calculate_radius(m_e, hv, mfield):
+    """
+    m_e= mass/charge
+    hv= accelerating voltage (V)
+    mfield= magnet field (H)
+    """
+    r = ((2 * m_e * hv) / mfield ** 2) ** 0.5
+
+    return r
 
 # ============= EOF =============================================
