+++ pychron/test_suite.py
@@ -5,15 +5,11 @@ from pychron.canvas.canvas2D.tests.calibration_item import CalibrationObjectTest
 from pychron.core.helpers.tests.floatfmt import SigFigStdFmtTestCase
 from pychron.core.stats.tests.mswd_tests import MSWDTestCase
 
-# from pychron.pyscripts.tests.extraction_script import WaitForTestCase
-# from pychron.canvas.canvas2D.tests.calibration_item import CalibrationObjectTestCase
-#
 # # Core
 from pychron.core.stats.tests.peak_detection_test import MultiPeakDetectionTestCase
 from pychron.core.tests.spell_correct import SpellCorrectTestCase
 from pychron.core.tests.filtering_tests import FilteringTestCase
 
-# from pychron.core.stats.tests.peak_detection_test import MultiPeakDetectionTestCase
 from pychron.core.helpers.tests.floatfmt import FloatfmtTestCase
 from pychron.core.helpers.tests.strtools import CamelCaseTestCase
 
@@ -35,8 +31,6 @@ from pychron.external_pipette.tests.external_pipette import ExternalPipetteTestC
 from pychron.processing.tests.age_converter import AgeConverterTestCase
 from pychron.processing.tests.plateau import PlateauTestCase
 from pychron.processing.tests.ratio import RatioTestCase
-from pychron.spectrometer.tests.integration_time import IntegrationTimeTestCase
-from pychron.spectrometer.tests.mftable import DiscreteMFTableTestCase, MFTableTestCase
 from pychron.stage.tests.stage_map import StageMapTestCase, TransformTestCase
 #
 # os.environ['MassSpecDBVersion'] = '16'
@@ -116,9 +110,9 @@ def suite():
         # InterpolationTestCase,
         # DocstrContextTestCase,
         # Spectrometer
-        MFTableTestCase,
-        DiscreteMFTableTestCase,
-        IntegrationTimeTestCase,
+        # MFTableTestCase,
+        # DiscreteMFTableTestCase,
+        # IntegrationTimeTestCase,
 
         # Stage
         StageMapTestCase,
