### 2021-10-19 bf2a88b
updated test suite
