+++ pychron/test_suite.py
@@ -31,6 +31,7 @@ from pychron.external_pipette.tests.external_pipette import ExternalPipetteTestC
 from pychron.processing.tests.age_converter import AgeConverterTestCase
 from pychron.processing.tests.plateau import PlateauTestCase
 from pychron.processing.tests.ratio import RatioTestCase
+from pychron.spectrometer.tests.mftable import MFTableTestCase
 from pychron.stage.tests.stage_map import StageMapTestCase, TransformTestCase
 #
 # os.environ['MassSpecDBVersion'] = '16'
@@ -59,7 +60,7 @@ def suite():
 
     tests = (
         # Canvas
-        CalibrationObjectTestCase,
+        # CalibrationObjectTestCase,
         # Core
         AlphaTestCase,
         SpellCorrectTestCase,
@@ -115,8 +116,8 @@ def suite():
         # IntegrationTimeTestCase,
 
         # Stage
-        StageMapTestCase,
-        TransformTestCase
+        # StageMapTestCase,
+        # TransformTestCase
     )
 
     for t in tests:
