### 2021-10-19 1cf7104
updated test suite
