+++ pychron/core/pychron_traits.py
@@ -55,11 +55,11 @@ class URLStr(BaseStr):
 class HostStr(BaseStr):
     def validate(self, obj, name, value):
         if (
-                not value
-                or value == "localhost"
-                or IPREGEX.match(value)
-                or URLREGEX.match(value)
-                or "\\" in value
+            not value
+            or value == "localhost"
+            or IPREGEX.match(value)
+            or URLREGEX.match(value)
+            or "\\" in value
         ):
 
             return value
@@ -129,7 +129,7 @@ class RestrictedStr(BaseStr):
             return value
 
 
-LOADNAME_REGEX = re.compile(r'^[a-zA-z]*(?<!O)\d+$')
+LOADNAME_REGEX = re.compile(r"^[a-zA-z]*(?<!O)\d+$")
 
 
 class LoadNameStr(BaseStr):
@@ -148,5 +148,6 @@ class BorderVGroup(VGroup):
 class BorderHGroup(HGroup):
     def _show_border_default(self):
         return True
+
 
 # ============= EOF =============================================
