+++ .github/workflows/unittests.yml
@@ -14,7 +14,7 @@ jobs:
     name:  Pychron Install via EDM (${{ matrix.python-version }}, ${{ matrix.os }})
     strategy:
       matrix:
-        os: [ macos-latest, windows-latest]
+        os: [ macos-latest]
     runs-on: ${{ matrix.os }}
     steps:
       - name: Checkout
