### 2021-10-19 e5a03c3
added device manager
