+++ pychron/hardware/turbo/hipace.py
@@ -23,12 +23,20 @@ http://mmrc.caltech.edu/Optical%20Furnace/Manuals/Pfeiffer%20TC%20110%20Elec.%20
 """
 
 from __future__ import absolute_import
+
+from traitsui.item import Item
+from traitsui.view import View
+
+from pychron.core.ui.lcd_editor import LCDEditor
 from pychron.hardware.core.communicators.serial_communicator import SerialCommunicator
 from pychron.hardware.core.core_device import CoreDevice
 import re
 
 BASE = r"^(?P<address>\d{3})(?P<action>\d{2})(?P<parameter>\d{2})(?P<datalength>\d{2})"
 RESPONSE_RE = re.compile(BASE)
+
+PARAMETERS = {'set_speed': 308,
+              'actual_speed': 309}
 
 
 def make_pattern(dl):
@@ -46,36 +54,28 @@ def calc_checksum(msg):
 
 
 class HiPace(CoreDevice):
+    scan_func = 'update'
+
+    def update(self):
+        self.read_set_speed()
+        self.read_actual_speed()
+
     def read_set_speed(self):
-        return self._read_parameter(308)
+        return self._read_parameter('set_speed')
 
     def read_actual_speed(self):
-        return self._read_parameter(309)
+        return self._read_parameter('actual_speed')
 
-    def _read_parameter(self, param):
-        resp = self.ask("read", param)
+    def _read_parameter(self, attr):
+        param = PARAMETERS[attr]
+        cmd = self._assemble("read", param)
+        resp = self.ask(cmd)
         data = self._get_data(resp)
         try:
-            return int(data)
+            v = int(data)
+            setattr(self, attr, v)
         except (ValueError, TypeError):
             pass
-
-    def _get_data(self, resp):
-        match = RESPONSE_RE.search(resp)
-        if match:
-            dl = match.group("datalength")
-            pattern = make_pattern(dl)
-            match = re.search(pattern, resp)
-            if match:
-                if check_checksum(resp, match.group("checksum")):
-                    data = match.group("data")
-                    return data
-
-
-class HiPaceCommunicator(SerialCommunicator):
-    def ask(self, action, parameter, data=None, *args, **kw):
-        cmd = self._assemble(action, parameter, data)
-        return super(HiPaceCommunicator, self).ask(cmd, *args, **kw)
 
     def _assemble(self, action, parameter, data):
         if action == "read":
@@ -93,7 +93,46 @@ class HiPaceCommunicator(SerialCommunicator):
             self.address, action, parameter, data_len, data
         )
 
-        return "{}{:03n}".format(msg, self._calc_checksum(msg))
+        return "{}{:03n}".format(msg, calc_checksum(msg))
+
+    def _get_data(self, resp):
+        match = RESPONSE_RE.search(resp)
+        if match:
+            dl = match.group("datalength")
+            pattern = make_pattern(dl)
+            match = re.search(pattern, resp)
+            if match:
+                if check_checksum(resp, match.group("checksum")):
+                    data = match.group("data")
+                    return data
+
+    def pump_view(self):
+        v = View(Item('setspeed', label='SetSpeed (308)', editor=LCDEditor(width=100, height=50)),
+                 Item('actualspeed', label='ActualSpeed (309)', editor=LCDEditor(width=100, height=50)))
+        return v
+
+# class HiPaceCommunicator(SerialCommunicator):
+#     def ask(self, action, parameter, data=None, *args, **kw):
+#         cmd = self._assemble(action, parameter, data)
+#         return super(HiPaceCommunicator, self).ask(cmd, *args, **kw)
+#
+#     def _assemble(self, action, parameter, data):
+#         if action == "read":
+#             # read parameter
+#             action = 0
[…]
