+++ pychron/experiment/automated_run/automated_run.py
@@ -315,10 +315,11 @@ class AutomatedRun(Loggable):
         new measurement interface for just sinking the data from a ring buffer
         """
         import csv
+
         spec = self.spectrometer_manager.spectrometer
         spec.set_data_pump_mode(1)
-        p, _ = unique_path2(paths.csv_data_dir, self.runid, extension='.csv')
-        with open(p, 'w') as rfile:
+        p, _ = unique_path2(paths.csv_data_dir, self.runid, extension=".csv")
+        with open(p, "w") as rfile:
             writer = csv.writer(rfile)
             ig = spec.sink_data(writer, n, delay)
 
