+++ pychron/extraction_line/pump_manager.py
@@ -33,4 +33,5 @@ class PumpManager(DeviceManager):
             )
             return 0
 
+
 # ============= EOF =============================================
