+++ .github/workflows/unittests.yml
@@ -1,13 +1,13 @@
 # This workflow will install Python dependencies, run tests and lint with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Pychron Install via EDM with pytest suite
+name: Pychron Install via EDM with unittest suite
 
 on:
   push:
-    branches: [ dev/dr, hot/dr, hot/dc]
+    branches: [ main, develop, dev/dr, hot/dr, hot/dc]
   pull_request:
-    branches: [ dev/dr , hot/dr, hot/dc]
+    branches: [ main, develop, dev/dr, hot/dr, hot/dc]
 
 jobs:
   build:
