+++ .github/workflows/install.yml
@@ -1,50 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: Python application
-
-on:
-  push:
-    branches: [ dev/dr ]
-  pull_request:
-    branches: [ dev/dr ]
-
-jobs:
-  job-1:
-    name: Pychron Install (${{ matrix.python-version }}, ${{ matrix.os }})
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-#        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-        os: [ "macos-latest"]
-        python-version: [ "3.9",]
-    steps:
-      - uses: conda-incubator/setup-miniconda@v2
-        with:
-          activate-environment: test
-          auto-activate-base: false
-          auto-update-conda: true
-          python-version: ${{ matrix.python-version }}
-      - uses: actions/checkout@v2
-#      - name: Set up Python 3.9
-#        uses: actions/setup-python@v2
-#        with:
-#          python-version: 3.9
-#      - name: Install swig 3
-#        run: |
-#          sudo apt-get install swig=3.0.\*
-      - name: Install Conda dependencies
-        run: |
-          conda install -n test -c conda-forge pip flake8 pytest numpy statsmodels scikit-learn PyYAML yaml traits traitsui pyface envisage sqlalchemy Reportlab lxml xlrd xlwt xlsxwriter requests keyring pyparsing pillow gitpython pytables pyproj pymysql certifi jinja2 swig=3 cython importlib_resources
-          $CONDA/envs/test/bin/pip3 install -e git+https://github.com/enthought/enable.git#egg=enable
-          $CONDA/envs/test/bin/pip3 install -e git+https://github.com/enthought/chaco.git#egg=chaco
-      - name: Lint with flake8
-        run: |
-          # stop the build if there are Python syntax errors or undefined names
-          $CONDA/envs/test/bin/flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          $CONDA/envs/test/bin/flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-      - name: Test with pytest
-        run: |
-          $CONDA/envs/test/bin/pytest
