+++ pychron/dvc/dvc_database.py
@@ -213,7 +213,7 @@ def exclude_invalid_analyses(q):
 
 def extract_devices_query(analysis_types, extract_devices, q):
     if extract_devices and (
-        "air" not in analysis_types and "cocktail" not in analysis_types
+            "air" not in analysis_types and "cocktail" not in analysis_types
     ):
         a = any(
             (
@@ -346,12 +346,12 @@ class DVCDatabase(DatabaseAdapter):
             sess.add(change)
 
     def find_references_by_load(
-        self,
-        load,
-        analysis_types,
-        extract_devices=None,
-        mass_spectrometers=None,
-        exclude_invalid=True,
+            self,
+            load,
+            analysis_types,
+            extract_devices=None,
+            mass_spectrometers=None,
+            exclude_invalid=True,
     ):
         with self.session_ctx() as sess:
             self.debug("----------- find references by load ------------")
@@ -380,14 +380,14 @@ class DVCDatabase(DatabaseAdapter):
             return records
 
     def find_references(
-        self,
-        times,
-        atypes,
-        hours=10,
-        exclude=None,
-        extract_devices=None,
-        mass_spectrometers=None,
-        exclude_invalid=True,
+            self,
+            times,
+            atypes,
+            hours=10,
+            exclude=None,
+            extract_devices=None,
+            mass_spectrometers=None,
+            exclude_invalid=True,
     ):
 
         with self.session_ctx():
@@ -699,7 +699,7 @@ class DVCDatabase(DatabaseAdapter):
             return self._add_item(a)
 
     def add_irradiation_level(
-        self, name, irradiation, holder, production_name, z=0, note=""
+            self, name, irradiation, holder, production_name, z=0, note=""
     ):
         with self.session_ctx():
             dblevel = self.get_irradiation_level(irradiation, name)
@@ -868,10 +868,10 @@ class DVCDatabase(DatabaseAdapter):
             q = q.distinct(IrradiationPositionTbl.id)
 
             comps = [
-                IrradiationPositionTbl.identifier.like("{}%".format(search_str)),
-                ProjectTbl.name == search_str,
-                ProjectTbl.id == search_str,
-            ] + self._fuzzy_sample_comps(search_str)
+                        IrradiationPositionTbl.identifier.like("{}%".format(search_str)),
+                        ProjectTbl.name == search_str,
+                        ProjectTbl.id == search_str,
+                    ] + self._fuzzy_sample_comps(search_str)
 
             f = or_(*comps)
             q = q.filter(f)
@@ -933,7 +933,7 @@ class DVCDatabase(DatabaseAdapter):
             return self._query_all(q)
 
     def get_sample_prep_steps(
-        self, worker, session, sample, project, material, grainsize
+            self, worker, session, sample, project, material, grainsize
     ):
         with self.session_ctx() as sess:
             q = sess.query(SamplePrepStepTbl)
@@ -969,6 +969,29 @@ class DVCDatabase(DatabaseAdapter):
             q = q.order_by(AnalysisTbl.timestamp.desc())
             q = q.limit(limit)
             return self._query_all(q)
+
+    def get_repository_analyses_date_range(self, repo):
+        with self.session_ctx() as sess:
+            r = self.get_repository(repo)
+            ts = {a.analysis.timestamp for a in r.repository_associations}
+            return min(ts), max(ts)
+
+    def get_repository_mass_spectrometers(self, repo):
+        with self.session_ctx() as sess:
+            r = self.get_repository(repo)
+            return {a.analysis.mass_spectrometer for a in r.repository_associations}
+
+    def get_repository_irradiations(self, repo):
+        with self.session_ctx() as sess:
+            q = sess.query(IrradiationTbl)
+            q = q.join(LevelTbl)
+            q = q.join(IrradiationPositionTbl)
[…]
