+++ pychron/entry/entry_views/repository_entry.py
@@ -14,9 +14,10 @@
 # limitations under the License.
 # ===============================================================================
 from __future__ import absolute_import
-from traits.api import Str, List
+from traits.api import Str, List, Enum
 from traitsui.api import VGroup, UItem, Item, EnumEditor
 
+from pychron.core.pychron_traits import BorderHGroup
 from pychron.core.ui.strings import SpacelessStr
 from pychron.entry.entry_views.entry import BaseEntry, OKButton, STYLESHEET
 
@@ -26,6 +27,11 @@ class RepositoryIdentifierEntry(BaseEntry):
     principal_investigator = Str
     principal_investigators = List
     value = SpacelessStr
+    readme = Str
+    license_template_name = Enum({'GNU General Public License v3.0': 'gpl-3.0',
+                                  'Apache 2.0': 'apache-2.0',
+                                  'Creative Commons Attribution 4.0': 'cc-by-4.0',
+                                  'Academic Free License v3.0': 'afl-3.0'})
 
     def _add_item(self):
         with self.dvc.session_ctx(use_parent_session=False):
@@ -44,12 +50,16 @@ class RepositoryIdentifierEntry(BaseEntry):
                 return
 
             ret = True
-            if not self.dvc.add_repository(self.value, self.principal_investigator):
+            if not self.dvc.add_repository(self.value, self.principal_investigator,
+                                           license_template=self.license_template_name,
+                                           ):
                 ret = False
                 if not self.confirmation_dialog(
                     'Could not add "{}". Try a different name?'.format(self.value)
                 ):
                     ret = None
+            else:
+                self.dvc.add_readme(self.value, self.readme)
 
         return ret
 
@@ -62,6 +72,8 @@ class RepositoryIdentifierEntry(BaseEntry):
                 "principal_investigator",
                 editor=EnumEditor(name="principal_investigators"),
             ),
+            BorderHGroup(UItem("readme", style="custom"), label="ReadMe"),
+            Item("license_template_name", label="License"),
             UItem("error_message", style="readonly", style_sheet=STYLESHEET),
         )
         buttons = [OKButton(), "Cancel"]
@@ -69,5 +81,8 @@ class RepositoryIdentifierEntry(BaseEntry):
             a, width=400, title="Add {}".format(self.tag), buttons=buttons
         )
 
+if __name__ == '__main__':
+    d = RepositoryIdentifierEntry()
+    d.configure_traits()
 
 # ============= EOF =============================================

[…]
