+++ pychron/dvc/dvc.py
@@ -153,14 +153,14 @@ class DVC(Loggable):
             return True
 
     def fix_identifier(
-            self,
-            src_uuid,
-            src_id,
-            dest_id,
-            repo_identifier,
-            dest_identifier,
-            dest_aliquot,
-            dest_step,
+        self,
+        src_uuid,
+        src_id,
+        dest_id,
+        repo_identifier,
+        dest_identifier,
+        dest_aliquot,
+        dest_step,
     ):
         self.info("converting {} to {}".format(src_id, dest_id))
         err = self.db.map_runid(src_id, dest_id)
@@ -210,13 +210,13 @@ class DVC(Loggable):
             os.remove(sp)
 
             for modifier in (
-                    "baselines",
-                    "blanks",
-                    "extraction",
-                    "intercepts",
-                    "icfactors",
-                    "peakcenter",
-                    ".data",
+                "baselines",
+                "blanks",
+                "extraction",
+                "intercepts",
+                "icfactors",
+                "peakcenter",
+                ".data",
             ):
                 sp = analysis_path(
                     src_id, repo_identifier, modifier=modifier, root=root
@@ -246,8 +246,8 @@ class DVC(Loggable):
             return
 
         if not self.confirmation_dialog(
-                "Are you sure you want to generate current values for the entire database? "
-                "This could take a while!"
+            "Are you sure you want to generate current values for the entire database? "
+            "This could take a while!"
         ):
             return
 
@@ -260,7 +260,7 @@ class DVC(Loggable):
 
         def chunks(l, n):
             for i in range(0, len(l), n):
-                yield l[i: i + n]
+                yield l[i : i + n]
 
         def func(ai, prog, i, n):
             if prog:
@@ -308,11 +308,11 @@ class DVC(Loggable):
         with db.session_ctx():
             for repo in db.get_repositories():
                 if repo.name in (
-                        "JIRSandbox",
-                        "REEFenite",
-                        "Henry01184",
-                        "FractionatedRes",
-                        "PowerZPattern",
+                    "JIRSandbox",
+                    "REEFenite",
+                    "Henry01184",
+                    "FractionatedRes",
+                    "PowerZPattern",
                 ):
                     continue
                 self.debug("Updating currents for {}".format(repo.name))
@@ -377,24 +377,26 @@ class DVC(Loggable):
         mss = self.db.get_repository_mass_spectrometers(repo)
         mss = [mi.capitalize() for mi in mss]
 
-        self.debug('{} {}'.format(low, high))
+        self.debug("{} {}".format(low, high))
         repos = []
         if low.month <= 6:
             year = low.year
             year = str(year)[-2:]
-            repos.extend(['{}_air{}0'.format(mi, year) for mi in mss])
-            repos.extend(['{}_blank{}0'.format(mi, year) for mi in mss])
-            repos.extend(['{}_cocktail{}0'.format(mi, year) for mi in mss])
+            repos.extend(["{}_air{}0".format(mi, year) for mi in mss])
+            repos.extend(["{}_blank{}0".format(mi, year) for mi in mss])
+            repos.extend(["{}_cocktail{}0".format(mi, year) for mi in mss])
 
         if high.month >= 6 or low.month >= 6:
             year = high.year
             year = str(year)[-2:]
-            repos.extend(['{}_air{}1'.format(mi, year) for mi in mss])
-            repos.extend(['{}_blank{}1'.format(mi, year) for mi in mss])
-            repos.extend(['{}_cocktail{}1'.format(mi, year) for mi in mss])
+            repos.extend(["{}_air{}1".format(mi, year) for mi in mss])
+            repos.extend(["{}_blank{}1".format(mi, year) for mi in mss])
+            repos.extend(["{}_cocktail{}1".format(mi, year) for mi in mss])
 
[…]
