+++ pychron/dvc/dvc_database.py
@@ -213,7 +213,7 @@ def exclude_invalid_analyses(q):
 
 def extract_devices_query(analysis_types, extract_devices, q):
     if extract_devices and (
-            "air" not in analysis_types and "cocktail" not in analysis_types
+        "air" not in analysis_types and "cocktail" not in analysis_types
     ):
         a = any(
             (
@@ -346,12 +346,12 @@ class DVCDatabase(DatabaseAdapter):
             sess.add(change)
 
     def find_references_by_load(
-            self,
-            load,
-            analysis_types,
-            extract_devices=None,
-            mass_spectrometers=None,
-            exclude_invalid=True,
+        self,
+        load,
+        analysis_types,
+        extract_devices=None,
+        mass_spectrometers=None,
+        exclude_invalid=True,
     ):
         with self.session_ctx() as sess:
             self.debug("----------- find references by load ------------")
@@ -380,14 +380,14 @@ class DVCDatabase(DatabaseAdapter):
             return records
 
     def find_references(
-            self,
-            times,
-            atypes,
-            hours=10,
-            exclude=None,
-            extract_devices=None,
-            mass_spectrometers=None,
-            exclude_invalid=True,
+        self,
+        times,
+        atypes,
+        hours=10,
+        exclude=None,
+        extract_devices=None,
+        mass_spectrometers=None,
+        exclude_invalid=True,
     ):
 
         with self.session_ctx():
@@ -699,7 +699,7 @@ class DVCDatabase(DatabaseAdapter):
             return self._add_item(a)
 
     def add_irradiation_level(
-            self, name, irradiation, holder, production_name, z=0, note=""
+        self, name, irradiation, holder, production_name, z=0, note=""
     ):
         with self.session_ctx():
             dblevel = self.get_irradiation_level(irradiation, name)
@@ -868,10 +868,10 @@ class DVCDatabase(DatabaseAdapter):
             q = q.distinct(IrradiationPositionTbl.id)
 
             comps = [
-                        IrradiationPositionTbl.identifier.like("{}%".format(search_str)),
-                        ProjectTbl.name == search_str,
-                        ProjectTbl.id == search_str,
-                    ] + self._fuzzy_sample_comps(search_str)
+                IrradiationPositionTbl.identifier.like("{}%".format(search_str)),
+                ProjectTbl.name == search_str,
+                ProjectTbl.id == search_str,
+            ] + self._fuzzy_sample_comps(search_str)
 
             f = or_(*comps)
             q = q.filter(f)
@@ -933,7 +933,7 @@ class DVCDatabase(DatabaseAdapter):
             return self._query_all(q)
 
     def get_sample_prep_steps(
-            self, worker, session, sample, project, material, grainsize
+        self, worker, session, sample, project, material, grainsize
     ):
         with self.session_ctx() as sess:
             q = sess.query(SamplePrepStepTbl)
@@ -1150,14 +1150,14 @@ class DVCDatabase(DatabaseAdapter):
             return q.count()
 
     def get_analyses_advanced(
-            self,
-            advanced_filter,
-            uuids=None,
-            identifiers=None,
-            return_labnumbers=False,
-            include_invalid=False,
-            limit=None,
-            verbose=False,
+        self,
+        advanced_filter,
+        uuids=None,
+        identifiers=None,
+        return_labnumbers=False,
+        include_invalid=False,
+        limit=None,
+        verbose=False,
     ):
         with self.session_ctx() as sess:
             if return_labnumbers:
@@ -1217,7 +1217,7 @@ class DVCDatabase(DatabaseAdapter):
             return int(ret[0]) if ret else 0
 
     def get_last_nhours_analyses(
-            self, n, return_limits=False, mass_spectrometers=None, analysis_types=None
[…]
