+++ pychron/entry/entry_views/repository_entry.py
@@ -28,10 +28,14 @@ class RepositoryIdentifierEntry(BaseEntry):
     principal_investigators = List
     value = SpacelessStr
     readme = Str
-    license_template_name = Enum({'GNU General Public License v3.0': 'gpl-3.0',
-                                  'Apache 2.0': 'apache-2.0',
-                                  'Creative Commons Attribution 4.0': 'cc-by-4.0',
-                                  'Academic Free License v3.0': 'afl-3.0'})
+    license_template_name = Enum(
+        {
+            "GNU General Public License v3.0": "gpl-3.0",
+            "Apache 2.0": "apache-2.0",
+            "Creative Commons Attribution 4.0": "cc-by-4.0",
+            "Academic Free License v3.0": "afl-3.0",
+        }
+    )
 
     def _add_item(self):
         with self.dvc.session_ctx(use_parent_session=False):
@@ -50,9 +54,11 @@ class RepositoryIdentifierEntry(BaseEntry):
                 return
 
             ret = True
-            if not self.dvc.add_repository(self.value, self.principal_investigator,
-                                           license_template=self.license_template_name,
-                                           ):
+            if not self.dvc.add_repository(
+                self.value,
+                self.principal_investigator,
+                license_template=self.license_template_name,
+            ):
                 ret = False
                 if not self.confirmation_dialog(
                     'Could not add "{}". Try a different name?'.format(self.value)
@@ -81,7 +87,8 @@ class RepositoryIdentifierEntry(BaseEntry):
             a, width=400, title="Add {}".format(self.tag), buttons=buttons
         )
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     d = RepositoryIdentifierEntry()
     d.configure_traits()
 

[…]
