### 2021-11-02 60724cc
changed login to use yaml instead of pickle
