### 2021-11-02 6ccbb77
non fatal update launch
