### 2021-11-02 1636477
fixed read_temperature
