+++ pychron/hardware/turbo/hipace.py
@@ -26,24 +26,34 @@ from __future__ import absolute_import
 
 from traitsui.item import Item
 from traitsui.view import View
+from traits.api import Float, Int
 
 from pychron.core.ui.lcd_editor import LCDEditor
 from pychron.hardware.core.communicators.serial_communicator import SerialCommunicator
 from pychron.hardware.core.core_device import CoreDevice
+from pychron.core.pychron_traits import BorderVGroup
+
 import re
 
-BASE = r"^(?P<address>\d{3})(?P<action>\d{2})(?P<parameter>\d{2})(?P<datalength>\d{2})"
+BASE = r"^(?P<address>\d{3})(?P<action>\d{2})(?P<parameter>\d{3})(?P<datalength>\d{2})"
 RESPONSE_RE = re.compile(BASE)
 
-PARAMETERS = {"set_speed": 308, "actual_speed": 309}
+def dt2(v):
+    return float(v)/100.
+
+PARAMETERS = {"set_speed": (308, int), 
+"actual_speed": (309, int), 
+'drive_current': (310, dt2)}
 
 
 def make_pattern(dl):
-    return "{}(?P<data>\d{{{}}})(?P<checksum>\d{3}".format(BASE, dl)
+    return "{}(?P<data>\d{{{}}})(?P<checksum>\d{{3}})".format(BASE, dl)
 
 
 def check_checksum(resp, chksum):
     r = resp[:-3]
+
+    print('calc {} {}  {}'.format(calc_checksum(r) == int(chksum), r, calc_checksum(r), int(chksum)))
     return calc_checksum(r) == int(chksum)
 
 
@@ -54,10 +64,19 @@ def calc_checksum(msg):
 
 class HiPace(CoreDevice):
     scan_func = "update"
-
+    
+    set_speed = Int
+    actual_speed = Int
+    drive_current = Float
+    address = 1
     def update(self):
+        self.debug('update')
         self.read_set_speed()
         self.read_actual_speed()
+        self.read_drive_current()
+
+    def read_drive_current(self):
+        return self._read_parameter('drive_current')
 
     def read_set_speed(self):
         return self._read_parameter("set_speed")
@@ -66,17 +85,19 @@ class HiPace(CoreDevice):
         return self._read_parameter("actual_speed")
 
     def _read_parameter(self, attr):
-        param = PARAMETERS[attr]
+        param, datatype = PARAMETERS[attr]
         cmd = self._assemble("read", param)
-        resp = self.ask(cmd)
+        resp = self.ask(cmd, verbose=True)
         data = self._get_data(resp)
+        self.debug('set data {}={}'.format(attr,data))
         try:
-            v = int(data)
+            v = datatype(data)
+
             setattr(self, attr, v)
         except (ValueError, TypeError):
             pass
 
-    def _assemble(self, action, parameter, data):
+    def _assemble(self, action, parameter, data=None):
         if action == "read":
             # read parameter
             action = 0
@@ -100,6 +121,7 @@ class HiPace(CoreDevice):
             dl = match.group("datalength")
             pattern = make_pattern(dl)
             match = re.search(pattern, resp)
+            self.debug('{} {}'.format(match, pattern))
             if match:
                 if check_checksum(resp, match.group("checksum")):
                     data = match.group("data")
@@ -107,16 +129,21 @@ class HiPace(CoreDevice):
 
     def pump_view(self):
         v = View(
-            Item(
-                "setspeed",
-                label="SetSpeed (308)",
+            BorderVGroup(Item(
+                "set_speed",
+                label="SetSpeed hz (308)",
                 editor=LCDEditor(width=100, height=50),
             ),
             Item(
-                "actualspeed",
-                label="ActualSpeed (309)",
+                "actual_speed",
[…]
