+++ pychron/hardware/turbo/hipace.py
@@ -22,14 +22,12 @@ TC 110 Electronic Drive Unit
 http://mmrc.caltech.edu/Optical%20Furnace/Manuals/Pfeiffer%20TC%20110%20Elec.%20Drive%20Unt.pdf
 """
 
-from __future__ import absolute_import
+from traitsui.api import View, UItem, HGroup, ButtonEditor, Item
+from traits.api import Float, Int, Bool, Event, Property
 
-from traitsui.item import Item
-from traitsui.view import View
-from traits.api import Float, Int
-
+from pychron.core.helpers.strtools import to_bool
 from pychron.core.ui.lcd_editor import LCDEditor
-from pychron.hardware.core.communicators.serial_communicator import SerialCommunicator
+from pychron.hardware.OnOffMixin import OnOffMixin
 from pychron.hardware.core.core_device import CoreDevice
 from pychron.core.pychron_traits import BorderVGroup
 
@@ -47,6 +45,8 @@ PARAMETERS = {
     "set_speed": (308, int),
     "actual_speed": (309, int),
     "drive_current": (310, dt2),
+    "motor_pump": (23, to_bool),
+    "standby": (2, to_bool)
 }
 
 
@@ -70,19 +70,53 @@ def calc_checksum(msg):
     return checksum % 256
 
 
-class HiPace(CoreDevice):
+class HiPace(CoreDevice, OnOffMixin):
     scan_func = "update"
 
+    standby = Bool
+    motor_pump = Bool
     set_speed = Int
     actual_speed = Int
     drive_current = Float
     address = 1
+
+    standby_button = Event
+    standby_label = Property(depends_on="standby_state")
+
+    onoff_state_name = 'motor_pump'
+
+    def _get_standby_label(self):
+        return "Standby On" if self.standby else "Standby Off"
+
+    def _standby_button_fired(self):
+        self.standby_state = not self.standby
+        self.debug("set state = {}".format(self.standby))
+        self.set_standby(self.standby)
+
+    def _onoff_button_fired(self):
+        self.onoff_state = not self.onoff_state
+        self.debug("set state = {}".format(self.onoff_state))
+        self.set_active(self.onoff_state)
 
     def update(self):
         self.debug("update")
         self.read_set_speed()
         self.read_actual_speed()
         self.read_drive_current()
+        self.read_state()
+        self.read_standby_state()
+
+    def set_standby(self, state):
+        self._set_parameter("standby", int(state))
+
+    def set_active(self, state):
+        self._set_parameter("motor_pump", int(state))
+
+    def read_state(self):
+        return self._read_parameter("motor_pump")
+
+    def read_standby_state(self):
+        return self._read_parameter("standby")
 
     def read_drive_current(self):
         return self._read_parameter("drive_current")
@@ -93,12 +127,23 @@ class HiPace(CoreDevice):
     def read_actual_speed(self):
         return self._read_parameter("actual_speed")
 
-    def _read_parameter(self, attr):
+    def _set_parameter(self, attr, value, verbose=True):
+        param, datatype = PARAMETERS[attr]
+        cmd = self._assemble("set", param, value)
+        resp = self.ask(cmd, verbose=verbose)
+        raw = self._get_data(resp)
+        data = datatype(raw)
+        if verbose:
+            self.debug("set parameter data {}={},  raw={}".format(attr, data, raw))
+        return data
+
+    def _read_parameter(self, attr, verbose=False):
         param, datatype = PARAMETERS[attr]
         cmd = self._assemble("read", param)
-        resp = self.ask(cmd, verbose=True)
+        resp = self.ask(cmd, verbose=verbose)
         data = self._get_data(resp)
-        self.debug("set data {}={}".format(attr, data))
+        if verbose:
+            self.debug("read parameter data {}={}".format(attr, data))
         try:
[…]
