### 2021-11-05 9b73d56
added isochron calculations to spec/iso
