+++ pychron/database/core/database_adapter.py
@@ -78,14 +78,15 @@ class SessionCTX(object):
         self._psession = None
 
     def __enter__(self):
-        if self._use_parent_session:
-            self._parent.create_session()
-            return self._parent.session
-        else:
-            self._psession = self._parent.session
-            self._session = self._parent.session_factory()
-            self._parent.session = self._session
-            return self._session
+        if self._parent:
+            if self._use_parent_session:
+                self._parent.create_session()
+                return self._parent.session
+            else:
+                self._psession = self._parent.session
+                self._session = self._parent.session_factory()
+                self._parent.session = self._session
+                return self._session
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         if self._session:
