+++ pychron/options/views/composite_views.py
@@ -16,7 +16,7 @@
 from pychron.options.options import SubOptions, TitleSubOptions, GroupSubOptions
 from pychron.options.views.isochron_views import (
     InverseIsochronMainOptions,
-    InverseIsochronAppearance,
+    InverseIsochronAppearance, InverseIsochronCalculationOptions,
 )
 from pychron.options.views.spectrum_views import (
     SpectrumMainOptions,
@@ -47,6 +47,7 @@ VIEWS = {
     "calculations(spec.)": CalculationSubOptions,
     "appearance(spec.)": SpectrumAppearance,
     "appearance(iso.)": InverseIsochronAppearance,
+    "calculations(iso.)": InverseIsochronCalculationOptions,
     "layout": LayoutSubOptions,
     "title": TitleSubOptions,
     GROUPS.lower(): GroupSubOptions,
