+++ pychron/options/views/composite_views.py
@@ -16,7 +16,8 @@
 from pychron.options.options import SubOptions, TitleSubOptions, GroupSubOptions
 from pychron.options.views.isochron_views import (
     InverseIsochronMainOptions,
-    InverseIsochronAppearance, InverseIsochronCalculationOptions,
+    InverseIsochronAppearance,
+    InverseIsochronCalculationOptions,
 )
 from pychron.options.views.spectrum_views import (
     SpectrumMainOptions,
