### 2021-11-08 89b1dc3
addressed spectrum overlay nsigma. added fixed height #2038
