+++ pychron/pipeline/plot/plotter/arar_figure.py
@@ -123,7 +123,8 @@ class BaseArArFigure(SelectionFigure):
 
         graph = self.graph
 
-        vertical_resize = not all([p.height for p in plots])
+        print([p.height for p in plots])
+        vertical_resize = not all([p.height for p in plots[:-1]])
 
         graph.vertical_resize = vertical_resize
         graph.clear_has_title()
@@ -133,13 +134,41 @@ class BaseArArFigure(SelectionFigure):
             title = self.options.title
 
         nplots = len(plots)
+
+        fw, fh = self.options.layout.fixed_width, self.options.layout.fixed_height
+
+        oheights = sum([po.height for po in plots[1:]])
+
         for i, po in enumerate(plots):
             kw = {"ytitle": po.name}
             if plot_dict:
                 kw.update(plot_dict)
 
-            if po.height:
-                kw["bounds"] = [50, po.height]
+            if fw:
+                r = ''
+                if fh:
+                    if i == 0 and not po.height:
+                        height = fh - oheights
+                    else:
+                        height = po.height
+                else:
+                    height = po.height
+                    r = 'v'
+
+                kw['bounds'] = [fw, height]
+                kw['resizable'] = r
+            elif fh:
+                kw['resizable'] = 'h'
+                if i == 0 and not po.height:
+                    height = fh - oheights
+                else:
+                    height = po.height
+
+                kw['bounds'] = [50, height]
+
+            # if po.height:
+            #     kw["bounds"] = [, po.height]
+            #     kw['resizable'] = ''
 
             # if self.options.layout.fixed_width:
             #     kw['bounds'] = [self.options.layout.fixed_width, kw['bounds'][1]]
@@ -302,7 +331,7 @@ class BaseArArFigure(SelectionFigure):
         return x.timestamp or 0
 
     def _unpack_attr(
-        self, attr, scalar=1, exclude_omit=False, nonsorted=False, ans=None
+            self, attr, scalar=1, exclude_omit=False, nonsorted=False, ans=None
     ):
         if ans is None:
             ans = self.sorted_analyses
@@ -502,7 +531,7 @@ class BaseArArFigure(SelectionFigure):
         scatter.underlays.append(ov)
 
     def _add_error_bars(
-        self, scatter, errors, axis, nsigma, line_width=1, end_caps=True, visible=True
+            self, scatter, errors, axis, nsigma, line_width=1, end_caps=True, visible=True
     ):
         ebo = ErrorBarOverlay(
             component=scatter,
@@ -518,18 +547,18 @@ class BaseArArFigure(SelectionFigure):
         return ebo
 
     def _add_scatter_inspector(
-        self,
-        scatter,
-        inspector=None,
-        add_tool=True,
-        add_selection=True,
-        value_format=None,
-        additional_info=None,
-        index_tag=None,
-        index_attr=None,
-        convert_index=None,
-        items=None,
-        update_meta_func=None,
+            self,
+            scatter,
+            inspector=None,
+            add_tool=True,
+            add_selection=True,
+            value_format=None,
+            additional_info=None,
+            index_tag=None,
+            index_attr=None,
+            convert_index=None,
+            items=None,
+            update_meta_func=None,
     ):
         if add_tool:
             # broadcaster = BroadcasterTool()
@@ -537,12 +566,10 @@ class BaseArArFigure(SelectionFigure):
 
             if inspector is None:
                 if value_format is None:
-
                     def value_format(x):
                         return "{:0.5f}".format(x)
 
                 if convert_index is None:
-
                     def convert_index(x):
                         return "{:0.3f}".format(x)
 
@@ -617,15 +644,15 @@ class BaseArArFigure(SelectionFigure):
[…]
