+++ pychron/pipeline/plot/plotter/arar_figure.py
@@ -145,7 +145,7 @@ class BaseArArFigure(SelectionFigure):
                 kw.update(plot_dict)
 
             if fw:
-                r = ''
+                r = ""
                 if fh:
                     if i == 0 and not po.height:
                         height = fh - oheights
@@ -153,18 +153,18 @@ class BaseArArFigure(SelectionFigure):
                         height = po.height
                 else:
                     height = po.height
-                    r = 'v'
+                    r = "v"
 
-                kw['bounds'] = [fw, height]
-                kw['resizable'] = r
+                kw["bounds"] = [fw, height]
+                kw["resizable"] = r
             elif fh:
-                kw['resizable'] = 'h'
+                kw["resizable"] = "h"
                 if i == 0 and not po.height:
                     height = fh - oheights
                 else:
                     height = po.height
 
-                kw['bounds'] = [50, height]
+                kw["bounds"] = [50, height]
 
             # if po.height:
             #     kw["bounds"] = [, po.height]
@@ -331,7 +331,7 @@ class BaseArArFigure(SelectionFigure):
         return x.timestamp or 0
 
     def _unpack_attr(
-            self, attr, scalar=1, exclude_omit=False, nonsorted=False, ans=None
+        self, attr, scalar=1, exclude_omit=False, nonsorted=False, ans=None
     ):
         if ans is None:
             ans = self.sorted_analyses
@@ -531,7 +531,7 @@ class BaseArArFigure(SelectionFigure):
         scatter.underlays.append(ov)
 
     def _add_error_bars(
-            self, scatter, errors, axis, nsigma, line_width=1, end_caps=True, visible=True
+        self, scatter, errors, axis, nsigma, line_width=1, end_caps=True, visible=True
     ):
         ebo = ErrorBarOverlay(
             component=scatter,
@@ -547,18 +547,18 @@ class BaseArArFigure(SelectionFigure):
         return ebo
 
     def _add_scatter_inspector(
-            self,
-            scatter,
-            inspector=None,
-            add_tool=True,
-            add_selection=True,
-            value_format=None,
-            additional_info=None,
-            index_tag=None,
-            index_attr=None,
-            convert_index=None,
-            items=None,
-            update_meta_func=None,
+        self,
+        scatter,
+        inspector=None,
+        add_tool=True,
+        add_selection=True,
+        value_format=None,
+        additional_info=None,
+        index_tag=None,
+        index_attr=None,
+        convert_index=None,
+        items=None,
+        update_meta_func=None,
     ):
         if add_tool:
             # broadcaster = BroadcasterTool()
@@ -566,10 +566,12 @@ class BaseArArFigure(SelectionFigure):
 
             if inspector is None:
                 if value_format is None:
+
                     def value_format(x):
                         return "{:0.5f}".format(x)
 
                 if convert_index is None:
+
                     def convert_index(x):
                         return "{:0.3f}".format(x)
 
@@ -644,15 +646,15 @@ class BaseArArFigure(SelectionFigure):
         plot.tools.append(OverlayMoveTool(component=ov))
 
     def _add_data_label(
-            self,
-            s,
-            text,
-            point,
-            bgcolor="transparent",
-            label_position="top right",
-            color=None,
-            append=True,
-            **kw
+        self,
+        s,
+        text,
+        point,
+        bgcolor="transparent",
+        label_position="top right",
+        color=None,
+        append=True,
+        **kw
     ):
         if color is None:
             color = s.color
@@ -683,17 +685,17 @@ class BaseArArFigure(SelectionFigure):
         return label
 
     def _build_label_text(
-            self,
-            x,
-            we,
-            n,
[…]
