+++ pychron/experiment/tasks/experiment_actions.py
@@ -354,4 +354,30 @@ class MeltingPointCalibrationAction(UIAction):
         mpc.edit_traits()
 
 
+class AddExperimentNoteAction(UIAction):
+    name = 'Add Experiment Note'
+    image = icon("insert-comment")
+
+    def perform(self, event):
+        app = event.task.window.application
+        remote = app.preferences.get("pychron.general.remote")
+        if not remote:
+            information(
+                event.task.window.control,
+                'Please set an "Laboratory Repo" in General Preferences',
+            )
+            return
+
+        from pychron.experiment.labbook.expeirment_note import ExperimentNote
+
+        from pychron.git.hosts import IGitHost
+        service = app.get_service(IGitHost)
+        m = ExperimentNote()
+
+        m.set_labels(service.get_labels(remote))
+        info = m.edit_traits()
+        if info.result:
+            if service.post_issue(remote, m.to_issue()):
+                information(None, 'Note added!')
+
 # ============= EOF ====================================
