+++ pychron/git/hosts/__init__.py
@@ -304,4 +304,5 @@ class GitHostService(BaseGitHostService):
         #
         # return {"Authorization": auth}
 
+
 # ============= EOF =============================================
