### 2021-11-15 1f66f1f
co2 ethernet hot fixes
