### 2021-11-16 166683b
updated find flux
