### 2021-11-19 4114d0a
added rise rate timestamp and persistence
