+++ pychron/pipeline/tables/xlsx_table_writer.py
@@ -374,7 +374,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         )
 
     def _intercalibration_columns(
-            self, columns, detectors, ic_visible=True, disc_visible=True
+        self, columns, detectors, ic_visible=True, disc_visible=True
     ):
         disc = [
             SigFigColumn(
@@ -409,9 +409,9 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
     def _signal_columns(self, columns, ibit, bkbit):
         isos = (("Ar", 40), ("Ar", 39), ("Ar", 38), ("Ar", 37), ("Ar", 36))
         for bit, tag in (
-                (True, "disc_ic_corrected"),
-                (ibit, "intercept"),
-                (bkbit, "blank"),
+            (True, "disc_ic_corrected"),
+            (ibit, "intercept"),
+            (bkbit, "blank"),
         ):
             cols = [
                 c
@@ -524,14 +524,15 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         fmt = "correction"
 
         cols = []
-        for am, bm, e in [(40, 39, "K"),
-                          (38, 38, "K"),
-                          (37, 39, "K"),
-                          (39, 37, "Ca"),
-                          (38, 37, "Ca"),
-                          (36, 37, "Ca"),
-                          (36, 38, "Cl")
-                          ]:
+        for am, bm, e in [
+            (40, 39, "K"),
+            (38, 38, "K"),
+            (37, 39, "K"),
+            (39, 37, "Ca"),
+            (38, 37, "Ca"),
+            (36, 37, "Ca"),
+            (36, 38, "Cl"),
+        ]:
             attr = "{}{}{}".format(e, am, bm)
             c = Column(
                 label=(
@@ -642,8 +643,8 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         def get_age_error(attr):
             def f(ag, *args):
                 return (
-                        ag.scaled_age(std_dev(getattr(ag, attr)), opt.age_units)
-                        * opt.summary_age_nsigma
+                    ag.scaled_age(std_dev(getattr(ag, attr)), opt.age_units)
+                    * opt.summary_age_nsigma
                 )
 
             return f
@@ -786,11 +787,11 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
 
         # hide extra age columns
         for hidden in (
-                "WeightedMeanAge",
-                "ArithmeticMeanAge",
-                "IsochronAge",
-                "PlateauAge",
-                "IntegratedAge",
+            "WeightedMeanAge",
+            "ArithmeticMeanAge",
+            "IsochronAge",
+            "PlateauAge",
+            "IntegratedAge",
         ):
             hc = next((i for i, c in enumerate(cols) if c.label == hidden), None)
             if hc is not None:
@@ -1060,8 +1061,8 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
                     args = []
                     for cii in ci:
                         for reg, fmt in (
-                                (supreg, self._superscript),
-                                (subreg, self._subscript),
+                            (supreg, self._superscript),
+                            (subreg, self._subscript),
                         ):
                             m = reg.match(cii)
                             if m:
@@ -1188,7 +1189,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
         return fn
 
     def _make_analysis(
-            self, sh, cols, item, is_last=False, is_plateau_step=None, cum=""
+        self, sh, cols, item, is_last=False, is_plateau_step=None, cum=""
     ):
         row = self._current_row
 
@@ -1553,6 +1554,7 @@ class XLSXAnalysisTableWriter(BaseTableWriter):
[…]
