### 2021-11-30 1370570
handle queue/factory sync exceptions
