### 2021-11-30 2252a4c
fixed spectrum aux plots
