+++ pychron/options/views/regression_views.py
@@ -20,11 +20,13 @@ from traitsui.api import View, Item
 
 class RegressionMainOptions(SubOptions):
     def traits_view(self):
-        v = View(Item("regressor"),
-                 Item('xtitle', label='X title'),
-                 Item('ytitle', label='Y title'),
-                 Item('regression_color', label='Regression Line'),
-                 Item('regression_width', label='Line wt.'))
+        v = View(
+            Item("regressor"),
+            Item("xtitle", label="X title"),
+            Item("ytitle", label="Y title"),
+            Item("regression_color", label="Regression Line"),
+            Item("regression_width", label="Line wt."),
+        )
 
         return v
 
