+++ pychron/entry/legacy/nmgrl/importer/irradiation.py
@@ -1,0 +1,204 @@
+# ===============================================================================
+# Copyright 2020 ross
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ===============================================================================
+import os
+import re
+from datetime import datetime, time
+
+from pychron.dvc.dvc import DVC
+from pychron.dvc.meta_object import Production
+from pychron.entry.editors.chronology import IrradiationDosage
+from pychron.entry.editors.production import IrradiationProduction
+from pychron.entry.legacy.nmgrl.importer import BaseImporter
+from pychron.entry.legacy.util import CSVHeader, get_dvc
+
+NMRE_NOLEVEL = re.compile(r'^NM-\d{3}$')
+NMRE_LEVEL = re.compile(r'^NM-\d{3}[A-Z]$')
+
+
+class IrradiationImporter(BaseImporter):
+    def do_import(self, dbsam, run, j, irrad, level, position):
+        print('Irradiation doing import run: {}'.format(run.runid))
+
+        # get the irradiation info
+        for c in self._cache:
+            if c[0] == irrad and c[1] == level:
+                ir = c
+                break
+        else:
+            for c in self._cache:
+                if c[0] == irrad and c[1] == 'ALL':
+                    ir = c
+                    break
+            else:
+                print('failed to find irradiation info')
+                return
+
+        print('irradiation: {}{}, {}, dbsam={}'.format(irrad, level, ir, dbsam))
+
+        dvc = self._dvc
+        if dvc:
+            _, _, prod, doses = ir
+
+            # import the irradiation if doesn't already exist
+            # add irradiation
+            dvc.add_irradiation(irrad, doses=doses)
+
+            # add production to irradiation
+            dvc.add_production(irrad, prod.name, prod)
+
+            # add level
+            dvc.add_irradiation_level(level, irrad, 'LegacyHolder', prod.name)
+
+            # add position to level
+            identifier = run.identifier
+            dvc.add_irradiation_position(irrad, level, position, identifier, sample=dbsam)
+
+            # set j for position
+            e, mj, me = 0, 0, 0
+            print('updating flux {}{} {} {} {}'.format(irrad, level, position, identifier, j))
+            dvc.update_flux(irrad, level, position, identifier, j, e, mj, me)
+        return True
+
+    # private
+    def _load(self):
+        p = '../tests/data/IrradiationData.csv'
+        skipping = []
+        with open(p, 'r') as rfile:
+            header = None
+            for line in rfile:
+                row = line.strip().split(',')
+                if header is None:
+                    header = CSVHeader(row)
+                else:
+                    err = self.process_row(header, row)
+                    if isinstance(err, str):
+                        skipping.append('{},{}'.format(err, line))
+                    elif err:
+                        self._cache.append(err)
+
+        op = 'skipping_irradiations.csv'
+        with open(op, 'w') as wfile:
+            wfile.write(''.join(skipping))
+
+    def process_row(self, header, r):
+        # only import NM- irradiations
+        irradname = header.get(r, 'Irradiation')
+        if NMRE_NOLEVEL.match(irradname):
+            level = 'ALL'
[…]
