+++ pychron/entry/legacy/nmgrl/tests/binary_extractor.py
@@ -1,0 +1,299 @@
+from __future__ import absolute_import
+from pychron.entry.loaders.mass_spec_binary_extractor import MassSpecBinaryExtractor
+
+__author__ = "ross"
+
+import unittest
+
+from pychron.entry.legacy.nmgrl.mass_spec_binary_extractor import MassSpecBinaryExtractor
+
+
+class Expected(object):
+    runid = "59273-01A"
+    project = "Zimmerer"
+    sample = "ORGAN-8"
+    material = "Biotite"
+    investigator = "Zimmerer"
+    fits = "LLLLLL"
+    locality = ""
+    rundate = "3/3/2010"
+    irradlabel = "NM-227L"
+    runhour = 2.0
+    version = 7.875
+    emv = [1.4503029584884644, 0.0]
+    optemp = 0
+    history = (
+        "Multiplier Baseline = L @ --- @ ---; Ar40 = L @ --- @ ---; Ar39 = L @ --- @ ---; "
+        "Ar38 = L @ --- @ ---; Ar36 = L @ --- @ ---; Ar37 = L @ --- @ ---; "
+        "argonlab - 3/11/2010 @ 10:08:42 AM @ --- @ ---; Ar40 bk val,er = Bracketing blanks @ --- @ ---; "
+        "Ar39 bk val,er = Bracketing blanks @ --- @ ---; Ar38 bk val,er = Bracketing blanks @ --- @ ---; "
+        "Ar36 bk val,er = Bracketing blanks @ --- @ ---; Ar37 bk val,er = Bracketing blanks @ --- @ ---"
+    )
+    scalefactor = 1000000000.0, 1000000000.0
+    extract_device = "CO2"
+    extract_value = 0.0
+    final_set_power = 0.5
+    totdur_heating = 0
+    totdur_atsetpoint = 0
+    gain = [166124.90625, 0.0]
+    calc_with_ratio = False
+    system_number = 3
+    mol_ref_iso = 8.004109832880028e-16
+    disc = 1.00600004196167
+    disc_err = 0.0010000000474974513
+    j = 0.002384300110861659
+    j_err = 2.048499936790904e-06
+    resistor_values = [1.0, 0, 0]
+    isogrp = "Ar"
+    niso = 5
+    nratio = 0
+    detectors = [1, 1, 1, 1, 1]
+    ndet = 1
+    refdetnum = 1
+    signormfactor = 1
+    ncyc = 0
+    isokeys = ["Ar40", "Ar39", "Ar38", "Ar36", "Ar37"]
+    runday = 8462.0
+    ncnts = [48, 72, 24, 120, 24]
+    isotopes = {
+        "Ar40": {
+            "background": 0.016390634700655937,
+            "background_err": 0.0001810772664612159,
+            "intercept": 1,
+            "intercept_err": 1,
+            "counts_per_cycle": 0,
+        },
+        "Ar39": {
+            "background": 0.00033566050115041435,
+            "background_err": 1.5215453458949924e-05,
+            "intercept": 1,
+            "intercept_err": 1,
+            "counts_per_cycle": 0,
+        },
+        "Ar36": {
+            "background": 7.502062362618744e-05,
+            "background_err": 4.699999863078119e-06,
+            "intercept": 1,
+            "intercept_err": 1,
+            "counts_per_cycle": 0,
+        },
+    }
+
+    baselines = [{"ncnts": 42}]
+
+
+
+class MassSpecBinaryExtractorTestCase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.extractor = MassSpecBinaryExtractor()
+        cls.specs = cls.extractor.import_file("./data/MSDataFile_7875")
+        cls.expected = Expected()
+
[…]
