+++ pychron/entry/legacy/nmgrl/extractor/mass_spec_binary_extractor.py
@@ -21,13 +21,14 @@ import os
 import struct
 from numpy import array
 from datetime import datetime
+
 # ============= local library imports  ==========================
 
 from pychron.experiment.utilities.identifier import strip_runid
 from pychron.processing.isotope import Isotope
 from pychron.processing.isotope_group import IsotopeGroup
 
-FITS = {'L': 1, 'P': 2, 'A': 'average'}
+FITS = {"L": 1, "P": 2, "A": "average"}
 
 
 class BinarySpec(object):
@@ -45,7 +46,7 @@ class BinarySpec(object):
 
 
 class MeasurementObject(object):
-    name = ''
+    name = ""
     xs = None
     ys = None
     ncnts = 0
@@ -58,12 +59,12 @@ class Baseline(MeasurementObject):
 class MassSpecBinaryExtractor:
     def _get_next_str(self, rfile):
         def gen():
-            t = b''
-            a = b''
+            t = b""
+            a = b""
             while 1:
                 t += a
                 a = rfile.read(1)
-                if a == b'\t':
+                if a == b"\t":
                     break
             return t.decode()
             # t = b''
@@ -83,51 +84,51 @@ class MassSpecBinaryExtractor:
 
     def _get_single(self, rfile):
         def _gen():
-            t = struct.unpack('>f', rfile.read(4))[0]
+            t = struct.unpack(">f", rfile.read(4))[0]
             return t
 
         return _gen
 
     def _get_short(self, rfile):
         def _gen():
-            t = struct.unpack('>h', rfile.read(2))[0]
+            t = struct.unpack(">h", rfile.read(2))[0]
             return t
 
         return _gen
 
     def _get_long(self, rfile):
         def _gen():
-            t = struct.unpack('>l', rfile.read(4))[0]
+            t = struct.unpack(">l", rfile.read(4))[0]
             return t
 
         return _gen
 
     def _get_double(self, rfile):
         def _gen():
-            t = struct.unpack('>d', rfile.read(8))[0]
+            t = struct.unpack(">d", rfile.read(8))[0]
             return t
 
         return _gen
 
     def extract_analysis(self, p, start, runid):
         """
-            clone of RunData.Binary_Read from MassSpec 7.875
+        clone of RunData.Binary_Read from MassSpec 7.875
         """
         if not os.path.isfile(p):
-            print('invalid: {}'.format(p))
+            print("invalid: {}".format(p))
             return
 
-        with open(p, 'rb') as rfile:
+        with open(p, "rb") as rfile:
             rfile.seek((start - 1) * 256)
             return self._extract(rfile, runid)
 
     def _extract(self, rfile, runid):
         is_air, isref, isblank = False, False, False
         f = runid[0]
-        if f in ('A',):
+        if f in ("A",):
             is_air = True
             isref = True
-        elif f in ('B', 'P'):
+        elif f in ("B", "P"):
             isblank = True
 
         bspec = BinarySpec()
@@ -153,7 +154,7 @@ class MassSpecBinaryExtractor:
         bspec.rundate = gns()
         bspec.irradlabel = gns()
         bspec.fits = gns()
-        print('fits', bspec.fits)
+        print("fits", bspec.fits)
         bspec.comment = gns()
         bspec.runhour = gsingle()
 
@@ -178,7 +179,7 @@ class MassSpecBinaryExtractor:
             s1 = 0  # 10  10
 
         bspec.scalefactor = s1, s1
-        bspec.extract_device = gns() if ver >= 4.46 else ''
+        bspec.extract_device = gns() if ver >= 4.46 else ""
 
[…]
