+++ pychron/entry/legacy/nmgrl/importer/run.py
@@ -23,7 +23,9 @@ from pychron.dvc.dvc_persister import DVCPersister, format_repository_identifier
 from pychron.entry.legacy.nmgrl.extractor.jadj_extractor import JAdjustmentExtractor
 from pychron.entry.legacy.nmgrl.importer.irradiation import IrradiationImporter
 from pychron.entry.legacy.nmgrl.importer.sample import SampleImporter
-from pychron.entry.legacy.nmgrl.extractor.mass_spec_binary_extractor import MassSpecBinaryExtractor
+from pychron.entry.legacy.nmgrl.extractor.mass_spec_binary_extractor import (
+    MassSpecBinaryExtractor,
+)
 from pychron.entry.legacy.util import get_dvc
 from pychron.experiment.automated_run.persistence_spec import PersistenceSpec
 from pychron.experiment.automated_run.spec import AutomatedRunSpec
@@ -53,11 +55,11 @@ def add_repository(dvc, repository_identifier, creator, create_repo):
         repo = GitRepoManager()
         repo.open_repo(proot)
 
-        repo.add_ignore('.DS_Store')
+        repo.add_ignore(".DS_Store")
         # self.repo_man = repo
         if create_repo:
             # add repo to central location
-            print('create repo not enabled')
+            print("create repo not enabled")
             # create_github_repo(repository_identifier)
             # url = 'https://github.com/{}/{}.git'.format(ORG, repository_identifier)
             # print('Create repo at github. url={}'.format(url))
@@ -78,55 +80,57 @@ def run_import(persister, sample_spec, irrad_spec, meta_spec, run):
         dest = persister.dvc.db
 
         if dest.get_analysis_runid(run.identifier, run.aliquot, run.step):
-            print('run already exists: {}'.format(run.runid))
+            print("run already exists: {}".format(run.runid))
             # self.warning('{} already exists'.format(make_runid(idn, aliquot, step)))
             return
 
-    rs = AutomatedRunSpec(labnumber=run.identifier,
-                          username=meta_spec['username'],
-                          material=sample_spec.material.name,
-                          project=sample_spec.project.name,
-                          sample=sample_spec.name,
-                          irradiation=irrad_spec[0],
-                          irradiation_level=irrad_spec[1],
-                          irradiation_position=irrad_spec[2],
-                          repository_identifier=meta_spec['repository_identifier'],
-                          mass_spectrometer=meta_spec['mass_spectrometer'],
-                          uuid=str(uuid.uuid4()),
-                          step=run.step,
-                          comment=run.comment,
-                          aliquot=int(run.aliquot),
-                          extract_device=run.extract_device,
-                          duration=run.totdur_heating,
-                          extract_value=run.final_set_power,
-                          # cleanup=extraction.cleanup_duration,
-                          # beam_diameter=extraction.beam_diameter,
-                          # extract_units=extraction.extract_units or '',
-                          # pattern=extraction.pattern or '',
-                          # weight=extraction.weight,
-                          # ramp_duration=extraction.ramp_duration or 0,
-                          # ramp_rate=extraction.ramp_rate or 0,
-                          # queue_conditionals_name='',
-                          # tray=''
-                          )
+    rs = AutomatedRunSpec(
+        labnumber=run.identifier,
+        username=meta_spec["username"],
+        material=sample_spec.material.name,
+        project=sample_spec.project.name,
+        sample=sample_spec.name,
+        irradiation=irrad_spec[0],
+        irradiation_level=irrad_spec[1],
+        irradiation_position=irrad_spec[2],
+        repository_identifier=meta_spec["repository_identifier"],
[…]
