+++ pychron/entry/legacy/nmgrl/importer/__init__.py
@@ -29,4 +29,6 @@ class BaseImporter:
 
     def _load(self):
         pass
+
+
 # ============= EOF =============================================

[…]
