### 2022-08-05 #3386 @arfon
@aschaen @bjicha @stephen-e-cox @sidneyhemming @cmmercer @wrguenthner – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is [_Pychron: Automated Data Collection and Reduction for Noble Gas and Ar/Ar Geochronology_](https://github.com/NMGRL/pychron)
