### 2022-08-12 #3386 @stephen-e-cox
There's a fundamental problem here that you guys have already noted: no one outside of our user group has the hardware necessary to test this software, but none of us are really independent reviewers. I think all of us as users have probably made a (comparatively very small) number of contributions to the code, have regularly corresponded with Jake about new features and troubleshooting, and are both financially and intellectually invested in this project. But I think it's important that our field move beyond leaving the software we use every day languishing in the grey literature, so I'm happy to write a review it if the journal is ok with all of that.

We also have a new user at Columbia, @jennaeverard, who has been setting up a new pychron-controlled analytical setup in our lab as part of her senior thesis project. If she's willing to consider contributing a review, I think it would be valuable because she has done the setup pretty recently with minimal input from me and I believe none directly from Jake. But she's also a college student working on her senior thesis and might not have the time to take this on right now.
