### 2023-04-08 #3386 @pvermees
I can't tick the following check box in my review: `Installation: Does installation proceed as outlined in the documentation?`, because I still haven't managed to install PyChron on my system (Ubuntu 22.04). Consequently, I cannot complete the remainder of the review either.

1. The Docker image throws the following error:

```
Could not connect to any X display.
ERROR conda.cli.main_run:execute(33): Subprocess for 'conda run ['python', './pychron/launchers/launcher.py']' command failed.  (See above for error)
```

2. The main GitHub repo (`NMGRL/pychron`) tells me to install a command line utility called `pcm`. Following the instructions (`pip install pcm` followed by `pcm wizard`), I get:

```
pcm wizard
Usage: pcm [OPTIONS] COMMAND [ARGS]...
Try 'pcm --help' for help.
```

3. Following the link to [https://github.com/PychronLabsLLC/pcm](https://github.com/PychronLabsLLC/pcm), and reinstalling pcm by entering `pip install pychron-cm`; then trying `pcm wizard` again, I get lots of errors, the last of which is:

```
FileNotFoundError: [Errno 2] No such file or directory: 'edm'
```

4. Installed `edm`, and got some more luck with `pcm wizard`, although I still get the following warnings:

```
No candidate found for requirement 'keyring'
Did you mean this?
    pyparsing
```

5. I don't know how to proceed next. The documentation (which dates from 2019) is intimidating and unclear [https://pychron.readthedocs.io/en/latest/user_guide/setup/quick_start.html](https://pychron.readthedocs.io/en/latest/user_guide/setup/quick_start.html). It is unclear how the "Quick Start (ish)" section relates to a code base that is installed through `pip`.
