### 2023-04-08 #3386 @jirhiker
@pvermees @Kevin-Mattheus-Moerman
1. [NMGRL/pychron_docker](https://github.com/NMGRL/pychron_docker) has been update and tested on Ubuntu 22.04 and macOS 12.1
2,3,4. pcm is a separate project from PychronLabsLLC that shouldn't be considered for this review
5. Installation and configuration of pychron is almost always contracted out hence the lack of good open installation documentation.  Pychron_docker was the compromise to let users use a "test" version of pychron.

@pvermees https://github.com/NMGRL/pychron_docker

```
curl -O https://raw.githubusercontent.com/NMGRL/pychron_docker/main/pychron.sh
chmod +x ./pychron.sh
./pychron.sh
```

This worked for me on macOS and Ubuntu 22.04.  FYI On Ubuntu i did have an issue where new windows were black but resizing the window caused it to properly refresh.
