### 2023-04-08 #3386 @pvermees
@jirhiker Docker works now, so I've ticked the Installation check box in the review form. Can you share some example Peak Hop files so that I can test (part of) the functionality? Alternative/additionally, can you direct me to your unit tests (if you have them)?
