### 2023-04-08 #3386 @jirhiker
@pvermees Peak hop can only be used with a data collection version.

The unittests for peakhop (https://github.com/NMGRL/pychron/blob/dev/dc/pychron/test_suite.py, https://github.com/NMGRL/pychron/blob/dev/dc/pychron/experiment/tests/peak_hop_parse.py) are only for testing parsing of the peak hop files.

In general, the `test_suite` runs automatically via a Github Action, its coverage is relatively low but I do try to have it cover the core calculations.

Some example pyscripts are located at https://github.com/NMGRL/pychron/tree/dev/dc/docs/user_guide/operation/scripts/examples
