### 2023-04-17 #3386 @pvermees
I have checked the installation box, but am unable to go further than that. As @jirhiker explained:

> Installation and configuration of pychron is almost always contracted out hence the lack of good open installation documentation.

You would really need to ask an actual user to get feedback on these aspects of the software.
