### 2023-07-13 #3386 @jirhiker
@Kevin-Mattheus-Moerman Thanks for the update. Pychron is used by an increasing number of laboratories for cutting edge isotope geochemistry. It's was originally developed for only the Ar/Ar community but now is being used at a number of other noble gas laboratories.

 According to the Scope & submission requirements (https://joss.theoj.org/about#submitting) Pychron seems to be acceptable for publication in JOSS, most notably "supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets". Pychron is used to collect and process data used in hundreds of published manuscripts and geologic products every year and is becoming the go to software for noble gas isotope geochemistry. Thanks again for your consideration
