### 2023-07-28 #3386 @Kevin-Mattheus-Moerman
@jirhiker the editorial board has review this submission and the above review issue comments. Unfortunately it was determined that, as it stands, this work is not acceptable for JOSS. The core reason is that no reviewers can currently be found which 1) do not have a conflict of interest (e.g. a member of the submitting research group) and/or 2) which can install/evaluate the software.

We will therefore now proceed to reject this submission.

We sincerely apologize for the severe delays this submission has encountered in being processed. We can assure you this is not the normal process for JOSS. We have found it challenging to both find reviewers and suitable handling editors for this submission, which are the key reasons for the delays. However, we do feel an apology is in order, as the above conclusion should perhaps have been drawn much earlier in the process.

Note we do welcome a future re-submission if that is of interested. We would welcome this if the software user community has grown significantly outside the direct developer group. This would mean external reviewers could be easily recruited. Furthermore, any external individual should be able to fully install/evaluate the software.
